### 2020-07-23 #2512 @whedon
[PRE REVIEW]: BetaML: The Beta Machine Learning Toolkit, a self-contained repository of Machine Learning algorithms in Julia
**Submitting author:** @sylvaticus (<a href="http://orcid.org/0000-0002-1534-8697">Antonello Lobianco</a>)
**Repository:** <a href="https://github.com/sylvaticus/BetaML.jl" target ="_blank">https://github.com/sylvaticus/BetaML.jl</a>
**Version:** v0.2.2
