### 2020-07-27 1e726e5
Modified mybinder script
