+++ .github/workflows/binder.yaml
@@ -1,12 +1,12 @@
 name: Binder
-on: [push]
+on: push
 
 jobs:
-  Create-Binder-Badge:
+  Create-MyBinderOrg-Cache:
     runs-on: ubuntu-latest
     steps:
     - name: cache binder build on mybinder.org
-      uses: machine-learning-apps/repo2docker-action@0.2
+      uses: jupyterhub/repo2docker-action@master
       with:
         NO_PUSH: true
         MYBINDERORG_TAG: ${{ github.event.ref }} # This builds the container on mybinder.org with the branch that was pushed on.
