+++ .github/workflows/binder.yaml
@@ -10,3 +10,4 @@ jobs:
       with:
         NO_PUSH: true
         MYBINDERORG_TAG: ${{ github.event.ref }} # This builds the container on mybinder.org with the branch that was pushed on.
+        IMAGE_NAME: sylvaticus/betaml
