### 2020-08-11 af5a755
Added funcxtion "gini" and corrected NN related jupyter notebooks
