### 2020-08-15 1e524bb
Completed first working algo of classification using decision trees/gini
Todo:
- put order, including "standard"/ predict methods
- test with the Sepal database
- made it more julia-way / optimise it (e.g. just create bitarrays and then make the split of the whole matrix rather than append)
- add entropy/variance as split decision methods and introduce regression
- work on ensembling (random forest)
