### 2020-08-17 6c398fd
Working on DecisionTrees
The algo works but the fact is not split in X/Y it means that Y is also "Any" that is not good.
TODO: rewrite the algo with explicitly separate the features from the predictions.
