+++ test/Trees_test.jl
@@ -14,51 +14,50 @@ println("*** Testing Decision trees/Random Forest algorithms...")
 # ==================================
 println("Testing basic classification of decision trees...")
 # ---------------------
-training_data = [
-    "Green"  3.0 "Apple";
-    "Yellow" 3.0 "Apple";
-    "Red"    1.0 "Grape";
-    "Red"    1.0 "Grape";
-    "Yellow" 3.0 "Lemon";
+xtrain = [
+    "Green"  3.0;
+    "Yellow" 3.0;
+    "Red"    1.0;
+    "Red"    1.0;
+    "Yellow" 3.0;
 ]
 
-q1 = Question(1,"Green")
-q2 = Question(2,3)
+ytrain = ["Apple",  "Apple", "Grape", "Grape", "Lemon"]
 
-p1,p2 = partition(training_data,q1)
+#q1 = Question(1,"Green")
+#q2 = Question(2,3)
+#p1,p2 = partition(training_data,q1)
+#infoGain(p1,p2,gini(training_data))
+#p1,p2 = partition(training_data,q2)
 
-info_gain(p1,p2,gini(training_data))
+#a = ["a","b","a","c","d"]
+#c = class_counts(a)
+#gini([1 2 3; 2 3 3; 4 3 3])
+#gini([1 2 3; 2 3 1; 4 3 2; 2 3 4; 1 2 5])
+#gini(training_data)
+#findBestSplit(training_data)
 
-p1,p2 = partition(training_data,q2)
+myTree = buildTree(xtrain,ytrain)
 
-a = ["a","b","a","c","d"]
-c = class_counts(a)
-gini([1 2 3; 2 3 3; 4 3 3])
-gini([1 2 3; 2 3 1; 4 3 2; 2 3 4; 1 2 5])
-gini(training_data)
+#print(myTree)
 
-find_best_split(training_data)
+ŷtrain = predict(myTree, xtrain, ytrain)
+ytrain = convert(Array{Any,1},ytrain)
+accuracy(ŷtrain,ytrain)
 
-my_tree = build_tree(training_data)
-
-print_tree(tree)
-
-classify(training_data[1,:],my_tree)
-
-print_leaf(classify(training_data[1,:], my_tree))
-
-# Evaluate
-testing_data = [
-    "Green" 3 "Apple";
-    "Yellow" 4 "Apple";
-    "Red" 2 "Grape";
-    "Red" 1 "Grape";
-    "Yellow" 3 "Lemon"
+xtest = [
+    "Green"  3;
+    "Yellow" 4;
+    "Red"    2;
+    "Red"    1;
+    "Yellow" 3
 ]
 
-for row in eachrow(testing_data)
-    println("Actual: $(row[end]) Predicted: $(print_leaf(classify(row, my_tree)))")
-end
+ytest = Any["Apple","Apple","Grape","Grape","Lemon"]
+ŷtest = predict(myTree, xtest,ytest)
+accuracy(ŷtest,ytest)
+
+
 
 
 # ==================================
@@ -66,14 +65,18 @@ end
 # ==================================
 
 println("Testing classification of the sepal database using decision trees...")
-iris     = readdlm(joinpath(@__DIR__,"data","iris_shuffled.csv"),",",skipstart=1)
+iris     = readdlm(joinpath(@__DIR__,"data","iris_shuffled.csv"),',',skipstart=1)
 x = convert(Array{Float64,2}, iris[:,1:4])
-y = map(x->Dict("setosa" => 1, "versicolor" => 2, "virginica" =>3)[x],iris[:, 5])
[…]
