### 2020-08-17 531056f
Done decisiontrees for classification (only gini)
Todo:
- implement more metrics (entropy, variance)
- implement regression
