+++ src/Trees.jl
@@ -41,15 +41,13 @@ it appears in the rows from the training data that reach this leaf.
 mutable struct Leaf
     rawPredictions
     predictions
-    function Leaf(data)
-        T = eltype(data[:,end])
-        println(data)
-        println(T)
-        if eltype(data[:,end]) <: Number
-            rawPredictions = data[:,end]
-            prediction = mean(rawPredictions)
+    function Leaf(y)
+        T = eltype(y)
+        if T <: Number
+            rawPredictions = y
+            prediction     = mean(rawPredictions)
         else
-            rawPredictions = classCounts(data)
+            rawPredictions = classCounts(y)
             total = sum(values(rawPredictions))
             predictions = Dict{T,Float64}()
             [predictions[k] = rawPredictions[k] / total for k in keys(rawPredictions)]
@@ -94,39 +92,39 @@ so, add it to 'true rows', otherwise, add it to 'false rows'.
 
 Rows with missing values on the question column are assigned randomply proportionally to the assignment of the non-missing rows.
 """
-function partition(data,question::Question)
-    (N,D) = size(data)
-    true_rows, false_rows, missing_rows = Array{Any,2}(undef,0,D), Array{Any,2}(undef,0,D), Array{Any,2}(undef,0,D)
-    for row in eachrow(data)
+function partition(question::Question,x)
+    N = size(x,1)
+    trueIdx = fill(false,N); falseIdx = fill(false,N); missingIdx = fill(false,N)
+    for (rIdx,row) in enumerate(eachrow(x))
         #println(row)
         if(ismissing(row[question.column]))
-            missing_rows = vcat(missing_rows,permutedims(row))
+            missingIdx[rIdx] = true
         elseif match(question,row)
-            true_rows = vcat(true_rows,permutedims(row))
+            trueIdx[rIdx] = true
         else
-            false_rows = vcat(false_rows,permutedims(row))
+            falseIdx[rIdx] = true
         end
     end
     # Assigning missing rows randomly proportionally to non-missing rows
-    p = size(true_rows,1)/(size(true_rows,1)+size(false_rows,1))
-    r = rand(size(missing_rows,1))
-    for (ridx,row) in enumerate(eachrow(missing_rows))
-        if r[ridx] <= p
-            true_rows = vcat(true_rows,permutedims(row))
-        else
-            false_rows = vcat(false_rows,permutedims(row))
+    p = sum(trueIdx)/(sum(trueIdx)+sum(falseIdx))
+    r = rand(N)
+    for rIdx in 1:N
+        if missingIdx[rIdx]
+            if r[rIdx] <= p
+                trueIdx[rIdx] = true
+            else
+                falseIdx[rIdx] = true
+            end
         end
     end
-    return true_rows, false_rows
+    return trueIdx, falseIdx
 end
 
 """Counts the number of each type of example in a dataset."""
-function classCounts(data)
-    T = eltype(data[:,end])
+function classCounts(y)
+    T = eltype(y)
     counts = Dict{T,Int64}()  # a dictionary of label -> count.
-    for row in eachrow(data)
-        # in our dataset format, the label is always the last column
-        label = row[end]
+    for label in y
         if !(label in keys(counts))
             counts[label] = 1
         else
@@ -164,27 +162,27 @@ end
 
 """Find the best question to ask by iterating over every feature / value
 and calculating the information gain."""
-function findBestSplit(data)
+function findBestSplit(x,y;maxFeatures,splittingCriterion)
 
     best_gain           = 0  # keep track of the best information gain
     best_question       = nothing  # keep train of the feature / value that produced it
-    current_uncertainty = gini(data)
-    D                   = size(data,2)-1  # number of columns (the last column is the label)
[…]
