### 2020-08-17 8a412d0
Improved printing of Decision Trees classifiers
