### 2020-08-17 27d658f
Moved giniImpurity and classCounts to utils
