### 2020-08-17 8dd6ec7
Added "entropy" to the splitting criteria for decision trees
