+++ test/runtests.jl
@@ -1,6 +1,7 @@
 using Test
 
 include("Utils_test.jl")
+include("Trees_test.jl")
 include("Nn_tests.jl")
 include("Perceptron_test.jl")
 include("Clustering_test.jl")
