### 2020-08-18 87336b9
Done decision trees (incl regressions)
not very good for regression compared with NN.
Now: random forests (and putting order/doc!)
