### 2020-08-18 1dd1b86
Done also with Random Forests :-)
Overall the following modules/functions has been added:

- module `Trees` for Decision Trees and Random Forests
- utils functions `giniImpurity`, `classCounts`, `meanDicts` and `entropy`.
