### 2020-08-18 64194d9
Updated project's details and relaxed testing to pass Trevis testing on some julia versions
