+++ .travis.yml
@@ -6,7 +6,7 @@ os:
   - osx
 julia:
   - nightly
-  - 1.3.1
+  - 1.5.0
   - 1
 
 notifications:
@@ -18,7 +18,7 @@ jobs:
 
   include:
     - stage: "Documentation"
-      julia: 1.3
+      julia: 1.5
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
