### 2020-08-18 62e6a38
Removed package dependencies in Project.toml that has been added automatically and are not really needed.
