### 2020-08-18 74b166c
Corrected bug in documentation build
