### 2020-08-18 29e54da
Attached Docstring of Trees module
