+++ test/Utils_test.jl
@@ -204,7 +204,7 @@ c = Dict('b'=>0.6,'e'=>0.4)
 d = Dict('a'=>1)
 dicts = [a,b,c,d]
 @test meanDicts(dicts) == Dict('a' => 0.375,'c' => 0.125,'d' => 0.15,'e' => 0.1,'b' => 0.25)
-
+@test meanDicts(dicts,weights=[1,1,97,1]) == Dict('a' => 0.015, 'b' => 0.586, 'c' => 0.005, 'd' =>0.006, 'e' =>0.388)
 
 a = Dict(1=> 0.1,2=>0.4,3=>0.5)
 b = Dict(4=> 0.3,1=>0.1,2=>0.6)
