### 2020-08-24 945c9c1
Added oobEstimation to Random Trees
TODO:
- redesign the API, like this is shit
