### 2020-08-25 b75caa0
Optimising Trees
Note that while giving a type to the question optimised it dividing time and memory of a factor 3, giving explicit type
to the Leaf structure increased timing a bit.
