+++ src/Trees.jl
@@ -53,9 +53,9 @@ A question used to partition a dataset.
 
 This struct just records a 'column number' and a 'column value' (e.g., Green).
 """
-mutable struct Question
-    column
-    value
+struct Question{T}
+    column::Int64
+    value::T
 end
 
 function print(question::Question)
@@ -80,23 +80,40 @@ A tree's leaf (terminal) node.
 - `predictions`: Either the relative label's count (i.e. a PMF) or the mean
 - `depth`: The nodes's depth in the tree
 """
-mutable struct Leaf
-    rawPredictions
-    predictions
-    depth
-    function Leaf(y,depth)
+struct Leaf{T}
+    predictions::Union{T,Dict{T,Float64}}
+    depth::Int64
+    function Leaf(y::Array{T,1},depth::Int64) where {T}
         if eltype(y) <: Number
             rawPredictions = y
-            predictions     = mean(rawPredictions)
+            predictions    = mean(rawPredictions)
         else
+            #println(y)
+            #println(eltype(y))
             rawPredictions = classCounts(y)
             total = sum(values(rawPredictions))
             predictions = Dict{eltype(y),Float64}()
             [predictions[k] = rawPredictions[k] / total for k in keys(rawPredictions)]
         end
-        return new(rawPredictions,predictions,depth)
+        new{T}(predictions,depth)
     end
 end
+
+#=
+function Leaf(y::Array{T,1},depth::Int64) where {T}
+    if eltype(y) <: Number # TODO replace eltype with T
+        rawPredictions = y
+        predictions     = mean(rawPredictions)
+    else
+        rawPredictions = classCounts(y)
+        total = sum(values(rawPredictions))
+        predictions = Dict{eltype(y),Float64}()
+        [predictions[k] = rawPredictions[k] / total for k in keys(rawPredictions)]
+    end
+    return Leaf(predictions,depth)
+end
+=#
+
 
 """
 
@@ -118,11 +135,11 @@ A tree's non-terminal node.
 - `falseBranch`: A reference to the "false" branch of the trees
 - `depth`: The nodes's depth in the tree
 """
-mutable struct DecisionNode
+struct DecisionNode
     question
     trueBranch
     falseBranch
-    depth
+    depth::Int64
     function DecisionNode(question,trueBranch,falseBranch, depth)
         return new(question,trueBranch,falseBranch, depth)
     end
@@ -300,8 +317,9 @@ The given tree is then returned.
 
 Missing data (in the feature dataset) are supported.
 """
-function buildTree(x, y, depth=1; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), splittingCriterion = eltype(y) <: Number ? "variance" : "gini", forceClassification=false)
+function buildTree(x, y::Array{T,1}, depth=1; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), splittingCriterion = eltype(y) <: Number ? "variance" : "gini", forceClassification=false) where {T}
 
+    #println(depth)
     # Force what would be a regression task into a classification task
     if forceClassification && eltype(y) <: Number
         y = string.(y)
@@ -382,9 +400,9 @@ predictSingle(tree,x)
 Predict the label of a single feature record. See [`predict`](@ref).
 
 """
-function predictSingle(node::Union{DecisionNode,Leaf}, x)
+function predictSingle(node::Union{DecisionNode,Leaf{Tl}}, x) where {Tl}
     # Base case: we've reached a leaf
-    if typeof(node) == Leaf
+    if typeof(node) <: Leaf
         return node.predictions
     end
[…]
