+++ src/Trees.jl
@@ -317,11 +317,11 @@ The given tree is then returned.
 
 Missing data (in the feature dataset) are supported.
 """
-function buildTree(x, y::Array{T,1}, depth=1; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), splittingCriterion = eltype(y) <: Number ? "variance" : "gini", forceClassification=false) where {T}
+function buildTree(x, y::Array{T,1}, depth=1; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), splittingCriterion = T <: Number ? "variance" : "gini", forceClassification=false) where {T}
 
     #println(depth)
     # Force what would be a regression task into a classification task
-    if forceClassification && eltype(y) <: Number
+    if forceClassification && T <: Number
         y = string.(y)
     end
 
@@ -445,13 +445,12 @@ The predictions of the "forest" are then the aggregated predictions of the indiv
 The function returns a touple whose first argument is the forest ityself (array of Trees) and the second one is an array with the ids of the records that has _not_ being used to train the specific tree.
 """
 
-function buildForest(x, y, nTrees=30; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=Int(round(sqrt(size(x,2)))), splittingCriterion = eltype(y) <: Number ? "variance" : "gini", forceClassification=false)
+function buildForest(x, y::Array{Ty,1}, nTrees=30; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=Int(round(sqrt(size(x,2)))), splittingCriterion = Ty <: Number ? "variance" : "gini", forceClassification=false) where {Ty}
     # Force what would be a regression task into a classification task
-    if forceClassification && eltype(y) <: Number
+    if forceClassification && Ty <: Number
         y = string.(y)
     end
-    Tl = eltype(y)
-    forest = Union{DecisionNode,Leaf{Tl}}[]
+    forest = Union{DecisionNode,Leaf{Ty}}[]
     errors = Float64[]
     notSampledByTree = Array{Int64,1}[] # to later compute the Out of Bag Error
 
@@ -480,7 +479,7 @@ Predict the label of a single feature record. See [`predict`](@ref).
 Optionally a weighted mean of tree's prediction is used if the parameter `weights` is given.
 
 """
-function predictSingle(forest::Array{Union{DecisionNode,Leaf{Tl}},1}, x;weights=ones(length(forest))) where {Tl}
+function predictSingle(forest::Array{Union{DecisionNode,Leaf{Ty}},1}, x;weights=ones(length(forest))) where {Ty}
     predictions  = predictSingle.(forest,Ref(x))
     if eltype(predictions) <: AbstractDict   # categorical
         #weights = 1 .- treesErrors # back to the accuracy
@@ -505,7 +504,7 @@ In the first case (numerical predictions) use `meanRelError(ŷ,y)` to assess th
 Optionally a weighted mean of tree's prediction is used if the parameter `weights` is given.
 
 """
-function predict(forest::Array{Union{DecisionNode,Leaf{Tl}},1}, x; weights=ones(length(forest))) where {Tl}
+function predict(forest::Array{Union{DecisionNode,Leaf{Ty}},1}, x; weights=ones(length(forest))) where {Ty}
     predictions = predictSingle.(Ref(forest),eachrow(x); weights = weights)
     return predictions
 end
@@ -517,9 +516,9 @@ end
 Compute the weights of each tree (to use in the prediction of the forest) based on the error of the individual tree computed on the records on which it has not been trained.
 
 """
[…]
