+++ src/Trees.jl
@@ -53,7 +53,8 @@ A question used to partition a dataset.
 
 This struct just records a 'column number' and a 'column value' (e.g., Green).
 """
-struct Question{T}
+abstract type AbstractQuestion end
+struct Question{T} <: AbstractQuestion
     column::Int64
     value::T
 end
@@ -65,6 +66,10 @@ function print(question::Question)
     end
     print("Is col $(question.column) $condition $(question.value) ?")
 end
+
+abstract type AbstractNode end
+abstract type AbstractDecisionNode <: AbstractNode end
+abstract type AbstractLeaf <: AbstractNode end
 
 """
    Leaf(y,depth)
@@ -80,10 +85,10 @@ A tree's leaf (terminal) node.
 - `predictions`: Either the relative label's count (i.e. a PMF) or the mean
 - `depth`: The nodes's depth in the tree
 """
-struct Leaf{T}
-    predictions::Union{T,Dict{T,Float64}}
+struct Leaf{Ty} <: AbstractLeaf
+    predictions::Union{Ty,Dict{Ty,Float64}}
     depth::Int64
-    function Leaf(y::Array{T,1},depth::Int64) where {T}
+    function Leaf(y::Array{Ty,1},depth::Int64) where {Ty}
         if eltype(y) <: Number
             rawPredictions = y
             predictions    = mean(rawPredictions)
@@ -92,10 +97,10 @@ struct Leaf{T}
             #println(eltype(y))
             rawPredictions = classCounts(y)
             total = sum(values(rawPredictions))
-            predictions = Dict{eltype(y),Float64}()
+            predictions = Dict{Ty,Float64}()
             [predictions[k] = rawPredictions[k] / total for k in keys(rawPredictions)]
         end
-        new{T}(predictions,depth)
+        new{Ty}(predictions,depth)
     end
 end
 
@@ -135,13 +140,13 @@ A tree's non-terminal node.
 - `falseBranch`: A reference to the "false" branch of the trees
 - `depth`: The nodes's depth in the tree
 """
-struct DecisionNode
-    question
-    trueBranch
-    falseBranch
+struct DecisionNode{Ty} <: AbstractDecisionNode
+    question::AbstractQuestion
+    trueBranch::Union{DecisionNode{Ty},Leaf{Ty}}
+    falseBranch::Union{DecisionNode{Ty},Leaf{Ty}}
     depth::Int64
-    function DecisionNode(question,trueBranch,falseBranch, depth)
-        return new(question,trueBranch,falseBranch, depth)
+    function DecisionNode(question,trueBranch::Union{DecisionNode{Ty},Leaf{Ty}},falseBranch::Union{DecisionNode{Ty},Leaf{Ty}}, depth) where {Ty}
+        return new{Ty}(question,trueBranch,falseBranch, depth)
     end
 end
 
@@ -317,11 +322,11 @@ The given tree is then returned.
 
 Missing data (in the feature dataset) are supported.
 """
-function buildTree(x, y::Array{T,1}, depth=1; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), splittingCriterion = T <: Number ? "variance" : "gini", forceClassification=false) where {T}
+function buildTree(x, y::Array{Ty,1}, depth=1; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), splittingCriterion = Ty <: Number ? "variance" : "gini", forceClassification=false) where {Ty}
 
     #println(depth)
     # Force what would be a regression task into a classification task
-    if forceClassification && T <: Number
+    if forceClassification && Ty <: Number
         y = string.(y)
     end
 
[…]
