### 2020-08-25 4c3abc1
Fully "typified" the nodes (for as much as possible) and cvonverted splittingCriterion to a function
- Leaf nodes are fully tipified by Ty
- Decision Nodes are typified by the Tx of the question and the Ty, but they can not be fully typified as
there are many possible types as the trees goes down using different features of the X

- splittingCriterion has been replaced from a string to directly a function and in utils `giniImpurity`
has been renamed just `gini` and `variance` has been added as the population variance.
