+++ src/Trees.jl
@@ -54,9 +54,9 @@ A question used to partition a dataset.
 This struct just records a 'column number' and a 'column value' (e.g., Green).
 """
 abstract type AbstractQuestion end
-struct Question{T} <: AbstractQuestion
+struct Question{Tx} <: AbstractQuestion
     column::Int64
-    value::T
+    value::Tx
 end
 
 function print(question::Question)
@@ -69,6 +69,7 @@ end
 
 abstract type AbstractNode end
 abstract type AbstractDecisionNode <: AbstractNode end
+abstract type AbstractDecisionNodeTy{Ty} <: AbstractDecisionNode end
 abstract type AbstractLeaf <: AbstractNode end
 
 """
@@ -104,21 +105,6 @@ struct Leaf{Ty} <: AbstractLeaf
     end
 end
 
-#=
-function Leaf(y::Array{T,1},depth::Int64) where {T}
-    if eltype(y) <: Number # TODO replace eltype with T
-        rawPredictions = y
-        predictions     = mean(rawPredictions)
-    else
-        rawPredictions = classCounts(y)
-        total = sum(values(rawPredictions))
-        predictions = Dict{eltype(y),Float64}()
-        [predictions[k] = rawPredictions[k] / total for k in keys(rawPredictions)]
-    end
-    return Leaf(predictions,depth)
-end
-=#
-
 
 """
 
@@ -140,13 +126,14 @@ A tree's non-terminal node.
 - `falseBranch`: A reference to the "false" branch of the trees
 - `depth`: The nodes's depth in the tree
 """
-struct DecisionNode{Ty} <: AbstractDecisionNode
-    question::AbstractQuestion
-    trueBranch::Union{DecisionNode{Ty},Leaf{Ty}}
-    falseBranch::Union{DecisionNode{Ty},Leaf{Ty}}
+struct DecisionNode{Tx,Ty} <: AbstractDecisionNodeTy{Ty}
+    # Note that a decision node is indeed type unstable, as it host other decision nodes whose X type could be different (different X features can have different type)
+    question::Question{Tx}
+    trueBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}}
+    falseBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}}
     depth::Int64
-    function DecisionNode(question,trueBranch::Union{DecisionNode{Ty},Leaf{Ty}},falseBranch::Union{DecisionNode{Ty},Leaf{Ty}}, depth) where {Ty}
-        return new{Ty}(question,trueBranch,falseBranch, depth)
+    function DecisionNode(question::Question{Tx},trueBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}},falseBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}}, depth) where {Tx,Ty}
+        return new{Tx,Ty}(question,trueBranch,falseBranch, depth)
     end
 end
 
@@ -161,9 +148,10 @@ It compares the feature value in the given record to the value stored in the
 question.
 Numerical features are compared in terms of disequality (">="), while categorical features are compared in terms of equality ("==").
 """
-function match(question, x)
+function match(question::Question{Tx}, x::AbstractArray) where {Tx}
     val = x[question.column]
-    if isa(val, Number) # or isa(val, AbstractFloat) to consider "numeric" only floats
+    if Tx <: Number
+    #if isa(val, Number) # or isa(val, AbstractFloat) to consider "numeric" only floats
         return val >= question.value
     else
         return val == question.value
@@ -178,7 +166,7 @@ Dicotomically partitions a dataset `x` given a question.
 For each row in the dataset, check if it matches the question. If so, add it to 'true rows', otherwise, add it to 'false rows'.
[…]
