+++ src/Utils.jl
@@ -28,7 +28,7 @@ export @codeLocation,
        dtanh, sigmoid, dsigmoid, softmax, dsoftmax, softplus, dsoftplus, mish, dmish, # exp/trig based functions
        bic, aic,
        autoJacobian,
-       squaredCost, dSquaredCost, crossEntropy, dCrossEntropy, giniImpurity, classCounts, meanDicts, entropy,
+       squaredCost, dSquaredCost, crossEntropy, dCrossEntropy, classCounts, meanDicts, gini, entropy, variance,
        error, accuracy, meanRelError,
        l1_distance,l2_distance, l2²_distance, cosine_distance, lse, sterling,
        #normalFixedSd, logNormalFixedSd,
@@ -570,13 +570,13 @@ function meanDicts(dicts; weights=ones(length(dicts)))
 end
 
 """
-   giniImpurity(x)
+   gini(x)
 
 Calculate the Gini Impurity for a list of items (or rows).
 
 See: https://en.wikipedia.org/wiki/Decision_tree_learning#Information_gain
 """
-function giniImpurity(x)
+function gini(x)
     counts = classCounts(x)
     N = size(x,1)
     impurity = 1.0
@@ -604,6 +604,9 @@ function entropy(x)
     end
     return entr
 end
+
+"""variance(x) - population variance"""
+variance(x) = var(x,corrected=false)
 
 # ------------------------------------------------------------------------------
 # Various kernel functions (e.g. for Perceptron)
