+++ test/Utils_test.jl
@@ -169,9 +169,9 @@ a = ["a","b","a","c","d"]
 
 # ==================================
 # New test
-println("** Testing giniImpurity()...")
-@test giniImpurity(['a','b','c','c']) == 0.625 # (1/4) * (3/4) + (2/4) * (2/4) + (1/4)*(3/4)
-@test giniImpurity([1 10; 2 20; 3 30; 2 20]) == 0.625
+println("** Testing gini()...")
+@test gini(['a','b','c','c']) == 0.625 # (1/4) * (3/4) + (2/4) * (2/4) + (1/4)*(3/4)
+@test gini([1 10; 2 20; 3 30; 2 20]) == 0.625
 
 
 #a = -0.01*log2(0.01)
