### 2020-08-25 317ce41
Cleaned up Trees
Solved a problem when y was integer and updated Notebooks
