+++ src/Trees.jl
@@ -87,15 +87,13 @@ A tree's leaf (terminal) node.
 - `depth`: The nodes's depth in the tree
 """
 struct Leaf{Ty} <: AbstractLeaf
-    predictions::Union{Ty,Dict{Ty,Float64}}
+    predictions::Union{Number,Dict{Ty,Float64}}
     depth::Int64
     function Leaf(y::Array{Ty,1},depth::Int64) where {Ty}
         if eltype(y) <: Number
             rawPredictions = y
             predictions    = mean(rawPredictions)
         else
-            #println(y)
-            #println(eltype(y))
             rawPredictions = classCounts(y)
             total = sum(values(rawPredictions))
             predictions = Dict{Ty,Float64}()
@@ -105,9 +103,7 @@ struct Leaf{Ty} <: AbstractLeaf
     end
 end
 
-
 """
-
 
 A Decision Node asks a question.
 
@@ -148,7 +144,7 @@ It compares the feature value in the given record to the value stored in the
 question.
 Numerical features are compared in terms of disequality (">="), while categorical features are compared in terms of equality ("==").
 """
-function match(question::Question{Tx}, x::AbstractArray) where {Tx}
+function match(question::Question{Tx}, x) where {Tx}
     val = x[question.column]
     if Tx <: Number
     #if isa(val, Number) # or isa(val, AbstractFloat) to consider "numeric" only floats
@@ -291,7 +287,7 @@ The given tree is then returned.
 
 Missing data (in the feature dataset) are supported.
 """
-function buildTree(x, y::Array{Ty,1}, depth=1; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), splittingCriterion = Ty <: Number ? variance : gini, forceClassification=false) where {Ty}
+function buildTree(x, y::Array{Ty,1}, depth=1; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), forceClassification=false, splittingCriterion = (Ty <: Number && !forceClassification) ? variance : gini) where {Ty}
 
     #println(depth)
     # Force what would be a regression task into a classification task
@@ -408,18 +404,27 @@ end
 """
    buildForest(x, y, nTrees; maxDepth, minGain, minRecords, maxFeatures, splittingCriterion, forceClassification)
 
-Builds (define and train) a "forest of Decision Trees.
+Builds (define and train) a "forest" of Decision Trees.
 
-See [`buildForest`](@ref). The parameters are exactly the same except here we have `nTrees` to define the number of trees in the forest [def: `30`] and the `maxFeatures` default to `√D` instead of `D`.
 
-Each individual decision tree is built using bootstrap over the data, i.e. "sampling N records with replacement" (hence, some records appear multiple times and some records do not appear in the specific tree training). The `maxFeature` to square of the features inject further variability and reduce the correlation between the forest trees.
+# Parameters:
+See [`buildTree`](@ref). The function has all the parameters of `bildTree` (with the `maxFeatures` defaulting to `√D` instead of `D`) plus the following parameters:
+- `nTrees`: Number of trees in the forest [def: `30`]
+- `β`: Parameter that regulate the weights of the scoring of each tree, to be (optionally) used in prediction (see later) [def: `0`, i.e. uniform weigths]
+- `oob`: Wheter to report the out-of-bag error, an estimation of the generalization accuracy [def: `false`]
 
-The predictions of the "forest" are then the aggregated predictions of the individual trees (from which the name "bagging": **b**oostrap **agg**regat**ing**).
+# Output:
+The function returns a named touple with the following elements:
+- `forest`: the forest ityself (array of Trees)
[…]
