### 2020-08-25 6bf9f42
Added option to run normal vs "all" tests
`Pkg.test("BetaML"; test_args = ["all"])` would run aditional tests.
