+++ docs/src/Notebooks.md
@@ -1,6 +1,4 @@
-# Example
-
-## Notebooks
+# Notebooks
 The following notebooks provide runnable examples of the package functionality:
 
 - Pegasus classifiers: [[Static notebook](https://github.com/sylvaticus/BetaML.jl/blob/master/notebooks/Perceptron.ipynb)] - [[myBinder](https://mybinder.org/v2/gh/sylvaticus/BetaML.jl/master?filepath=notebooks%2FPerceptron.ipynb)]
@@ -13,7 +11,3 @@ The following notebooks provide runnable examples of the package functionality:
 Note: the live, runnable computational environment is a temporary new copy made at each connection. The first time after a commit is done on this repository a new environment has to be set (instead of just being copied), and the server may take several minutes.
 
 This is only if you are the unlucky user triggering the rebuild of the environment after the commit.
-
-## Code under the "test" folder
-
-While the primary reason of the scripts under the "test" folder is to provide automatic testing of the BetaML toolkit, they can also be used to see how functions should be called, as virtually all functions provided by BetaML are tested there. [Have a look](https://github.com/sylvaticus/BetaML.jl/tree/master/test) !
