### 2020-08-26 7fac8f3
Added multithreading to buildForest() and removed a Julia 1.5 warning in NN_tests.jl
Nice, time reduced of a factor 3 with 4 threads.
