### 2020-08-27 05f2853
Implemented unrecursive Tree rewrite up to buildTree
Before we continue with predict, let's see if we really have a performance gain ;-)
