+++ src/Trees.jl
@@ -69,8 +69,13 @@ end
 
 abstract type AbstractNode end
 abstract type AbstractDecisionNode <: AbstractNode end
-abstract type AbstractDecisionNodeTy{Ty} <: AbstractDecisionNode end
+#abstract type AbstractDecisionNodeTy{Ty} <: AbstractDecisionNode end
 abstract type AbstractLeaf <: AbstractNode end
+
+
+struct Tree
+    nodes::Array{AbstractNode,1}
+end
 
 """
    Leaf(y,depth)
@@ -89,12 +94,13 @@ A tree's leaf (terminal) node.
 struct Leaf{Ty} <: AbstractLeaf
     predictions::Union{Number,Dict{Ty,Float64}}
     depth::Int64
-    function Leaf(y::Array{Ty,1},depth::Int64) where {Ty}
+    function Leaf(y::Array{Ty,1},rIds,depth::Int64) where {Ty}
+        thisNodeY = @view y[rIds]
         if eltype(y) <: Number
-            rawPredictions = y
+            rawPredictions = thisNodeY
             predictions    = mean(rawPredictions)
         else
-            rawPredictions = classCounts(y)
+            rawPredictions = classCounts(thisNodeY)
             total = sum(values(rawPredictions))
             predictions = Dict{Ty,Float64}()
             [predictions[k] = rawPredictions[k] / total for k in keys(rawPredictions)]
@@ -122,14 +128,14 @@ A tree's non-terminal node.
 - `falseBranch`: A reference to the "false" branch of the trees
 - `depth`: The nodes's depth in the tree
 """
-struct DecisionNode{Tx,Ty} <: AbstractDecisionNodeTy{Ty}
+mutable struct DecisionNode{Tx} <: AbstractDecisionNode
     # Note that a decision node is indeed type unstable, as it host other decision nodes whose X type could be different (different X features can have different type)
     question::Question{Tx}
-    trueBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}}
-    falseBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}}
+    trueNodeId::Int64
+    falseNodeId::Int64
     depth::Int64
-    function DecisionNode(question::Question{Tx},trueBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}},falseBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}}, depth) where {Tx,Ty}
-        return new{Tx,Ty}(question,trueBranch,falseBranch, depth)
+    function DecisionNode(question::Question{Tx},trueNodeId,falseNodeId,depth) where {Tx}
+        return new{Tx}(question,trueNodeId,falseNodeId,depth)
     end
 end
 
@@ -144,8 +150,7 @@ It compares the feature value in the given record to the value stored in the
 question.
 Numerical features are compared in terms of disequality (">="), while categorical features are compared in terms of equality ("==").
 """
-function match(question::Question{Tx}, x) where {Tx}
-    val = x[question.column]
+function match(question::Question{Tx}, val::Tx) where {Tx}
     if Tx <: Number
     #if isa(val, Number) # or isa(val, AbstractFloat) to consider "numeric" only floats
         return val >= question.value
@@ -155,39 +160,40 @@ function match(question::Question{Tx}, x) where {Tx}
 end
 
 """
-   partition(question,x)
+   partition(question,x,xids)
 
 Dicotomically partitions a dataset `x` given a question.
 
-For each row in the dataset, check if it matches the question. If so, add it to 'true rows', otherwise, add it to 'false rows'.
[…]
