### 2020-08-27 6247ee3
Ended rewriting trees unrecursively
