+++ src/Trees.jl
@@ -73,8 +73,9 @@ abstract type AbstractDecisionNode <: AbstractNode end
 abstract type AbstractLeaf <: AbstractNode end
 
 
-struct Tree
+struct Tree{Ty}
     nodes::Array{AbstractNode,1}
+    type::DataType
 end
 
 """
@@ -173,7 +174,6 @@ function partition(question::Question{Tx},x,rIds) where {Tx}
     trueRIds = Int64[]; falseRIds = Int64[]; missingRIds = Int64[]
     for rid in rIds
         value = x[rid,question.column]
-        #println(row)
         if(ismissing(value))
             push!(missingRIds,rid)
         elseif match(question,value)
@@ -295,7 +295,7 @@ Missing data (in the feature dataset) are supported.
 """
 function buildTree(x, y::Array{Ty,1}, depth=1; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), forceClassification=false, splittingCriterion = (Ty <: Number && !forceClassification) ? variance : gini) where {Ty}
 
-    tree = Tree(Array{AbstractNode,1}[])
+    tree = Tree{Ty}(Array{Union{AbstractDecisionNode,Leaf{Ty}},1}[],Ty)
     N    = size(x,1)
 
     if forceClassification && Ty <: Number
@@ -320,7 +320,7 @@ function buildTree(x, y::Array{Ty,1}, depth=1; maxDepth = size(x,1), minGain=0.0
     end
 
     while length(idsToTreat) > 0
-        workingSet = popfirst!(idsToTreat)
+        workingSet = pop!(idsToTreat)
         thisNodeRIds = workingSet[1]
         #nodex = @view x[workingSet[1],:]
         #nodey = @view y[workingSet[1]]
@@ -361,29 +361,35 @@ function buildTree(x, y::Array{Ty,1}, depth=1; maxDepth = size(x,1), minGain=0.0
     return tree
 end
 
-#=
+#= TODO: need to think on how to print the tree without recursion..
 """
-  print(node)
+  print(tree)
 
 Print a Decision Tree (textual)
 
 """
-function print(node::AbstractNode, rootDepth="")
+function print(tree::Tree)
 
-    depth     = node.depth
+    depth     = 1
     fullDepth = rootDepth*string(depth)*"."
     spacing   = ""
-    if depth  == 1
-        println("*** Printing Decision Tree: ***")
-    else
-        spacing = join(["\t" for i in 1:depth],"")
-    end
+    rootDepth = ""
+    nId = 1
 
-    # Base case: we've reached a leaf
-    if typeof(node) <: Leaf
-        println("  $(node.predictions)")
-        return
-    end
+    while true
+        node = tree.nodes[nID]
+        depth = node.depth
+        if depth  == 1
+            println("*** Printing Decision Tree: ***")
+        else
+            spacing = join(["\t" for i in 1:depth],"")
+        end
+
+        # Base case: we've reached a leaf
+        if typeof(node) <: Leaf
+            println("  $(node.predictions)")
+            continue
+        end
 
     # Print the question at this node
     print("\n$spacing$fullDepth ")
@@ -397,8 +403,9 @@ function print(node::AbstractNode, rootDepth="")
     # Call this function recursively on the false branch
     print(spacing * "--> False:")
     print(node.falseBranch, fullDepth)
+    end
 end
-
+=#
 
 """
 predictSingle(tree,x)
@@ -406,18 +413,25 @@ predictSingle(tree,x)
 Predict the label of a single feature record. See [`predict`](@ref).
 
 """
-function predictSingle(node::Union{DecisionNode{Tx,Ty},Leaf{Ty}}, x) where {Tx,Ty}
-    # Base case: we've reached a leaf
-    if typeof(node) <: Leaf
[…]
