### 2020-08-27 59f687e
Simplified Forest api
TODO: notebook, documentation
