+++ src/Trees.jl
@@ -69,15 +69,35 @@ end
 
 abstract type AbstractNode end
 abstract type AbstractDecisionNode <: AbstractNode end
-#abstract type AbstractDecisionNodeTy{Ty} <: AbstractDecisionNode end
 abstract type AbstractLeaf <: AbstractNode end
 
+"""
+    Tree{Ty}
 
+Type representing an individual Decision Tree.
+
+Nodes are stored in the array `nodes`. The "type" of the tree is given by the type of its labels
+
+"""
 struct Tree{Ty}
-    nodes::Array{AbstractNode,1}
+    nodes::Array{AbstractNode,1} # Array of nodes
     type::DataType
 end
 
+"""
+    Forest{Ty}
+
+Type representing a Random Forest.
+
+Individual trees are stored in the array `trees`. The "type" of the forest is given by the type of the labels on which it has been trained.
+
+# Struct members:
+- `trees`:        The individual Decision Trees
+- `isRegression`: Wheter the forest is to be used for regression jobs or classification
+- `oobData`:      For each tree, the rows number if the data that have _not_ being used to train the specific tree
+- `oobError`:     The out of bag error (if it has been computed)
+- `weights`:      A weight for each tree depending on the tree's score on the oobData (see [`buildForest`](@ref))
+"""
 mutable struct Forest{Ty}
     trees::Array{Tree{Ty},1}
     isRegression::Bool
@@ -92,11 +112,10 @@ end
 A tree's leaf (terminal) node.
 
 # Constructor's arguments:
-- `y`: The labels assorciated to each record (either numerical or categorical)
+- `y`: The labels associated to each record (either numerical or categorical)
 - `depth`: The nodes's depth in the tree
 
 # Struct members:
-- `rawPredictions`: Either the label's count or the numerical labels of the members of the node
 - `predictions`: Either the relative label's count (i.e. a PMF) or the mean
 - `depth`: The nodes's depth in the tree
 """
@@ -119,23 +138,15 @@ struct Leaf{Ty} <: AbstractLeaf
 end
 
 """
-
-A Decision Node asks a question.
-
-This holds a reference to the question, and to the two child nodes.
-"""
-
-
-"""
-   DecisionNode(question,trueBranch,falseBranch, depth)
+   DecisionNode(question,trueNodeId,falseNodeId,depth)
 
 A tree's non-terminal node.
 
 # Constructor's arguments and struct members:
-- `question`: The question asked in this node
-- `trueBranch`: A reference to the "true" branch of the trees
-- `falseBranch`: A reference to the "false" branch of the trees
-- `depth`: The nodes's depth in the tree
+- `question`:    The question asked in this node
+- `trueNodeId`:  The position in the array of the node representing the "true" branch
+- `falseNodeId`: The position in the array of the node representing the  "false" branch of the trees
+- `depth`:       The nodes's depth in the tree
 """
 mutable struct DecisionNode{Tx} <: AbstractDecisionNode
     # Note that a decision node is indeed type unstable, as it host other decision nodes whose X type could be different (different X features can have different type)
@@ -151,7 +162,7 @@ end
 
 """
 
-   match(question, x)
+   match(question, val)
 
 Return a dicotomic answer of a question when applied to a given feature record.
 
@@ -161,7 +172,6 @@ Numerical features are compared in terms of disequality (">="), while categorica
 """
 function match(question::Question{Tx}, val::Tx) where {Tx}
     if Tx <: Number
-    #if isa(val, Number) # or isa(val, AbstractFloat) to consider "numeric" only floats
         return val >= question.value
     else
         return val == question.value
@@ -169,12 +179,12 @@ function match(question::Question{Tx}, val::Tx) where {Tx}
 end
 
 """
-   partition(question,x,xids)
[…]
