### 2020-08-28 f17d17f
Small working on trees.jl optimisation
