### 2020-08-28 1a9300e
DON'T use this commit, extra slow trees :-(
In trees, attempt to call everything with Tx and Ty, but when Tx is just Float64 it doesn't change anytthing, when it is Any it is a disaster
