+++ src/Trees.jl
@@ -46,12 +46,6 @@ using LinearAlgebra, Random, Statistics, Reexport
 export buildTree, buildForest, updateTreesWeights!, predictSingle, predict, print
 import Base.print
 
-import Base.zero
-zero(x::AbstractString) = ""
-zero(x::Type{AbstractString}) = ""
-zero(x::Type{Any}) = undef
-zero(x::Any) = undef
-
 """
    Question
 
@@ -59,8 +53,8 @@ A question used to partition a dataset.
 
 This struct just records a 'column number' and a 'column value' (e.g., Green).
 """
-#abstract type AbstractQuestion end
-struct Question{Tx}
+abstract type AbstractQuestion end
+struct Question{Tx} <: AbstractQuestion
     column::Int64
     value::Tx
 end
@@ -73,9 +67,44 @@ function print(question::Question)
     print("Is col $(question.column) $condition $(question.value) ?")
 end
 
-#abstract type AbstractNode end
-#abstract type AbstractDecisionNode <: AbstractNode end
-#abstract type AbstractLeaf <: AbstractNode end
+abstract type AbstractNode end
+abstract type AbstractDecisionNode <: AbstractNode end
+abstract type AbstractLeaf <: AbstractNode end
+
+"""
+    Tree{Ty}
+
+Type representing an individual Decision Tree.
+
+Nodes are stored in the array `nodes`. The "type" of the tree is given by the type of its labels
+
+"""
+struct Tree{Ty}
+    nodes::Array{AbstractNode,1} # Array of nodes
+    type::DataType
+end
+
+"""
+    Forest{Ty}
+
+Type representing a Random Forest.
+
+Individual trees are stored in the array `trees`. The "type" of the forest is given by the type of the labels on which it has been trained.
+
+# Struct members:
+- `trees`:        The individual Decision Trees
+- `isRegression`: Wheter the forest is to be used for regression jobs or classification
+- `oobData`:      For each tree, the rows number if the data that have _not_ being used to train the specific tree
+- `oobError`:     The out of bag error (if it has been computed)
+- `weights`:      A weight for each tree depending on the tree's score on the oobData (see [`buildForest`](@ref))
+"""
+mutable struct Forest{Ty}
+    trees::Array{Tree{Ty},1}
+    isRegression::Bool
+    oobData::Array{Array{Int64,1},1}
+    oobError::Float64
+    weights::Array{Float64,1}
+end
 
 """
    Leaf(y,depth)
@@ -90,7 +119,7 @@ A tree's leaf (terminal) node.
 - `predictions`: Either the relative label's count (i.e. a PMF) or the mean
 - `depth`: The nodes's depth in the tree
 """
-struct Leaf{Ty}
+struct Leaf{Ty} <: AbstractLeaf
     predictions::Union{Number,Dict{Ty,Float64}}
     depth::Int64
     function Leaf(y::Array{Ty,1},rIds,depth::Int64) where {Ty}
@@ -119,7 +148,7 @@ A tree's non-terminal node.
 - `falseNodeId`: The position in the array of the node representing the  "false" branch of the trees
 - `depth`:       The nodes's depth in the tree
 """
-mutable struct DecisionNode{Tx}
+mutable struct DecisionNode{Tx} <: AbstractDecisionNode
     # Note that a decision node is indeed type unstable, as it host other decision nodes whose X type could be different (different X features can have different type)
     question::Question{Tx}
     trueNodeId::Int64
@@ -128,40 +157,6 @@ mutable struct DecisionNode{Tx}
     function DecisionNode(question::Question{Tx},trueNodeId,falseNodeId,depth) where {Tx}
[…]
