### 2020-08-28 d813a9c
Revert "Small working on trees.jl optimisation"
This reverts commit f17d17f3ed72164debd47d62dcd21df928c3a979.
