+++ src/Trees.jl
@@ -43,7 +43,7 @@ module Trees
 using LinearAlgebra, Random, Statistics, Reexport
 @reexport using ..Utils
 
-export buildTree, buildForest, updateTreesWeights!, predictSingle, predict, print
+export buildTree, buildForest, computeTreesWeights, oobEstimation, predictSingle, predict, print
 import Base.print
 
 """
@@ -76,14 +76,6 @@ abstract type AbstractLeaf <: AbstractNode end
 struct Tree{Ty}
     nodes::Array{AbstractNode,1}
     type::DataType
-end
-
-mutable struct Forest{Ty}
-    trees::Array{Tree{Ty},1}
-    isRegression::Bool
-    oobData::Array{Array{Int64,1},1}
-    oobError::Float64
-    weights::Array{Float64,1}
 end
 
 """
@@ -488,12 +480,12 @@ function buildForest(x, y::Array{Ty,1}, nTrees=30; maxDepth = size(x,1), minGain
     if forceClassification && Ty <: Number
         y = string.(y)
     end
-    trees           = Array{Tree{Ty},1}(undef,nTrees)
+    forest           = Array{Tree{Ty},1}(undef,nTrees)
     notSampledByTree = Array{Array{Int64,1},1}(undef,nTrees) # to later compute the Out of Bag Error
 
     errors = Float64[]
 
-    jobIsRegression = (forceClassification || !(eltype(y) <: Number)) ? false : true # we don't need the tertiary operator here, but it is more clear with it...
+    #jobIsRegression = (forceClassification || !(eltype(y) <: Number ) ? false : true # we don't need the tertiary operator here, but it is more clear with it...
     (N,D) = size(x)
 
     Threads.@threads for i in 1:nTrees
@@ -505,19 +497,19 @@ function buildForest(x, y::Array{Ty,1}, nTrees=30; maxDepth = size(x,1), minGain
         #controly = y[notToSample]
         tree = buildTree(bootstrappedx, bootstrappedy; maxDepth = maxDepth, minGain=minGain, minRecords=minRecords, maxFeatures=maxFeatures, splittingCriterion = splittingCriterion, forceClassification=forceClassification)
         #ŷ = predict(tree,controlx)
-        trees[i] = tree
+        forest[i] = tree
         notSampledByTree[i] = notToSample
     end
 
-    weights = ones(Float64,nTrees)
+    weigths = ones(Float64,nTrees)
     if β > 0
-        weights = updateTreesWeights!(Forest{Ty}(trees,jobIsRegression,notSampledByTree,0.0, weights), x, y, β=β)
+        weigths = computeTreesWeights(forest, notSampledByTree, x, y, forceClassification=forceClassification, β=β)
     end
     oobE = +Inf
     if oob
-        oobE = oobError(Forest{Ty}(trees,jobIsRegression,notSampledByTree,0.0,weights),x,y)
+        oobE = oobError(forest,notSampledByTree,x,y,forceClassification = forceClassification)
     end
-    return Forest{Ty}(trees,jobIsRegression,notSampledByTree,oobE,weights)
+    return (forest=forest,weights=weigths,oobError=oobE)
 end
 
 """
@@ -527,10 +519,8 @@ Predict the label of a single feature record. See [`predict`](@ref).
 Optionally a weighted mean of tree's prediction is used if the parameter `weights` is given.
 
 """
-function predictSingle(forest::Forest{Ty}, x) where {Ty}
-    trees   = forest.trees
-    weights = forest.weights
-    predictions  = predictSingle.(trees,Ref(x))
+function predictSingle(forest::Array{Tree{Ty},1}, x;weights=ones(length(forest))) where {Ty}
+    predictions  = predictSingle.(forest,Ref(x))
[…]
