### 2020-08-28 2bac10f
Revert "Ended rewriting trees unrecursively"
This reverts commit 6247ee37799e77e7d33e45249320827ec7de2c37.
