### 2020-08-28 c7638a3
Revert "Implemented unrecursive Tree rewrite up to buildTree"
This reverts commit 05f2853b1bfecfbf80d259791f25749423dc30d4.
