+++ notebooks/Project.toml
@@ -1,11 +1,0 @@
-[deps]
-BetaML = "024491cd-cc6b-443e-8034-08ea7eb7db2b"
-CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
-DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-DecisionTree = "7806a523-6efd-50cb-b5f6-3fa6f1930dbb"
-Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
-StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
-
-[compat]
-julia = "1.3.1"
