### 2020-08-28 9a54314
The last serie of commit revert everything to recursive Trees
The work on using unrecursive trees didn't get the expected gain, I need to introduce qome other ways to do

The last few commits revert hence everything to the commit:

7fac8f383eb666890c021ddbefa56d3afc5d0dc6
Added multithreading to buildForest() and removed a Julia 1.5 warning in NN_tests.jl

Note that a few things of what it has been reverted should be reimplemented, like better API
