### 2020-08-28 312c188
Small improvement in trees
- Made random only if needed
- bestGain set initially to float in findBestSplit
