### 2020-08-28 aa32e6d
Reimplemented the better api
