+++ src/Trees.jl
@@ -43,7 +43,7 @@ module Trees
 using LinearAlgebra, Random, Statistics, Reexport
 @reexport using ..Utils
 
-export buildTree, buildForest, computeTreesWeights, oobEstimation, predictSingle, predict, print
+export buildTree, buildForest, updateTreesWeights!, predictSingle, predict, print
 import Base.print
 
 """
@@ -57,14 +57,6 @@ abstract type AbstractQuestion end
 struct Question{Tx} <: AbstractQuestion
     column::Int64
     value::Tx
-end
-
-function print(question::Question)
-    condition = "=="
-    if isa(question.value, Number)
-        condition = ">="
-    end
-    print("Is col $(question.column) $condition $(question.value) ?")
 end
 
 abstract type AbstractNode end
@@ -82,7 +74,6 @@ A tree's leaf (terminal) node.
 - `depth`: The nodes's depth in the tree
 
 # Struct members:
-- `rawPredictions`: Either the label's count or the numerical labels of the members of the node
 - `predictions`: Either the relative label's count (i.e. a PMF) or the mean
 - `depth`: The nodes's depth in the tree
 """
@@ -133,6 +124,35 @@ struct DecisionNode{Tx,Ty} <: AbstractDecisionNodeTy{Ty}
     end
 end
 
+"""
+    Forest{Ty}
+
+Type representing a Random Forest.
+
+Individual trees are stored in the array `trees`. The "type" of the forest is given by the type of the labels on which it has been trained.
+
+# Struct members:
+- `trees`:        The individual Decision Trees
+- `isRegression`: Wheter the forest is to be used for regression jobs or classification
+- `oobData`:      For each tree, the rows number if the data that have _not_ being used to train the specific tree
+- `oobError`:     The out of bag error (if it has been computed)
+- `weights`:      A weight for each tree depending on the tree's score on the oobData (see [`buildForest`](@ref))
+"""
+mutable struct Forest{Ty}
+    trees::Array{Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}},1}
+    isRegression::Bool
+    oobData::Array{Array{Int64,1},1}
+    oobError::Float64
+    weights::Array{Float64,1}
+end
+
+function print(question::Question)
+    condition = "=="
+    if isa(question.value, Number)
+        condition = ">="
+    end
+    print("Is col $(question.column) $condition $(question.value) ?")
+end
 
 """
 
@@ -147,7 +167,6 @@ Numerical features are compared in terms of disequality (">="), while categorica
 function match(question::Question{Tx}, x) where {Tx}
     val = x[question.column]
     if Tx <: Number
-    #if isa(val, Number) # or isa(val, AbstractFloat) to consider "numeric" only floats
         return val >= question.value
     else
         return val == question.value
@@ -415,28 +434,26 @@ See [`buildTree`](@ref). The function has all the parameters of `bildTree` (with
 - `oob`: Wheter to report the out-of-bag error, an estimation of the generalization accuracy [def: `false`]
 
 # Output:
-The function returns a named touple with the following elements:
-- `forest`: the forest ityself (array of Trees)
-- `weights`: the per-tree weight based on their accuracy [def: to array of ones if `β ≤ 0`]
-- `oob`:   the estimate of the oob error [def: to `+Inf` if `oob` == `false`]
+- The function returns a Forest object (see [`Forest`](@ref)).
+- The forest weights default to array of ones if `β ≤ 0` and the oob error to `+Inf` if `oob` == `false`.
 
 # Notes :
[…]
