+++ src/Trees.jl
@@ -181,25 +181,37 @@ Dicotomically partitions a dataset `x` given a question.
 For each row in the dataset, check if it matches the question. If so, add it to 'true rows', otherwise, add it to 'false rows'.
 Rows with missing values on the question column are assigned randomly proportionally to the assignment of the non-missing rows.
 """
-function partition(question::Question{Tx},x) where {Tx}
+function partition(question::Question{Tx},x,mCols) where {Tx}
     N = size(x,1)
-    trueIdx = fill(false,N); falseIdx = fill(false,N); missingIdx = fill(false,N)
-    for (rIdx,row) in enumerate(eachrow(x))
-        #println(row)
-        if(ismissing(row[question.column]))
-            missingIdx[rIdx] = true
-        elseif match(question,row)
-            trueIdx[rIdx] = true
-        else
-            falseIdx[rIdx] = true
+
+    trueIdx = fill(false,N); falseIdx = fill(false,N);
+
+    if  in(question.column,mCols) # do we have missings in this col ?
+        missingIdx = fill(false,N)
+        @inbounds for (rIdx,row) in enumerate(eachrow(x))
+            if(ismissing(row[question.column]))
+                missingIdx[rIdx] = true
+            elseif match(question,row)
+                trueIdx[rIdx] = true
+            else
+                falseIdx[rIdx] = true
+            end
         end
-    end
-    # Assigning missing rows randomly proportionally to non-missing rows
-    p = sum(trueIdx)/(sum(trueIdx)+sum(falseIdx))
-    for rIdx in 1:N
-        if missingIdx[rIdx]
-            r = rand()
-            if r[rIdx] <= p
+        # Assigning missing rows randomly proportionally to non-missing rows
+        p = sum(trueIdx)/(sum(trueIdx)+sum(falseIdx))
+        @inbounds for rIdx in 1:N
+            if missingIdx[rIdx]
+                r = rand()
+                if r <= p
+                    trueIdx[rIdx] = true
+                else
+                    falseIdx[rIdx] = true
+                end
+            end
+        end
+    else
+        @inbounds for (rIdx,row) in enumerate(eachrow(x))
+            if match(question,row)
                 trueIdx[rIdx] = true
             else
                 falseIdx[rIdx] = true
@@ -250,7 +262,7 @@ Find the best question to ask by iterating over every feature / value and calcul
 - `splittingCriterion`: The metric to define the "impurity" of the labels
 
 """
-function findBestSplit(x,y::Array{Ty,1};maxFeatures,splittingCriterion=gini) where {Ty}
+function findBestSplit(x,y::Array{Ty,1}, mCols;maxFeatures,splittingCriterion=gini) where {Ty}
     bestGain           = 0.0  # keep track of the best information gain
     bestQuestion       = nothing # keep train of the feature / value that produced it
     currentUncertainty = splittingCriterion(y)
@@ -262,7 +274,7 @@ function findBestSplit(x,y::Array{Ty,1};maxFeatures,splittingCriterion=gini) whe
             question = Question(d, val)
             # try splitting the dataset
             #println(question)
-            trueIdx, falseIdx = partition(question,x)
+            trueIdx, falseIdx = partition(question,x,mCols)
             # Skip this split if it doesn't divide the
             # dataset.
             if sum(trueIdx) == 0 || sum(falseIdx) == 0
@@ -307,7 +319,7 @@ The given tree is then returned.
 
 Missing data (in the feature dataset) are supported.
 """
[…]
