### 2020-08-28 92190a3
Replaced the output of partition from two bitarray to a single one
