### 2020-08-28 4222e9b
Completed second attempt of unrecursion but not changing anything
Even this attempt doesn' change much
