+++ src/Trees.jl
@@ -61,7 +61,6 @@ end
 
 abstract type AbstractNode end
 abstract type AbstractDecisionNode <: AbstractNode end
-abstract type AbstractDecisionNodeTy{Ty} <: AbstractDecisionNode end
 abstract type AbstractLeaf <: AbstractNode end
 
 """
@@ -94,6 +93,14 @@ struct Leaf{Ty} <: AbstractLeaf
     end
 end
 
+struct TempNode
+    trueBranch::Bool
+    parentNode::AbstractDecisionNode
+    depth::Int64
+    x
+    y
+end
+
 """
 
 A Decision Node asks a question.
@@ -113,14 +120,14 @@ A tree's non-terminal node.
 - `falseBranch`: A reference to the "false" branch of the trees
 - `depth`: The nodes's depth in the tree
 """
-struct DecisionNode{Tx,Ty} <: AbstractDecisionNodeTy{Ty}
+mutable struct DecisionNode{Tx} <: AbstractDecisionNode
     # Note that a decision node is indeed type unstable, as it host other decision nodes whose X type could be different (different X features can have different type)
     question::Question{Tx}
-    trueBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}}
-    falseBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}}
+    trueBranch::Union{Nothing,AbstractNode}
+    falseBranch::Union{Nothing,AbstractNode}
     depth::Int64
-    function DecisionNode(question::Question{Tx},trueBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}},falseBranch::Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}}, depth) where {Tx,Ty}
-        return new{Tx,Ty}(question,trueBranch,falseBranch, depth)
+    function DecisionNode(question::Question{Tx},trueBranch::Union{Nothing,AbstractNode},falseBranch::Union{Nothing,AbstractNode}, depth) where {Tx}
+        return new{Tx}(question,trueBranch,falseBranch, depth)
     end
 end
 
@@ -139,7 +146,7 @@ Individual trees are stored in the array `trees`. The "type" of the forest is gi
 - `weights`:      A weight for each tree depending on the tree's score on the oobData (see [`buildForest`](@ref))
 """
 mutable struct Forest{Ty}
-    trees::Array{Union{AbstractDecisionNodeTy{Ty},Leaf{Ty}},1}
+    trees::Array{Union{AbstractDecisionNode,Leaf{Ty}},1}
     isRegression::Bool
     oobData::Array{Array{Int64,1},1}
     oobError::Float64
@@ -263,7 +270,9 @@ function findBestSplit(x,y::Array{Ty,1}, mCols;maxFeatures,splittingCriterion=gi
     currentUncertainty = splittingCriterion(y)
     (N,D)  = size(x)  # number of columns (the last column is the label)
 
-    for d in shuffle(1:D)[1:maxFeatures]      # for each feature (we consider only maxFeatures features randomly)
+    featuresToConsider = (maxFeatures == D) ? (1:D) : shuffle(1:D)[1:maxFeatures]
+
+    for d in featuresToConsider      # for each feature (we consider only maxFeatures features randomly)
         values = Set(skipmissing(x[:,d]))  # unique values in the column
         for val in values  # for each value
             question = Question(d, val)
@@ -316,6 +325,7 @@ Missing data (in the feature dataset) are supported.
 """
[…]
