### 2020-08-28 1b20922
Imrproved timing by presorting the x matrix for the colum looked up in partition
