### 2020-08-28 05d28bf
Further (and I guess final) small optimisations in Trees.
