+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"index.html#![BLogos](assets/BetaML_logo_30x30.png)-BetaML.jl-Documentation-1","page":"Index","title":"(Image: BLogos) BetaML.jl Documentation","text":"","category":"section"},{"location":"index.html#","page":"Index","title":"Index","text":"Welcome to the documentation of the Beta Machine Learning toolkit.","category":"page"},{"location":"index.html#Installation-1","page":"Index","title":"Installation","text":"","category":"section"},{"location":"index.html#","page":"Index","title":"Index","text":"The BetaML package is now included in the standard Julia register, install it with:","category":"page"},{"location":"index.html#","page":"Index","title":"Index","text":"] add BetaML","category":"page"},{"location":"index.html#Loading-the-module(s)-1","page":"Index","title":"Loading the module(s)","text":"","category":"section"},{"location":"index.html#","page":"Index","title":"Index","text":"This package is split in several submodules. You can access its functionality either by using the specific submodule of interest and then directly the provided functionality (utilities are re-exported by each of the other submodules, so normally you don't need to implicitly import them) or by using the root module BetaML and then prefix with BetaML. each object/function you want to use, e.g.:","category":"page"},{"location":"index.html#","page":"Index","title":"Index","text":"using BetaML.Nn\nmyLayer = DenseLayer(2,3)","category":"page"},{"location":"index.html#","page":"Index","title":"Index","text":"or, equivalently,","category":"page"},{"location":"index.html#","page":"Index","title":"Index","text":"using BetaML\nres = BetaML.kernelPerceptron([1.1 2.1; 5.3 4.2; 1.8 1.7], [-1,1,-1])","category":"page"},{"location":"index.html#Usage-1","page":"Index","title":"Usage","text":"","category":"section"},{"location":"index.html#","page":"Index","title":"Index","text":"Documentation for most algorithms can be retrieved using the inline Julia help system (just press the question mark ? and then, on the special help prompt help?>, type the function name).","category":"page"},{"location":"index.html#","page":"Index","title":"Index","text":"For a list of supported algorithms please look at the individual modules:","category":"page"},{"location":"index.html#","page":"Index","title":"Index","text":"BetaML.Perceptron: The Perceptron, Kernel Perceptron and Pegasos classification algorithms;\nBetaML.Trees: The Decision Trees and Random Forests algorithms for classification or regression (with missing values supported);\nBetaML.Nn: Implementation of Artificial Neural Networks;\nBetaML.Clustering`: Clustering algorithms (Kmeans, Mdedoids, EM/GMM) and missing imputation / collaborative filtering / recommandation systems using clusters;\nBetaML.Utils`: Various utility functions (scale, one-hot, distances, kernels, pca, accuracy/error measures..).","category":"page"},{"location":"index.html#Examples-1","page":"Index","title":"Examples","text":"","category":"section"},{"location":"index.html#","page":"Index","title":"Index","text":"Using an Artificial Neural Network for multinomial categorisation","category":"page"},{"location":"index.html#","page":"Index","title":"Index","text":"# Load Modules\nusing BetaML.Nn, DelimitedFiles, Random, StatsPlots # Load the main module and ausiliary modules\nR
[…]
