+++ notebooks/Project.toml
@@ -1,26 +1,21 @@
-name = "BetaML"
-uuid = "024491cd-cc6b-443e-8034-08ea7eb7db2b"
-authors = ["Antonello Lobianco <antonello@lobianco.org>"]
-version = "0.3.0"
-
 [deps]
+BetaML = "024491cd-cc6b-443e-8034-08ea7eb7db2b"
+CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+DecisionTree = "7806a523-6efd-50cb-b5f6-3fa6f1930dbb"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 PDMats = "90014a1f-27ba-587c-ab20-58faa44d9150"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
-Combinatorics = "1"
-Distributions = "0.23"
-PDMats = "0.9, 0.10"
-ProgressMeter = "1.3"
-Reexport = "0.2"
-Zygote = "0.4, 0.5"
 julia = "1.3.1"
