### 2020-08-28 bb4a0bd
Trees: corrected a bug on weights and added possibility to accept missing values also in  prediction.
