+++ src/Trees.jl
@@ -16,11 +16,11 @@ Decision Trees implementation (Module BetaML.Trees)
 
 Implement the functionality required to build a Decision Tree or a whole Random Forest, predict data and assess its performances.
 
-Both Decision Trees and Random Forests can be used for regression or classification problems, based on the type of the labels (numerical or not). You can override the automatic selection with the parameter `forceClassification=true`, typically if your labels are integer representing some categories rather than numbers.
+Both Decision Trees and Random Forests can be used for regression or classification problems, based on the type of the labels (numerical or not). You can override the automatic selection with the parameter `forceClassification=true`, typically if your labels are integer representing some categories rather than numbers. For classification problems the output of `predictSingle` is a dictionary with the key being the labels with non-zero probabilitity and the corresponding value its proobability; for regression it is a numerical value.
 
-Missing data on features are supported.
+Please be aware that, differently from most other implementations, the Random Forest algorithm collects and averages the probabilities from the trees, rather than just repording the mode, i.e. no information is lost and the output of the forest classifier is still a PMF.
 
-Based originally on the [Josh Gordon's code](https://www.youtube.com/watch?v=LDRbO9a6XPU)
+Missing data on features are supported, both on training and on prediction.
 
 The module provide the following functions. Use `?[type or function]` to access their full signature and detailed documentation:
 
@@ -37,6 +37,8 @@ The module provide the following functions. Use `?[type or function]` to access 
 - `Utils.mearRelError(tree or forest,p)`: L-p norm based error
 
 Features are expected to be in the standard format (nRecords × nDimensions matrices) and the labels (either categorical or numerical) as a nRecords column vector.
+
+Acknowlegdments: originally based on the [Josh Gordon's code](https://www.youtube.com/watch?v=LDRbO9a6XPU)
 """
 module Trees
 
@@ -126,8 +128,9 @@ mutable struct DecisionNode{Tx} <: AbstractDecisionNode
     trueBranch::Union{Nothing,AbstractNode}
     falseBranch::Union{Nothing,AbstractNode}
     depth::Int64
-    function DecisionNode(question::Question{Tx},trueBranch::Union{Nothing,AbstractNode},falseBranch::Union{Nothing,AbstractNode}, depth) where {Tx}
-        return new{Tx}(question,trueBranch,falseBranch, depth)
+    pTrue::Float64
+    function DecisionNode(question::Question{Tx},trueBranch::Union{Nothing,AbstractNode},falseBranch::Union{Nothing,AbstractNode}, depth,pTrue) where {Tx}
+        return new{Tx}(question,trueBranch,falseBranch, depth,pTrue)
     end
 end
 
@@ -367,9 +370,8 @@ function buildTree(x, y::Array{Ty,1}, depth=1; maxDepth = size(x,1), minGain=0.0
     # we'll return a leaf.
     if gain <= minGain  return Leaf(y, depth)  end
 
-    rootNode = DecisionNode(question,nothing,nothing,1)
-
     trueIdx = partition(question,x,mCols)
+    rootNode = DecisionNode(question,nothing,nothing,1,sum(trueIdx)/length(trueIdx))
 
     push!(nodes,TempNode(true,rootNode,2,x[trueIdx,:],y[trueIdx]))
     push!(nodes,TempNode(false,rootNode,2,x[map(!,trueIdx),:],y[map(!,trueIdx)]))
@@ -394,7 +396,7 @@ function buildTree(x, y::Array{Ty,1}, depth=1; maxDepth = size(x,1), minGain=0.0
[…]
