### 2020-08-28 bf2c282
Removed @benchmark from testing script :-/
There is still a strange thing on Trees-test.jl that local test don't pass but I can't reproduce when I run the test manually.
