### 2020-08-30 #2512 @sylvaticus
Hello, in the meantime I did update the library, adding a module on Decision Trees/Random Forests.
Whot is the standard procedure for Joss submissions? Should I also update the paper, or that remains the paper for the version of the software as at the time of the submission?
