### 2020-09-04 #2512 @terrytangyuan
@sylvaticus You will be able to update both the version of the software and the paper.
