+++ Project.toml
@@ -18,7 +18,7 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
 Combinatorics = "1"
-Distributions = "0.23"
+Distributions = "0.23, 0.24"
 PDMats = "0.9, 0.10"
 ProgressMeter = "1.3"
 Reexport = "0.2"
