### 2020-10-19 c620c93
Updated the paper with the Trees module
