+++ paper.md
@@ -6,6 +6,8 @@ tags:
   - neural networks
   - deep learning
   - clustering
+  - decision trees
+  - random forest
   - perceptron
   - data science
 authors:
@@ -37,8 +39,8 @@ bibliography: docs/paper/paper.bib
 
 # Summary
 
-A serie of _machine learning_ algorithms has been implemented and bundled in a single package for the Julia language.
-Currently, algorithms are available in the areas of classification (perceptron, kernel perceptron, pegasos), neural networks (feed-forward) and clustering (kmeans, kmenoids, EM, missing values attribution). Development of these algorithms started following the theoretical notes of the
+A serie of _machine learning_ algorithms has been implemented and bundled in a single package for the Julia programming language.
+Currently, algorithms are available in the areas of classification (perceptron, kernel perceptron, pegasos), neural networks (feed-forward), clustering (kmeans, kmenoids, EM, missing values attribution) and decision trees/random forests. Development of these algorithms started following the theoretical notes of the
 MOOC class "Machine Learning with Python: from Linear Models to Deep Learning" from MITx/edX.
 
 This paper presents the motivations and the general approach of the package and gives an overview of its organisation. We refer the reader to the [package documentation](https://sylvaticus.github.io/BetaML.jl/stable) for instructions on how to use the various algorithms provided or to the MOOC notes available on GitHub [@Lobianco:2020] for their mathematical backgrounds.
@@ -81,13 +83,7 @@ Of course you can get much better results (in general) by scaling the variables,
 -->
 
 We believe that the BetaML flexibility and simplicity, together with the efficiency and usability of a Just in Time compiled language like Julia and the convenience to have several ML algorithms and data-science utilities all in the same package,
-will support the needs of that community of
-<!-- can address significantly better the needs of  -->
-students and researchers  that, contrary to industrial practitioners or computer science specialists, don't necessarily need to work  with very large datasets that don't fit in memory or algorithms that require distributed computation.
-
-
-
-
+will support the needs of that community of <!-- can address significantly better the needs of  --> students and researchers  that, contrary to industrial practitioners or computer science specialists, don't necessarily need to work  with very large datasets that don't fit in memory or algorithms that require distributed computation.
 
 
 <!--
@@ -104,14 +100,14 @@ Knet: @Yuret:2016
 
 # Package organisation
 
-The BetaML toolkit is currently composed of 4 modules: `Utils` provides common data-science utility functions to be used in the other modules, `Perceptron` supplies linear and non-linear classifiers based on the classical Perceptron algorithm, `Nn` allows implementing and training artificial neural networks, and `Clustering` includes several clustering algorithms and missing value attribution / collaborative filtering algorithms based on clustering.
+The BetaML toolkit is currently composed of 5 modules: `Utils` provides common data-science utility functions to be used in the other modules, `Perceptron` supplies linear and non-linear classifiers based on the classical Perceptron algorithm, `Nn` allows implementing and training artificial neural networks, `Clustering` includes several clustering algorithms and missing value attribution / collaborative filtering algorithms based on clustering and finally `Trees` implements decision trees classifiers/regressors together with their most common ensemble method, random forests.
 
-`Perceptron`, `Nn` and `Clustering` all import and re-export the `Utils` function, so the final users normally doesn't need to deal with `Utils`, but just with the module of interest.
+`Perceptron`, `Nn`, `Clustering` and `Trees` all import and re-export the `Utils` function, so the final users normally doesn't need to deal with `Utils`, but just with the module of interest.
 
 
 ## The `Utils` module
 
[…]
