### 2020-10-19 9dcd720
Corrected minor doc
