### 2020-10-19 #2512 @sylvaticus
[Paper author here]
Thank you.
I have updated the paper to include the Decision Trees / Random Forests methods that I have added in the meantime.
