### 2020-10-19 #2512 @sylvaticus
Uh, how do I regenerate the PDF after I changed the paper.md file ? I thought it was enough to cite "@whedon generate pdf" in this issue comment..

Does the `@whedon` command is ran straight away, or it takes some time before it is ran ?
