### 2020-10-19 #2512 @danielskatz
commands to whedon have to be the only or first thing in a new comment
