### 2020-11-29 9131f2b
In the JOS paper, added citations/references to ScikitLearn.jl and AutoMLPipeline.jl
