+++ docs/paper/paper.bib
@@ -121,3 +121,21 @@
     primaryClass={cs.LG},
     url={https://arxiv.org/abs/1412.6980}
 }
+
+% ScikitLearn.jl package...
+@Misc{St-Jean:2020,
+  author = {St-Jean, Cédric and Okon, Samuel},
+  title = {ScikitLearn.jl: Julia implementation of the scikit-learn API},
+  howpublished = {GitHub repository},
+  year = {2020},
+  url = {https://github.com/cstjean/ScikitLearn.jl},
+}
+
+% AutoMLPipeline.jl package...
+@Misc{Paulito:2020,
+  author = {Palmes, Paulito},
+  title = { AutoMLPipeline.jl: A package that makes it trivial to create and evaluate machine learning pipeline architectures.},
+  howpublished = {GitHub repository},
+  year = {2020},
+  url = {https://github.com/IBM/AutoMLPipeline.jl},
+}
