### 2020-11-29 0367f51
Renamed em to gmm in clustering
