+++ src/Clustering.jl
@@ -20,10 +20,10 @@ The module provides the following functions. Use `?[function]` to access their f
 - [`initRepresentatives(X,K;initStrategy,Z₀)`](@ref initRepresentatives): Initialisation strategies for Kmean and Kmedoids
 - [`kmeans(X,K;dist,initStrategy,Z₀)](@ref kmeans)`: Classical KMean algorithm
 - [`kmedoids(X,K;dist,initStrategy,Z₀)](@ref kmedoids)`: Kmedoids algorithm
-- [`em(X,K;p₀,mixtures,tol,verbosity,minVariance,minCovariance,initStrategy)](@ref em)`: EM algorithm over GMM
-- [`predictMissing(X,K;p₀,mixtures,tol,verbosity,minVariance,minCovariance)](@ref predictMissing)`: Fill mixing values / collaborative filtering using EM as backbone
+- [`gmm(X,K;p₀,mixtures,tol,verbosity,minVariance,minCovariance,initStrategy)](@ref gmm)`: gmm algorithm over GMM
+- [`predictMissing(X,K;p₀,mixtures,tol,verbosity,minVariance,minCovariance)](@ref predictMissing)`: Fill mixing values / collaborative filtering using gmm as backbone
 
-{Spherical|Diagonal|Full}Gaussian mixtures for `em` / `predictMissing` are already provided. User defined mixtures can be used defining a struct as subtype of `Mixture` and implementing for that mixture the following functions:
+{Spherical|Diagonal|Full}Gaussian mixtures for `gmm` / `predictMissing` are already provided. User defined mixtures can be used defining a struct as subtype of `Mixture` and implementing for that mixture the following functions:
 - `initMixtures!(mixtures, X; minVariance, minCovariance, initStrategy)`
 - `lpdf(m,x,mask)` (for the e-step)
 - `updateParameters!(mixtures, X, pₙₖ; minVariance, minCovariance)` (the m-step)
@@ -36,7 +36,7 @@ using LinearAlgebra, Random, Statistics, Reexport
 
 @reexport using ..Utils
 
-export initRepresentatives, kmeans, kmedoids, em, predictMissing
+export initRepresentatives, kmeans, kmedoids, gmm, predictMissing
 
 abstract type Mixture end
 include("Mixtures.jl")
@@ -268,12 +268,12 @@ function kmedoids(X,K;dist=(x,y) -> norm(x-y),initStrategy="shuffle",Z₀=nothin
 end
 
 
-## The EM algorithm (Lecture/segment 16.5 of https://www.edx.org/course/machine-learning-with-python-from-linear-models-to)
+## The gmm algorithm (Lecture/segment 16.5 of https://www.edx.org/course/machine-learning-with-python-from-linear-models-to)
 
 # no longer true with the numerical trick implemented
 # - For mixtures with full covariance matrix (i.e. `FullGaussian(μ,σ²)`) the minCovariance should NOT be set equal to the minVariance, or if the covariance matrix goes too low, it will become singular and not invertible.
 """
-  em(X,K;p₀,mixtures,tol,verbosity,minVariance,minCovariance,initStrategy)
+  gmm(X,K;p₀,mixtures,tol,verbosity,minVariance,minCovariance,initStrategy)
 
[…]
