### 2020-11-30 1055317
Updated TagBot to work without cron (but not CompactHelper)
