+++ .github/workflows/TagBot.yml
@@ -1,12 +1,30 @@
 name: TagBot
 on:
-  schedule:
-    - cron: 0 0 * * *
+  issue_comment:  # THIS BIT IS NEW
+    types:
+      - created
+  workflow_dispatch:
 jobs:
   TagBot:
+    # THIS 'if' LINE IS NEW
+    if: github.event_name == 'workflow_dispatch' || github.actor == 'JuliaTagBot'
+    # NOTHING BELOW HAS CHANGED
     runs-on: ubuntu-latest
     steps:
       - uses: JuliaRegistries/TagBot@v1
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
-          
+          ssh: ${{ secrets.DOCUMENTER_KEY }}
+
+# OLD:
+#name: TagBot
+#on:
+#  schedule:
+#    - cron: 0 0 * * *
+#jobs:
+#  TagBot:
+#    runs-on: ubuntu-latest
+#    steps:
+#      - uses: JuliaRegistries/TagBot@v1
+#        with:
+#          token: ${{ secrets.GITHUB_TOKEN }}
