### 2020-12-07 3621981
Updated documentation in the form of the readme
