+++ docs/Project.toml
@@ -16,6 +16,6 @@ julia = "1.4"
 ProgressMeter = "1.3"
 Reexport = "0.2"
 Zygote = "0.4"
-Documenter = "0.24"
+Documenter = "0.25"
 PDMats = "0.9"
 Distributions = "0.23"
