### 2020-12-07 91ed2b6
Added Utils.partition to partition multiple matrices at once
TODO: rename all rShuffle and sequential parameters to shuffle and use inside the functions Random.shuffle to avoid collision
