### 2020-12-07 2be32d6
Added Random Forest example to the readme.md
TODO: convert all the example to use the newly provided partition function
