+++ README.md
@@ -32,10 +32,31 @@ We also provide some [notebooks](https://sylvaticus.github.io/BetaML.jl/dev/Note
 
 If you are looking for an introductory book on Julia, have a look on "[Julia Quick Syntax Reference](https://www.julia-book.com/)"(Apress,2019).
 
-The package can be easily used in R or Python employing [JuliaCall](https://github.com/Non-Contradiction/JuliaCall) or [PyJulia](https://github.com/JuliaPy/pyjulia) respectively (see the Documentation).
+The package can be easily used in R or Python employing [JuliaCall](https://github.com/Non-Contradiction/JuliaCall) or [PyJulia](https://github.com/JuliaPy/pyjulia) respectively.
 
 
 ### Examples
+
+We see how to use three different algorithms to learn the relation between floral sepals and petal measures (first 4 columns) and the specie's name (5th column) in the famous [iris flower dataset](https://en.wikipedia.org/wiki/Iris_flower_data_set).
+
+The first two algorithms are example of _supervised_ learning, the third one of _unsupervised_ learning.
+
+- **Using Random Forests for classification**
+
+```julia
+using DelimitedFiles, BetaML.Trees
+
+iris  = readdlm(joinpath(abspath("BetaML"),"test","data","iris.csv"),',',skipstart=1)
+x = iris[:,1:4]
+x = iris[:,1:4]
+((xtrain,xtest),(ytrain,ytest)) = partition([x,y],[0.7,0.3])
+(ytrain,ytest) = dropdims.([ytrain,ytest],dims=2)
+myForest       = buildForest(xtrain,ytrain,30)
+ŷtrain         = Trees.predict(myForest, xtrain)
+ŷtest          = Trees.predict(myForest, xtest)
+trainAccuracy  = accuracy(ŷtrain,ytrain) # 1.00
+testAccuracy   = accuracy(ŷtest,ytest)   # 0.956
+```
 
 - **Using an Artificial Neural Network for multinomial categorisation**
 
@@ -45,7 +66,7 @@ using BetaML.Nn, DelimitedFiles, Random, StatsPlots # Load the main module and a
 Random.seed!(123); # Fix the random seed (to obtain reproducible results)
 
 # Load the data
-iris     = readdlm(joinpath(dirname(Base.find_package("BetaML")),"..","test","data","iris.csv"),',',skipstart=1)
+iris     = readdlm(joinpath(abspath("BetaML"),"test","data","iris.csv"),',',skipstart=1)
 iris     = iris[shuffle(axes(iris, 1)), :] # Shuffle the records, as they aren't by default
 x        = convert(Array{Float64,2}, iris[:,1:4])
 y        = map(x->Dict("setosa" => 1, "versicolor" => 2, "virginica" =>3)[x],iris[:, 5]) # Convert the target column to numbers
@@ -75,7 +96,7 @@ trainAccuracy = accuracy(ŷtrain,ytrain,tol=1) # 0.983
 testAccuracy  = accuracy(ŷtest,ytest,tol=1)   # 1.0
 
 # Visualise results
-testSize = size(ŷtest,1)
+testSize    = size(ŷtest,1)
 ŷtestChosen =  [argmax(ŷtest[i,:]) for i in 1:testSize]
 groupedbar([ytest ŷtestChosen], label=["ytest" "ŷtest (est)"], title="True vs estimated categories") # All records correctly labelled !
 plot(0:res.epochs,res.ϵ_epochs, ylabel="epochs",xlabel="error",legend=nothing,title="Avg. error per epoch on the Sepal dataset")
[…]
