### 2020-12-08 0365510
Updated documentation to use partition() and to correctly load the data in a current-folder-position independent way
