+++ docs/paper/revision/reviewer1_r1.md
@@ -1,0 +1,200 @@
+Okay, here's an **update** of my review from the [pre-review thread](https://github.com/openjournals/joss-reviews/issues/2512)
+
+## What the package provides
+
+The package under review provides pure-julia implementations of two
+tree-based models, three clustering models, a perceptron model (with 3
+variations) and a basic neural network model. In passing, it should be
+noted that all or almost all of these algorithms have existing julia
+implementations (e.g., DecisionTree.jl, Clustering.jl, Flux.jl). The package
+is used in a course on Machine Learning but integration between the
+package and the course is quite loose, as far as I could ascertain
+(more on this below).
+
+~~Apart from a library of loss functions, the package provides no
+other tools.~~ In addition to the models the package provides a number
+of loss functions, as well as activation functions for the neural
+network models, and some tools to rescale data. I did not see tools to
+automate resampling (such as cross-validation), hyper parameter
+optimization, and no model composition (pipelining). The quality of
+the model implementations looks good to me, although the author warns
+us that "the code is not heavily optimized and GPU [for neural
+networks] is not supported "
+
+## Existing machine learning toolboxes in Julia
+
+For context, consider the following multi-paradigm ML
+toolboxes written in Julia which are relatively mature, by Julia standards:
+
+package          | number of models | resampling  | hyper-parameter optimization | composition
+-----------------|------------------|-------------|------------------------------|-------------
+[ScikitLearn.jl](https://github.com/cstjean/ScikitLearn.jl)   | > 150            | yes         | yes                          | basic
+[AutoMLPipeline.jl](https://github.com/IBM/AutoMLPipeline.jl)| > 100            | no          | no                           | medium
+[MLJ.jl](https://joss.theoj.org/papers/10.21105/joss.02704)           | 151              | yes         | yes                          | advanced
+
+In addition to these are several excellent and mature packages
+dedicated to neural-networks, the most popular being the AD-driven
+Flux.jl package. So far, these provide limited meta-functionality,
+although MLJ now provides an interface to certain classes of Flux
+models ([MLJFlux](https://github.com/alan-turing-institute/MLJFlux.jl)) and
+ScikitLearn.jl provides interfaces to python neural network models
+sufficient for small datasets and pedagogical use.
+
+Disclaimer: I am a designer/contributor to MLJ.
+
+**According to the [JOSS requirements](https://joss.theoj.org/about),
+Submissions should "Have an obvious research application."**  In its
+current state of maturity, BetaML is not a serious competitor to the
+frameworks above, for contributing directly to research. However, the
+author argues that it has pedagogical advantages over existing tools.
+
+## Value as pedagogical tool
+
+I don't think there are many rigorous machine learning courses or
+texts closely integrated with models and tools implemented in julia
+and it would be useful to have more of these. ~~The degree of
+integration in this case was difficult for me to ascertain because I
+couldn't see how to access the course notes without formally
+registering for the course (which is, however, free).~~ I was also
+disappointed to find only one link from doc-strings to course
+materials; from this "back door" to the course notes I could find no
+reference back to the package, however. Perhaps there is better
+integration in course exercises? I couldn't figure this out.
+
+**edit** Okay, I see that I missed the link to the course notes, as
+opposed to the course itself. However the notes make only references
+to python code and so do not appear to be directly integrated with the
+package BetaML.
+
+The remaining argument for BetaML's pedagogical value rests on a
+number of perceived drawbacks of existing toolboxes, for the
+beginner. Quoting from the JOSS manuscript:
+
+1. "For example the popular Deep Learning library Flux (Mike Innes,
[…]
