+++ docs/paper/revision/reviewer1_r1_reply.md
@@ -1,0 +1,251 @@
+I would like to thanks the reviewer for having taken the time to provide the review and the useful suggestions in it. I have implemented most of them.
+
+My detailed response is below.
+
+```
+Okay, here's an **update** of my review from the [pre-review thread](https://github.com/openjournals/joss-reviews/issues/2512)
+
+## What the package provides
+
+The package under review provides pure-julia implementations of two
+tree-based models, three clustering models, a perceptron model (with 3
+variations) and a basic neural network model. In passing, it should be
+noted that all or almost all of these algorithms have existing julia
+implementations (e.g., DecisionTree.jl, Clustering.jl, Flux.jl).
+```
+While "most" of the functionality is indeed already present, from the user point of view, they are not necessarily accessed in the same way and for some functionality, like missing imputation using GMM models, I am not aware of implementations in Julia. Also the kind of output is often different from current implementations. For example most classifiers report the whole PMF of the various items rather than the mode. Together with the fact that the function `accuracy` has an extra optional parameter for selecting the range of items to consider the estimate correct, one can train a classifier that is best in returning a correct value for example within the most probable 2 results (rather than the single most probable one). This can be useful in some applications where the second-best is also acceptable.
+
+```
+The package
+is used in a course on Machine Learning but integration between the
+package and the course is quite loose, as far as I could ascertain
+(more on this below).
+```
+I am sorry for the misunderstanding here. I am not affiliated with that course. The course referenced uses Python to teach the algorithms, while I believe a Julia approach, when dealing with the internals of the algorithms (conversely to "just" using some API), is more appropriate, this is why I translated, and generalised, the code in Julia.
+
+```
+~~Apart from a library of loss functions, the package provides no
+other tools.~~ In addition to the models the package provides a number
+of loss functions, as well as activation functions for the neural
+network models, and some tools to rescale data. I did not see tools to
+automate resampling (such as cross-validation), hyper parameter
+optimization, and no model composition (pipelining). The quality of
+the model implementations looks good to me, although the author warns
+us that "the code is not heavily optimized and GPU [for neural
+networks] is not supported "
+```
+While tools for automatic sampling and cross-validation may be in scope with BetaML, I believe that the added value for pipeling in a language like Julia is not so strong like it is for other programming languages.
+In R and Python for example loops are slow, and it definitely helps having a fast library implementing them in the workload.
+Julia is instead highly expressive and has fast loops at the same time. The computational and convenience benefits to use a specific framework to build a chain of models or tune the hyper-parameters balance again the flexibility and easiness of using just the "core" Julia functionalities to do the same, so that the advantage is partially shaded and depends from the situation.
+
+```
+## Existing machine learning toolboxes in Julia
+
+For context, consider the following multi-paradigm ML
+toolboxes written in Julia which are relatively mature, by Julia standards:
+
+package          | number of models | resampling  | hyper-parameter optimization | composition
+-----------------|------------------|-------------|------------------------------|-------------
+[ScikitLearn.jl](https://github.com/cstjean/ScikitLearn.jl)   | > 150            | yes         | yes                          | basic
+[AutoMLPipeline.jl](https://github.com/IBM/AutoMLPipeline.jl)| > 100            | no          | no                           | medium
+[MLJ.jl](https://joss.theoj.org/papers/10.21105/joss.02704)           | 151              | yes         | yes                          | advanced
+
+In addition to these are several excellent and mature packages
+dedicated to neural-networks, the most popular being the AD-driven
[…]
