### 2020-12-28 d20fc3a
Moved CI to GitHub actions
