+++ docs/paper/revision/reviewer1_r1_reply.md
@@ -1,4 +1,6 @@
-I would like to thanks the reviewer for having taken the time to provide the review and the useful suggestions in it. I have implemented most of them.
+# Author's response to @ablaom review 1
+
+Above all, I would like to thanks the reviewer for having taken the time to provide the review and the useful suggestions he brings. I have implemented most of them, as they helped improving the software.
 
 My detailed response is below.
 
@@ -13,7 +15,7 @@ variations) and a basic neural network model. In passing, it should be
 noted that all or almost all of these algorithms have existing julia
 implementations (e.g., DecisionTree.jl, Clustering.jl, Flux.jl).
 ```
-While "most" of the functionality is indeed already present, from the user point of view, they are not necessarily accessed in the same way and for some functionality, like missing imputation using GMM models, I am not aware of implementations in Julia. Also the kind of output is often different from current implementations. For example most classifiers report the whole PMF of the various items rather than the mode. Together with the fact that the function `accuracy` has an extra optional parameter for selecting the range of items to consider the estimate correct, one can train a classifier that is best in returning a correct value for example within the most probable 2 results (rather than the single most probable one). This can be useful in some applications where the second-best is also acceptable.
+While "most" of the functionality is indeed already present, from the user point of view, they are not necessarily accessed in the same way and for some functionality, like missing imputation using GMM models, I am not aware of implementations in Julia. Also the kind of output is often different from current implementations. For example most classifiers in BetaML report the whole PMF of the various items rather than the mode. Together with the fact that the function `accuracy` has an extra optional parameter for selecting the range of items to consider the estimate correct, one can train a classifier that is best in returning a correct value for example within the most probable 2 results (rather than the single most probable one). This can be useful in some applications where the second-best is also an acceptable value.
 
 ```
 The package
@@ -35,7 +37,7 @@ us that "the code is not heavily optimized and GPU [for neural
 networks] is not supported "
 ```
 While tools for automatic sampling and cross-validation may be in scope with BetaML, I believe that the added value for pipeling in a language like Julia is not so strong like it is for other programming languages.
-In R and Python for example loops are slow, and it definitely helps having a fast library implementing them in the workload.
+In R and Python for example loops are slow, and it definitely helps having a fast library implementing for example hyper-parameters tuning.
 Julia is instead highly expressive and has fast loops at the same time. The computational and convenience benefits to use a specific framework to build a chain of models or tune the hyper-parameters balance again the flexibility and easiness of using just the "core" Julia functionalities to do the same, so that the advantage is partially shaded and depends from the situation.
 
 ```
@@ -166,6 +168,7 @@ community would be to integrate the author's course materials with one
 of the mature ML toolboxes. In the case of MLJ, I would be more than
 happy to provide guidance for such a project.
 ```
+
 I do appreciate both the Reviewer comments and the MLJ as a mature, state-of-the art framework, I just believe that there is space for a different approach with different user cases.
 
 ---
@@ -182,7 +185,7 @@ environment (see commment below).
   is true, the paper should state this clearly in the "Summary" (not
   just that it was developed in response to the course).
 ```
-I specified that the focus is on usability, whether for didactic or applied research purposes.
+As specified on a previous comment, the focus is on usability, whether this is important for didactic or applied research purposes.
 
 ```
 - **added** The authors should reference for comparison the toolboxes
@@ -209,7 +212,7 @@ Being a much smaller package than MLJ, I believe the "Installation" and "Loading
   model. A more "Hello World" example might be to train a Decision
   Tree.
 ```
[…]
