+++ .travis.yml
@@ -1,33 +1,0 @@
-codecov: true
-
-language: julia
-os:
-  - linux
-  - osx
-julia:
-  - nightly
-  - 1.5.0
-  - 1
-
-notifications:
-  email: false
-
-jobs:
-  allow_failures:
-    - julia: nightly
-
-  include:
-    - stage: "Documentation"
-      julia: 1.5
-      os: linux
-      script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
-                                               Pkg.instantiate()'
-        - julia --project=docs/ docs/make.jl
-      after_success: skip
-
-after_success:
-  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(process_folder())'
-
-#script:
-#  - $TESTCMD -e 'using Pkg; Pkg.build(); Pkg.test(coverage=true)'
