+++ Project.toml
@@ -22,5 +22,5 @@ Distributions = "0.23, 0.24"
 PDMats = "0.9, 0.10"
 ProgressMeter = "1.3"
 Reexport = "0.2, 1.0"
-Zygote = "0.4, 0.5, O.6"
+Zygote = "0.4, 0.5, 0.6"
 julia = "1.3.1"
