### 2021-02-08 #2849 @arfon
> I would suggest you can reply to @ablaom review questions/comments because it can help also hasten the review process so that I can just focus on those issues not covered by both of your conversations if I still need more clarifications.

:wave: @ppalmes - I think this review could definitely benefit from your input here 🙏
