### 2021-02-12 #2849 @ppalmes
My decision is Major Revision.

The main contribution of the package is the reimplementation in pure Julia of the various algorithms in supervised and unsupervised learning for teaching purposes.

I agree with @ablaom that in terms of usability, other existing toolkit are more straightforward and consistent to use.  Among the things that I consider to be a major issue is the absence of pipeline API. All related packages mentioned support this API which is a big factor for usability.

Here are my list of suggestions:
1. Improve the online documentation. If the target audience are students, the online documentation needs more examples and tutorials. There is only one page showing examples in the online documentation. Notebooks are great but static documentation (html or pdf) is faster to read without any installation issues.
2. Include good use-cases in the documentation that employ the toolkit to solve real problems incorporating different strategies from the toolbox. This will add value to the scholarly effort. Perform some benchmark and discussions about the results and implementation choices including internal data structures.
3. Implement the pipeline API. It's important for the ML toolkit to make the data preprocessing steps composable for easier usage and experimentation from the perspective of students.
