### 2021-02-24 166d558
Added basic MLJ interface for DecisionTree and RandomForest **regressors**
TODO:
- change the API to accept symbols for inner functions instead of name
- add the MLJ api for the classifiers
- add the info about the package, the model and the parameters
- complete the API for the trees
- extend the api for the other algorithms of BetaML
