+++ src/Trees.jl
@@ -44,6 +44,7 @@ Acknowlegdments: originally based on the [Josh Gordon's code](https://www.youtub
 module Trees
 
 using LinearAlgebra, Random, Statistics, Reexport
+
 @reexport using ..Utils
 
 export buildTree, buildForest, updateTreesWeights!, predictSingle, predict, print
@@ -82,7 +83,7 @@ A tree's leaf (terminal) node.
 struct Leaf{Ty} <: AbstractLeaf
     predictions::Union{Number,Dict{Ty,Float64}}
     depth::Int64
-    function Leaf(y::Array{Ty,1},depth::Int64) where {Ty}
+    function Leaf(y::AbstractArray{Ty,1},depth::Int64) where {Ty}
         if eltype(y) <: Number
             rawPredictions = y
             predictions    = mean(rawPredictions)
@@ -279,13 +280,13 @@ Find the best question to ask by iterating over every feature / value and calcul
 - `splittingCriterion`: The metric to define the "impurity" of the labels
 
 """
-function findBestSplit(x,y::Array{Ty,1}, mCols;maxFeatures,splittingCriterion=gini) where {Ty}
+function findBestSplit(x,y::AbstractArray{Ty,1}, mCols;maxFeatures,splittingCriterion=gini) where {Ty}
     bestGain           = 0.0  # keep track of the best information gain
     bestQuestion       = Question(1,1.0) # keep train of the feature / value that produced it
     currentUncertainty = splittingCriterion(y)
     (N,D)  = size(x)  # number of columns (the last column is the label)
 
-    featuresToConsider = (maxFeatures == D) ? (1:D) : shuffle(1:D)[1:maxFeatures]
+    featuresToConsider = (maxFeatures >= D) ? (1:D) : shuffle(1:D)[1:maxFeatures]
 
     for d in featuresToConsider      # for each feature (we consider only maxFeatures features randomly)
         values = Set(skipmissing(x[:,d]))  # unique values in the column
@@ -341,7 +342,7 @@ The given tree is then returned.
 
 Missing data (in the feature dataset) are supported.
 """
-function buildTree(x, y::Array{Ty,1}; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), forceClassification=false, splittingCriterion = (Ty <: Number && !forceClassification) ? variance : gini, mCols=nothing) where {Ty}
+function buildTree(x, y::AbstractArray{Ty,1}; maxDepth = size(x,1), minGain=0.0, minRecords=2, maxFeatures=size(x,2), forceClassification=false, splittingCriterion = (Ty <: Number && !forceClassification) ? variance : gini, mCols=nothing) where {Ty}
 
 
     #println(depth)
@@ -513,7 +514,7 @@ See [`buildTree`](@ref). The function has all the parameters of `bildTree` (with
 - This function optionally reports a weight distribution of the performances of eanch individual trees, as measured using the records he has not being trained with. These weights can then be (optionally) used in the `predict` function. The parameter `β ≥ 0` regulate the distribution of these weights: larger is `β`, the greater the importance (hence the weights) attached to the best-performing trees compared to the low-performing ones. Using these weights can significantly improve the forest performances (especially using small forests), however the correct value of β depends on the problem under exam (and the chosen caratteristics of the random forest estimator) and should be cross-validated to avoid over-fitting.
[…]
