### 2021-02-24 b7f2956
Trees: set the splittingCriterion parameter as a symbol instead of a function
This hovewer has the inconvenient that would preclude users to use their own functions,
so I will gonna revert it
