+++ src/Trees.jl
@@ -44,7 +44,6 @@ Acknowlegdments: originally based on the [Josh Gordon's code](https://www.youtub
 module Trees
 
 using LinearAlgebra, Random, Statistics, Reexport
-
 @reexport using ..Utils
 
 export buildTree, buildForest, updateTreesWeights!, predictSingle, predict, print
@@ -83,7 +82,7 @@ A tree's leaf (terminal) node.
 struct Leaf{Ty} <: AbstractLeaf
     predictions::Union{Number,Dict{Ty,Float64}}
     depth::Int64
-    function Leaf(y::AbstractArray{Ty,1},depth::Int64) where {Ty}
+    function Leaf(y::Array{Ty,1},depth::Int64) where {Ty}
         if eltype(y) <: Number
             rawPredictions = y
             predictions    = mean(rawPredictions)
@@ -256,15 +255,14 @@ Compare the "information gain" my measuring the difference betwwen the "impurity
 - `splittingCriterion`: Metric to adopt to determine the "impurity" (see below)
 
 You can use your own function as the metric. We provide the following built-in metrics:
-- `:gini` (categorical)
-- `:entropy` (categorical)
-- `:variance` (numerical)
+- `gini` (categorical)
+- `entropy` (categorical)
+- `variance` (numerical)
 
 """
-function infoGain(leftY, rightY, parentUncertainty; splittingCriterion=:gini)
+function infoGain(leftY, rightY, parentUncertainty; splittingCriterion=gini)
     p = size(leftY,1) / (size(leftY,1) + size(rightY,1))
-    #return parentUncertainty - p * splittingCriterion(leftY) - (1 - p) * splittingCriterion(rightY)
-    return parentUncertainty - p * getfield(Trees, Symbol(splittingCriterion))(leftY) - (1 - p) * getfield(Trees, Symbol(splittingCriterion))(rightY)
+    return parentUncertainty - p * splittingCriterion(leftY) - (1 - p) * splittingCriterion(rightY)
 end
 
 """
@@ -281,13 +279,13 @@ Find the best question to ask by iterating over every feature / value and calcul
 - `splittingCriterion`: The metric to define the "impurity" of the labels
 
 """
-function findBestSplit(x,y::AbstractArray{Ty,1}, mCols;maxFeatures,splittingCriterion=:gini) where {Ty}
+function findBestSplit(x,y::Array{Ty,1}, mCols;maxFeatures,splittingCriterion=gini) where {Ty}
     bestGain           = 0.0  # keep track of the best information gain
     bestQuestion       = Question(1,1.0) # keep train of the feature / value that produced it
-    currentUncertainty = getfield(Trees, Symbol(splittingCriterion))(y)
+    currentUncertainty = splittingCriterion(y)
     (N,D)  = size(x)  # number of columns (the last column is the label)
 
-    featuresToConsider = (maxFeatures >= D) ? (1:D) : shuffle(1:D)[1:maxFeatures]
+    featuresToConsider = (maxFeatures == D) ? (1:D) : shuffle(1:D)[1:maxFeatures]
 
     for d in featuresToConsider      # for each feature (we consider only maxFeatures features randomly)
         values  = Set(skipmissing(x[:,d]))  # unique values in the column
@@ -335,7 +333,7 @@ The given tree is then returned.
 - `minGain`: The minimum information gain to allow for a node's partition [def: `0`]
 - `minRecords`:  The minimum number of records a node must holds to consider for a partition of it [def: `2`]
 - `maxFeatures`: The maximum number of (random) features to consider at each partitioning [def: `D`, i.e. look at all features]
[…]
