+++ Project.toml
@@ -8,7 +8,6 @@ Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-MLJModelInterface = "e80e1ace-859a-464e-9ed9-23947d8ae3ea"
 PDMats = "90014a1f-27ba-587c-ab20-58faa44d9150"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
@@ -20,7 +19,6 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 [compat]
 Combinatorics = "1"
 Distributions = "0.23, 0.24"
-MLJModelInterface = "0.3.7, 0.4.0"
 PDMats = "0.9, 0.10"
 ProgressMeter = "1.3"
 Reexport = "0.2, 1.0"
