### 2021-02-25 4fa0dcd
Re-added MLJ interface, including regressor and testing + test segmentation
Note: trees regressors MLJ interface not yet working
Partial tests can be run with Pkg.test("BetaML", test_args=["Trees","Clustering","all"])
