### 2021-02-25 14b7d73
Working on MLJ interface for probabilistic classifier
And solved a bug in testing documentation (I was using MyAwesomePackage)
