+++ .github/workflows/ci.yml
@@ -59,8 +59,8 @@ jobs:
       - run: |
           julia --project=docs -e '
             using Documenter: doctest
-            using MyAwesomePackage
-            doctest(MyAwesomePackage)' # change MYPACKAGE to the name of your package
+            using BetaML
+            doctest(BetaML)' # change MYPACKAGE to the name of your package
       - run: julia --project=docs docs/make.jl
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
