### 2021-02-26 5d55af4
CompatHelper: add new compat entry for "CategoricalArrays" at version "0.9"
