+++ Project.toml
@@ -19,6 +19,7 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
+CategoricalArrays = "0.9"
 Combinatorics = "1"
 Distributions = "0.23, 0.24"
 MLJModelInterface = "0.3.7, 0.4.0"
