### 2021-02-26 8485cd5
Expanded Utils.error and Utils.accuracy to work with any T categorical value, not just Int64
