+++ src/Trees.jl
@@ -657,39 +657,39 @@ end
 # MLJ interface
 import MLJModelInterface
 const MMI = MLJModelInterface
-export BetaMLDecisionTreeRegressor, BetaMLRandomForestRegressor, BetaMLDecisionTreeClassifier, BetaMLRandomForestClassifier
+export DecisionTreeRegressor, RandomForestRegressor, DecisionTreeClassifier, RandomForestClassifier
 
-mutable struct BetaMLDecisionTreeRegressor <: MMI.Deterministic
+mutable struct DecisionTreeRegressor <: MMI.Deterministic
     maxDepth::Int64
     minGain::Float64
     minRecords::Int64
     maxFeatures::Int64
     splittingCriterion::Function
 end
-BetaMLDecisionTreeRegressor(;
-   maxDepth=typemax(Int),
+DecisionTreeRegressor(;
+   maxDepth=0, #typemax(Int)
    minGain=0.0,
    minRecords=2,
-   maxFeatures=typemax(Int),
+   maxFeatures=0,
    splittingCriterion=variance
-   ) = BetaMLDecisionTreeRegressor(maxDepth,minGain,minRecords,maxFeatures,splittingCriterion)
+   ) = DecisionTreeRegressor(maxDepth,minGain,minRecords,maxFeatures,splittingCriterion)
 
-mutable struct BetaMLDecisionTreeClassifier <: MMI.Probabilistic
+mutable struct DecisionTreeClassifier <: MMI.Probabilistic
    maxDepth::Int64
    minGain::Float64
    minRecords::Int64
    maxFeatures::Int64
    splittingCriterion::Function
 end
-BetaMLDecisionTreeClassifier(;
-  maxDepth=typemax(Int),
+DecisionTreeClassifier(;
+  maxDepth=0,
   minGain=0.0,
   minRecords=2,
-  maxFeatures=typemax(Int),
+  maxFeatures=0,
   splittingCriterion=gini
-  ) = BetaMLDecisionTreeClassifier(maxDepth,minGain,minRecords,maxFeatures,splittingCriterion)
+  ) = DecisionTreeClassifier(maxDepth,minGain,minRecords,maxFeatures,splittingCriterion)
 
-mutable struct BetaMLRandomForestRegressor <: MMI.Deterministic
+mutable struct RandomForestRegressor <: MMI.Deterministic
    nTrees::Int64
    maxDepth::Int64
    minGain::Float64
@@ -698,17 +698,17 @@ mutable struct BetaMLRandomForestRegressor <: MMI.Deterministic
    splittingCriterion::Function
    β::Float64
 end
-BetaMLRandomForestRegressor(;
+RandomForestRegressor(;
   nTrees=30,
-  maxDepth=typemax(Int),
+  maxDepth=0,
   minGain=0.0,
   minRecords=2,
-  maxFeatures=typemax(Int),
+  maxFeatures=0,
   splittingCriterion=variance,
   β=0.0
-  ) = BetaMLRandomForestRegressor(nTrees,maxDepth,minGain,minRecords,maxFeatures,splittingCriterion,β)
+  ) = RandomForestRegressor(nTrees,maxDepth,minGain,minRecords,maxFeatures,splittingCriterion,β)
 
-mutable struct BetaMLRandomForestClassifier <: MMI.Probabilistic
+mutable struct RandomForestClassifier <: MMI.Probabilistic
     nTrees::Int64
     maxDepth::Int64
     minGain::Float64
@@ -717,38 +717,44 @@ mutable struct BetaMLRandomForestClassifier <: MMI.Probabilistic
     splittingCriterion::Function
     β::Float64
 end
[…]
