+++ test/Trees_test.jl
@@ -174,28 +174,28 @@ myTree3 = buildTree(xtrain,ytrainInt, forceClassification=true)
 # NEW Test
 println("Testing MLJ interface for Trees models....")
 X, y                           = Mlj.@load_boston
-model_dtr                      = BetaMLDecisionTreeRegressor()
+model_dtr                      = DecisionTreeRegressor()
 regressor_dtr                  = Mlj.machine(model_dtr, X, y)
 (fitresult_dtr, cache, report) = Mlj.fit(model_dtr, 0, X, y)
 yhat_dtr                       = Mlj.predict(model_dtr, fitresult_dtr, X)
 @test meanRelError(yhat_dtr,y) < 0.02
 #(l1_distance(y,yhat)/length(y))/mean(y)
 #evaluate!(regressor, resampling=CV(), measure=rms, verbosity=0)
-model_rfr                      = BetaMLRandomForestRegressor()
+model_rfr                      = RandomForestRegressor()
 regressor_rfr                  = Mlj.machine(model_rfr, X, y)
 (fitresult_rfr, cache, report) = Mlj.fit(model_rfr, 0, X, y)
 yhat_rfr                       = Mlj.predict(model_rfr, fitresult_rfr, X)
 @test meanRelError(yhat_rfr,y) < 0.05
 
 X, y                           = Mlj.@load_iris
-model_dtc                      = BetaMLDecisionTreeClassifier()
+model_dtc                      = DecisionTreeClassifier()
 regressor_dtc                  = Mlj.machine(model_dtc, X, y)
 (fitresult_dtc, cache, report) = Mlj.fit(model_dtc, 0, X, y)
 yhat_dtc                       = Mlj.predict(model_dtc, fitresult_dtc, X)
 @test Mlj.mean(Mlj.LogLoss(tol=1e-4)(yhat_dtc, y)) < 0.0002
 #Mlj.evaluate!(regressor_dtc, resampling=Mlj.CV(), measure=Mlj.LogLoss())
 
-model_rfc                      = BetaMLRandomForestClassifier()
+model_rfc                      = RandomForestClassifier(maxFeatures=3)
 regressor_rfc                  = Mlj.machine(model_rfc, X, y)
 (fitresult_rfc, cache, report) = Mlj.fit(model_rfc, 0, X, y)
 yhat_rfc                       = Mlj.predict(model_rfc, fitresult_rfc, X)
