+++ src/Trees.jl
@@ -788,4 +788,34 @@ function MMI.predict(model::Union{DecisionTreeClassifier,RandomForestClassifier}
 end
 
 
+# Model registration
+MMI.metadata_model(DecisionTreeRegressor,
+    input_scitype    = MMI.Table{<: Union{MMI.Missing, MMI.Known}}, # MMI.Table(MMI.Continuous),  # what input data is supported?
+    target_scitype   = AbstractVector{<: MMI.Continuous},           # for a supervised model, what target?
+    supports_weights = false,                                       # does the model support sample weights?
+    descr            = "A simple Decision Tree for regression with support for Missing data, from the Beta Machine Learning Toolkit (BetaML).",
+	load_path        = "BetaML.Trees.DecisionTreeRegressor"
+    )
+MMI.metadata_model(RandomForestRegressor,
+    input_scitype    = MMI.Table{<: Union{MMI.Missing,MMI.Known}}, # MMI.Table(MMI.Continuous),  # what input data is supported?
+    target_scitype   = AbstractVector{<: MMI.Continuous},          # for a supervised model, what target?
+    supports_weights = false,                                      # does the model support sample weights?
+    descr            = "A simple Random Forest ensemble for regression with support for Missing data, from the Beta Machine Learning Toolkit (BetaML).",
+	load_path        = "BetaML.Trees.RandomForestRegressor"
+    )
+MMI.metadata_model(DecisionTreeClassifier,
+    input_scitype    = MMI.Table{<: Union{MMI.Missing, MMI.Known}},                 # MMI.Table(MMI.Continuous),  # what input data is supported?
+    target_scitype   = AbstractVector{<: Union{MMI.Missing,MMI.Finite,MMI.Count}},  # for a supervised model, what target?
+    supports_weights = false,                                                       # does the model support sample weights?
+    descr            = "A simple Decision Tree for classification with support for Missing data, from the Beta Machine Learning Toolkit (BetaML).",
+	load_path        = "BetaML.Trees.DecisionTreeClassifier"
+    )
+MMI.metadata_model(RandomForestClassifier,
+    input_scitype    = MMI.Table{<: Union{MMI.Missing,MMI.Known}},                  # MMI.Table(MMI.Continuous),  # what input data is supported?
+    target_scitype   = AbstractVector{<: Union{MMI.Missing,MMI.Finite,MMI.Count}},  # for a supervised model, what target?
+    supports_weights = false,                                                       # does the model support sample weights?
+    descr            = "A simple Random Forest ensemble for classification with support for Missing data, from the Beta Machine Learning Toolkit (BetaML).",
+	load_path        = "BetaML.Trees.RandomForestClassifier"
+    )
+
 end # end module
