+++ src/Trees_MLJ.jl
@@ -1,0 +1,170 @@
+# MLJ interface for Decision Trees/Random Forests models
+
+export DecisionTreeRegressor, RandomForestRegressor, DecisionTreeClassifier, RandomForestClassifier
+
+
+# ------------------------------------------------------------------------------
+# Model Structure declarations..
+
+mutable struct DecisionTreeRegressor <: MMI.Deterministic
+    maxDepth::Int64
+    minGain::Float64
+    minRecords::Int64
+    maxFeatures::Int64
+    splittingCriterion::Function
+end
+DecisionTreeRegressor(;
+   maxDepth=0, #typemax(Int)
+   minGain=0.0,
+   minRecords=2,
+   maxFeatures=0,
+   splittingCriterion=variance
+   ) = DecisionTreeRegressor(maxDepth,minGain,minRecords,maxFeatures,splittingCriterion)
+
+mutable struct DecisionTreeClassifier <: MMI.Probabilistic
+   maxDepth::Int64
+   minGain::Float64
+   minRecords::Int64
+   maxFeatures::Int64
+   splittingCriterion::Function
+end
+DecisionTreeClassifier(;
+  maxDepth=0,
+  minGain=0.0,
+  minRecords=2,
+  maxFeatures=0,
+  splittingCriterion=gini
+  ) = DecisionTreeClassifier(maxDepth,minGain,minRecords,maxFeatures,splittingCriterion)
+
+mutable struct RandomForestRegressor <: MMI.Deterministic
+   nTrees::Int64
+   maxDepth::Int64
+   minGain::Float64
+   minRecords::Int64
+   maxFeatures::Int64
+   splittingCriterion::Function
+   β::Float64
+end
+RandomForestRegressor(;
+  nTrees=30,
+  maxDepth=0,
+  minGain=0.0,
+  minRecords=2,
+  maxFeatures=0,
+  splittingCriterion=variance,
+  β=0.0
+  ) = RandomForestRegressor(nTrees,maxDepth,minGain,minRecords,maxFeatures,splittingCriterion,β)
+
+mutable struct RandomForestClassifier <: MMI.Probabilistic
+    nTrees::Int64
+    maxDepth::Int64
+    minGain::Float64
+    minRecords::Int64
+    maxFeatures::Int64
+    splittingCriterion::Function
+    β::Float64
+end
+RandomForestClassifier(;
+    nTrees=30,
+    maxDepth=0,
+    minGain=0.0,
+    minRecords=2,
+    maxFeatures=0,
+    splittingCriterion=gini,
+    β=0.0
+) = RandomForestClassifier(nTrees,maxDepth,minGain,minRecords,maxFeatures,splittingCriterion,β)
+
+# ------------------------------------------------------------------------------
+# Fit functions...
+
+function MMI.fit(model::Union{DecisionTreeRegressor,RandomForestRegressor}, verbosity, X, y)
+   x = MMI.matrix(X)                     # convert table to matrix
+   maxDepth         = model.maxDepth == 0 ? size(x,1) : model.maxDepth
+   if (typeof(model) == DecisionTreeRegressor)
+       maxFeatures = model.maxFeatures == 0 ? size(x,2) : model.maxFeatures
+       fitresult   = buildTree(x, y, maxDepth=maxDepth, minGain=model.minGain, minRecords=model.minRecords, maxFeatures=maxFeatures, splittingCriterion=model.splittingCriterion)
+   else
+       maxFeatures = model.maxFeatures == 0 ? Int(round(sqrt(size(x,2)))) : model.maxFeatures
[…]
