+++ src/BetaML.jl
@@ -31,7 +31,8 @@ myLayer = BetaML.DenseLayer(2,3)
 """
 module BetaML
 
-using MLJModelInterface
+import MLJModelInterface
+const MMI = MLJModelInterface
 
 include("Utils.jl")
 using .Utils
@@ -44,11 +45,12 @@ using .Trees
 include("Clustering.jl")
 using .Clustering
 
+# ------------------------------------------------------------------------------
+#MLJ interface...
 
-#MLJ interface
 const ALL_MODELS = (DecisionTreeClassifier, DecisionTreeRegressor, RandomForestClassifier, RandomForestRegressor)
 
-MLJModelInterface.metadata_pkg.(ALL_MODELS,
+MMI.metadata_pkg.(ALL_MODELS,
     name       = "BetaML",
     uuid       = "024491cd-cc6b-443e-8034-08ea7eb7db2b", # see your Project.toml
     url        = "https://github.com/sylvaticus/BetaML.jl",  # URL to your package repo
@@ -56,8 +58,5 @@ MLJModelInterface.metadata_pkg.(ALL_MODELS,
     license    = "MIT",       # your package license
     is_wrapper = false,    # does it wrap around some other package?
 )
-
-
-
 
 end # module
