### 2021-02-26 39a7484
Improved script for doctests
