### 2021-02-26 4676aa1
Corrected the scientific type of the input of the MLJ Trees models
