+++ test/Clustering_test.jl
@@ -84,8 +84,8 @@ out = predictMissing(X,3,mixtures=[SphericalGaussian() for i in 1:3],verbosity=N
 
 X = [1 10.5;1.5 missing; 1.8 8; 1.7 15; 3.2 40; missing missing; 3.3 38; missing -2.3; 5.2 -2.4]
 out2 = predictMissing(X,3,mixtures=[DiagonalGaussian() for i in 1:3],verbosity=NONE)
-@test out2.X̂[2,2] == 11.438358350316872
+@test out2.X̂[2,2] ≈ 11.438358350316872
 
 X = [1 10.5;1.5 missing; 1.8 8; 1.7 15; 3.2 40; missing missing; 3.3 38; missing -2.3; 5.2 -2.4]
 out3 = predictMissing(X,3,mixtures=[FullGaussian() for i in 1:3],verbosity=NONE)
-@test out3.X̂[2,2] == 11.166652292936876
+@test out3.X̂[2,2] ≈ 11.166652292936876
