### 2021-02-26 b853f28
Added the first "additional tests"
To run these test use:

Pkg.test("BetaML", test_args = ["all"])
