+++ test/runtests.jl
@@ -28,4 +28,5 @@ end
 
 if "all" in ARGS
     # run optional long tests
+    include("Trees_test_additional.jl")
 end
