### 2021-02-26 a6fc030
Removed autotesting of docstrings
As loading the BetaML.X modules was creating a strange error in CI
"ERROR: LoadError: Evaluation into the closed module `Trees` breaks incremental compilation because the side effects
will not be permanent. This is likely due to some other module mutating `Trees` with `eval` during precompilation
- don't do this."
