+++ .github/workflows/ci.yml
@@ -57,11 +57,11 @@ jobs:
             Pkg.develop(PackageSpec(path=pwd()))
             Pkg.instantiate()'
       # Test the documentation itself
-      - run: |
-          julia --project=docs -e '
-            using Documenter: doctest
-            using BetaML.Perceptron, BetaML.Nn, BetaML.Clustering, BetaML.Trees
-            doctest(BetaML)' # change MYPACKAGE to the name of your package
+      #- run: |
+      #    julia --project=docs -e '
+      #      using Documenter: doctest
+      #      using BetaML.Perceptron, BetaML.Nn, BetaML.Clustering, BetaML.Trees
+      #      doctest(BetaML)' # change MYPACKAGE to the name of your package
       - run: julia --project=docs docs/make.jl
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
