+++ src/BetaML.jl
@@ -31,8 +31,8 @@ myLayer = BetaML.DenseLayer(2,3)
 """
 module BetaML
 
-import MLJModelInterface
-const MMI = MLJModelInterface
+#import MLJModelInterface
+#const MMI = MLJModelInterface
 
 include("Utils.jl")
 using .Utils
@@ -47,7 +47,7 @@ using .Clustering
 
 # ------------------------------------------------------------------------------
 #MLJ interface...
-
+#=
 const ALL_MODELS = (DecisionTreeClassifier, DecisionTreeRegressor, RandomForestClassifier, RandomForestRegressor)
 
 MMI.metadata_pkg.(ALL_MODELS,
@@ -58,5 +58,5 @@ MMI.metadata_pkg.(ALL_MODELS,
     license    = "MIT",       # your package license
     is_wrapper = false,    # does it wrap around some other package?
 )
-
+=#
 end # module
