### 2021-02-26 570eee1
Removed the tests also concerning MLJ interface....
