### 2021-02-26 27f3982
Readded the import of MLJModelInterface at the very top of the module, but not yet metadata_pkg
