+++ src/BetaML.jl
@@ -47,7 +47,7 @@ using .Clustering
 
 # ------------------------------------------------------------------------------
 #MLJ interface...
-#=
+
 const ALL_MODELS = (DecisionTreeClassifier, DecisionTreeRegressor, RandomForestClassifier, RandomForestRegressor)
 
 MMI.metadata_pkg.(ALL_MODELS,
@@ -58,5 +58,5 @@ MMI.metadata_pkg.(ALL_MODELS,
     license    = "MIT",       # your package license
     is_wrapper = false,    # does it wrap around some other package?
 )
-=#
+
 end # module
