### 2021-02-27 c92612d
Wrapped MLJ.metadata_pkg inside __init__ function
