### 2021-02-27 2f92e3c
Added Utils.issortable(), enabled Trees models with unsortable feratures, remove MLJ related const from __init__ function
