### 2021-02-27 7595ed4
Extended Utils.issortable to all AbstractArrays
