### 2021-03-04 3d89591
Completed multiclass for perceptron_like algorithms and added MJ interface
