+++ test/Perceptron_test.jl
@@ -1,6 +1,8 @@
 using Statistics
 using Test
 using DelimitedFiles
+import MLJBase
+const Mlj = MLJBase
 
 using StableRNGs
 rng = StableRNG(123)
@@ -20,12 +22,37 @@ xtrain = x[1:ntrain,:]
 ytrain = y[1:ntrain]
 xtest = x[ntrain+1:end,:]
 ytest = y[ntrain+1:end]
+classes = unique(y)
+out      = perceptron(xtrain, ytrain, shuffle=false,nMsgs=0)
+ŷtrain    = Perceptron.predict(xtrain,out.θ,out.θ₀,out.classes)
+ϵtrain    = error(ytrain, mode(ŷtrain))
+ŷtest    = Perceptron.predict(xtest,out.θ,out.θ₀,classes)
+outTest  = perceptron(xtrain, ytrain, shuffle=false,nMsgs=0,returnMeanHyperplane=true)
+ŷavgtest = Perceptron.predict(xtest,outTest.θ,outTest.θ₀,outTest.classes)
+ϵ        = error(ytest, mode(ŷtest))
+ϵavg     = error(ytest, mode(ŷavgtest))
+@test ϵ    < 0.03
+@test ϵavg < 0.2
 
-out      = perceptron(xtrain, ytrain, rShuffle=false,nMsgs=0)
-ŷtest    = Perceptron.predict(xtest,out.θ,out.θ₀)
-ŷavgtest = Perceptron.predict(xtest,out.avgθ,out.avgθ₀)
-ϵ        = error(ytest, ŷtest)
-ϵavg     = error(ytest, ŷavgtest)
+println("Testing multiple classes...")
+
+#xtrain = [3 4 5; 5 3 5; 3 7 2; 8 5 3; 4 2 3; 3 2 1; 8 3 4; 3 5 1; 1 9 3; 4 2 1]
+xtrain = rand(100,3)
+ytt    = [(0.5*x[1]+0.2*x[2]^2+0.3*x[3]+1) for (i,x) in enumerate(eachrow(xtrain))]
+ytrain = [i > median(ytt)*1.1 ? "big" :  i > median(ytt)*0.9 ? "avg" : "small" for i in ytt]
+#xtest  = [2 2 3; 3 2 2; 4 1 2; 4 3 2; 3 7 2]
+xtest = rand(20,3)
+ytt2   = [(0.5*x[1]+0.2*x[2]^2+0.3*x[3]+1) for (i,x) in enumerate(eachrow(xtest))]
+ytest  = [i > median(ytt2)*1.1 ? "big" :  i > median(ytt2)*0.9 ? "avg" : "small" for i in ytt2]
+
+out    = perceptron(xtrain,  ytrain, shuffle=false,nMsgs=0)
+ŷtrain = Perceptron.predict(xtrain,out.θ,out.θ₀,out.classes)
+ŷtest  = Perceptron.predict(xtest,out.θ,out.θ₀,out.classes)
+ϵtrain = error(ytrain, mode(ŷtrain))
+ϵtest  = error(ytest, mode(ŷtest))
+
+@test ϵtrain  < 0.4
+@test ϵavg    < 0.4
 
 
 
@@ -38,19 +65,48 @@ println("Going through Test2 (Kernel Perceptron)...")
 xtrain = [3 4 5; 5 3 5; 3 7 2; 8 5 3; 4 2 3; 3 2 1; 8 3 4; 3 5 1; 1 9 3; 4 2 1]
[…]
