+++ src/Perceptron_MLJ.jl
@@ -1,0 +1,162 @@
+# MLJ interface for Decision Trees/Random Forests models
+
+import MLJModelInterface       # It seems that having done this in the top module is not enought
+const MMI = MLJModelInterface  # We need to repoeat it here
+
+export PerceptronClassifier, KernelPerceptronClassifier, PegasosClassifier
+
+
+# ------------------------------------------------------------------------------
+# Model Structure declarations..
+
+mutable struct PerceptronClassifier <: MMI.Probabilistic
+   initialθ::Vector{Float64}
+   initialθ₀::Float64
+   maxEpochs::Int64
+   shuffle::Bool
+   forceOrigin::Bool
+   returnMeanHyperplane::Bool
+end
+PerceptronClassifier(;
+  initialθ=Float64[],
+  initialθ₀=0.0,
+  maxEpochs=1000,
+  shuffle=false,
+  forceOrigin=false,
+  returnMeanHyperplane=false
+  ) = PerceptronClassifier(initialθ,initialθ₀,maxEpochs,shuffle,forceOrigin,returnMeanHyperplane)
+
+#kernelPerceptron(x, y; K=radialKernel, T=100, α=zeros(Int64,length(y)), nMsgs=0, shuffle=false)
+ mutable struct KernelPerceptronClassifier <: MMI.Probabilistic
+     K::Function
+     maxEpochs::Int64
+     initialα::Vector{Int64}
+     shuffle::Bool
+ end
+ KernelPerceptronClassifier(;
+    K=radialKernel,
+    maxEpochs=100,
+    initialα = Int64[],
+    shuffle=false,
+    ) = KernelPerceptronClassifier(K,maxEpochs,initialα,shuffle)
+
+# pegasos(x, y; θ=zeros(size(x,2)),θ₀=0.0, λ=0.5,η= (t -> 1/sqrt(t)), T=1000, nMsgs=10, shuffle=false,forceOrigin=false,returnMeanHyperplane=false
+mutable struct PegasosClassifier <: MMI.Probabilistic
+   initialθ::Vector{Float64}
+   initialθ₀::Float64
+   λ::Float64
+   η::Function
+   maxEpochs::Int64
+   shuffle::Bool
+   forceOrigin::Bool
+   returnMeanHyperplane::Bool
+end
+PegasosClassifier(;
+  initialθ=Float64[],
+  initialθ₀=0.0,
+  λ = 0.5,
+  η = (t -> 1/sqrt(t)),
+  maxEpochs=1000,
+  shuffle=false,
+  forceOrigin=false,
+  returnMeanHyperplane=false,
+  ) = PegasosClassifier(initialθ,initialθ₀,λ,η,maxEpochs,shuffle,forceOrigin,returnMeanHyperplane)
+
+# ------------------------------------------------------------------------------
+# Fit functions...
+
+function MMI.fit(model::PerceptronClassifier, verbosity, X, y)
+ x = MMI.matrix(X)                     # convert table to matrix
+ initialθ  = length(model.initialθ) == 0 ? zeros(size(x,2)) : model.initialθ
+ fitresult = perceptron(x, y; θ=initialθ, θ₀=model.initialθ₀, T=model.maxEpochs, nMsgs=0, shuffle=model.shuffle, forceOrigin=model.forceOrigin, returnMeanHyperplane=model.returnMeanHyperplane)
+ cache=nothing
+ report=nothing
+ return fitresult, cache, report
+end
+
+function MMI.fit(model::KernelPerceptronClassifier, verbosity, X, y)
+ x = MMI.matrix(X)                     # convert table to matrix
[…]
