### 2021-03-04 0a43705
Removes autoJacobian of softmax testing that was creating problems in CI due to zygote craching in Julia < 1.6
