### 2021-03-04 df1c155
Increased a perceptron test output tollerance
