### 2021-03-07 1c5a7ac
MLJ interface: optimised fit function for probabilistic models
Implemented `predictions = MMI.UnivariateFinite(classes, predMatrix)` instead of

`predictions = [MMI.UnivariateFinite(classes, predMatrix[i,:]) for i in 1:nRecords]`
