+++ test/Clustering_test.jl
@@ -3,6 +3,9 @@ using Test
 
 using StableRNGs
 rng = StableRNG(123)
+
+import MLJBase
+const Mlj = MLJBase
 using BetaML.Clustering
 
 
@@ -89,3 +92,15 @@ out2 = predictMissing(X,3,mixtures=[DiagonalGaussian() for i in 1:3],verbosity=N
 X = [1 10.5;1.5 missing; 1.8 8; 1.7 15; 3.2 40; missing missing; 3.3 38; missing -2.3; 5.2 -2.4]
 out3 = predictMissing(X,3,mixtures=[FullGaussian() for i in 1:3],verbosity=NONE)
 @test out3.X̂[2,2] ≈ 11.166652292936876
+
+# ==================================
+# NEW TEST
+println("Testing MLJ interface for Clustering models....")
+X, y                           = Mlj.@load_iris
+model                          = KMeans()
+modelMachine                   = Mlj.machine(model, X)
+(fitResults, cache, report)    = Mlj.fit(model, 0, X)
+distances                      = Mlj.transform(model,fitResults,X)
+yhat                           = Mlj.predict(model, fitResults, X)
+acc = accuracy(Mlj.levelcode.(yhat),Mlj.levelcode.(y),ignoreLabels=true)
+@test acc > 0.8
