### 2021-03-09 3500fae
Added MLJ interface for kmedoids and gmm (started), standardised test file names using the plural (test**s**)
