+++ test/runtests.jl
@@ -11,23 +11,24 @@ else
 end
 
 if "all" in ARGS || "Utils" in ARGS || nArgs == 0
-    include("Utils_test.jl")
+    include("Utils_tests.jl")
 end
 if "all" in ARGS || "Trees" in ARGS || nArgs == 0
-    include("Trees_test.jl")
+    include("Trees_tests.jl")
 end
 if "all" in ARGS || "Nn" in ARGS || nArgs == 0
     include("Nn_tests.jl")
 end
 if "all" in ARGS || "Perceptron" in ARGS || nArgs == 0
-    include("Perceptron_test.jl")
+    include("Perceptron_tests.jl")
 end
 if "all" in ARGS || "Clustering" in ARGS || nArgs == 0
-    include("Clustering_test.jl")
+    include("Clustering_tests.jl")
 end
 
 if "all" in ARGS
     # run optional long tests
-    include("Perceptron_test_additional.jl")
-    include("Trees_test_additional.jl")
+    include("Perceptron_tests_additional.jl")
+    include("Trees_tests_additional.jl")
+    include("Clustering_tests_additional.jl")
 end
