+++ src/Clustering_MLJ.jl
@@ -3,7 +3,7 @@
 import MLJModelInterface       # It seems that having done this in the top module is not enought
 const MMI = MLJModelInterface  # We need to repoeat it here
 
-export KMeans
+export KMeans, KMedoids, GMM
 
 # ------------------------------------------------------------------------------
 # Model Structure declarations..
@@ -19,39 +19,84 @@ KMeans(;
    dist         = dist=(x,y) -> norm(x-y),
    initStrategy = "grid",
    Z₀           = nothing
-   ) = KMeans(K,dist,initStrategy,Z₀)
+ ) = KMeans(K,dist,initStrategy,Z₀)
+
+ mutable struct KMedoids <: MMI.Unsupervised
+    K::Int64
+    dist::Function
+    initStrategy::String
+    Z₀::Union{Nothing,Matrix{Float64}}
+ end
+ KMedoids(;
+    K            = 3,
+    dist         = dist=(x,y) -> norm(x-y),
+    initStrategy = "grid",
+    Z₀           = nothing
+  ) = KMedoids(K,dist,initStrategy,Z₀)
+
+# function gmm(X,K;p₀=nothing,mixtures=[DiagonalGaussian() for i in 1:K],tol=10^(-6),verbosity=STD,minVariance=0.05,minCovariance=0.0,initStrategy="grid")
+mutable struct GMM{TM <: AbstractMixture} <: MMI.Unsupervised
+  K::Int64
+  p₀::Union{Nothing,AbstractArray{Float64,1}}
+  mixtures::AbstractArray{TM,1}
+  minVariance::Float64
+  minCovariance::Float64
+  initStrategy::String
+end
+GMM(;
+    K             = 3,
+    p₀            = nothing,
+    mixtures      = [DiagonalGaussian() for i in 1:K],
+    minVariance   = 0.05,
+    minCovariance = 0.0,
+    initStrategy  = "kmeans",
+) = GMM(K,p₀,mixtures, minVariance, minCovariance,initStrategy)
 
 # ------------------------------------------------------------------------------
 # Fit functions...
-function MMI.fit(m::KMeans, verbosity, X)
-    x                = MMI.matrix(X)                        # convert table to matrix
-    (assignedClasses,representatives) = kmeans(x,m.K,dist=m.dist,initStrategy=m.initStrategy,Z₀=m.Z₀)
+function MMI.fit(m::Union{KMeans,KMedoids}, verbosity, X)
+    x  = MMI.matrix(X)                        # convert table to matrix
+    if typeof(m) == KMeans
+        (assignedClasses,representatives) = kmeans(x,m.K,dist=m.dist,initStrategy=m.initStrategy,Z₀=m.Z₀)
+    else
+        (assignedClasses,representatives) = kmedoids(x,m.K,dist=m.dist,initStrategy=m.initStrategy,Z₀=m.Z₀)
+    end
     cache=nothing
     report=nothing
     return ((assignedClasses,representatives,m.dist), cache, report)
 end
 
+function MMI.fit(m::GMM, verbosity, X)
+    x          = MMI.matrix(X) # convert table to matrix
+    res        = gmm(x,m.K,p₀=m.p₀,mixtures=m.mixtures, minVariance=m.minVariance, minCovariance=m.minCovariance,initStrategy=m.initStrategy,verbosity=NONE)
+    fitResults = (res.pₙₖ,res.pₖ,res.mixtures)
+    cache      = nothing
+    report     = (res.ϵ,res.lL,res.BIC,res.AIC)
+    return (fitResults, cache, report)
+end
+
+
+
+
 # ------------------------------------------------------------------------------
 # Transform functions...
 
[…]

[…]
