### 2021-03-09 61dca76
Seems ok but gmm testing now doesn't work !
