### 2021-03-09 518abf8
Corrected a test for the changement of the default init strategy for gmm()
Other changes in the previous commits:
- Mixture renamed to AbstractMixture
- Added Utils.getPermutations()
- Added parameter maxIter to gmm
- Added option "given" to `initStrategy` to `gmm` and `initMixtures!`
- Working on MLJ.predict(m::GMM)
