+++ test/Clustering_tests.jl
@@ -74,7 +74,7 @@ initMixtures!(mixtures,X,minVariance=0.25)
 # New test
 # ==================================
 println("Testing gmm...")
-clusters = gmm([1 10.5;1.5 missing; 1.8 8; 1.7 15; 3.2 40; missing missing; 3.3 38; missing -2.3; 5.2 -2.4],3,verbosity=NONE)
+clusters = gmm([1 10.5;1.5 missing; 1.8 8; 1.7 15; 3.2 40; missing missing; 3.3 38; missing -2.3; 5.2 -2.4],3,verbosity=NONE, initStrategy="grid")
 @test isapprox(clusters.BIC,119.04816608007282)
 
 # ==================================
