+++ src/BetaML.jl
@@ -49,8 +49,9 @@ using .Clustering
 #MLJ interface...
 
 const MLJ_PERCEPTRON_MODELS = (PerceptronClassifier, KernelPerceptronClassifier, PegasosClassifier)
-const MLJ_TREES_MODELS = (DecisionTreeClassifier, DecisionTreeRegressor, RandomForestClassifier, RandomForestRegressor)
-const MLJ_INTERFACED_MODELS = (MLJ_PERCEPTRON_MODELS..., MLJ_TREES_MODELS...)
+const MLJ_TREES_MODELS      = (DecisionTreeClassifier, DecisionTreeRegressor, RandomForestClassifier, RandomForestRegressor)
+const MLJ_CLUSTERING_MODELS = (KMeans, KMedoids, GMM, MissingImputator)
+const MLJ_INTERFACED_MODELS = (MLJ_PERCEPTRON_MODELS..., MLJ_TREES_MODELS..., MLJ_CLUSTERING_MODELS...)
 
 function __init__()
     MMI.metadata_pkg.(MLJ_INTERFACED_MODELS,
