### 2021-03-10 3dd36e3
Renamed notebooks to example and docs/src/Notebooks.md to Examples.md
Added Literate dependence to the docs folder Project.toml
