+++ docs/src/index.md
@@ -2,6 +2,16 @@
 
 Welcome to the documentation of the [_Beta Machine Learning toolkit_](https://github.com/sylvaticus/BetaML.jl).
 
+## About
+
+The `BetaML` toolkit provides classical algorithms written in the Julia programming language useful to "learn" the relationship between some inputs and some outputs, with the objective to make accurate predictions of the output given new inputs ("supervised machine learning") or to better understand the structure of the data, perhaps hidden because of the high dimensionality ("unsupervised machine learning").
+
+While specific packages exist for state-of-the art implementations of these algorithms (see the section "[Alternative Packages](https://github.com/sylvaticus/BetaML.jl#alternative-packages)"), thanks to the Just-In-Time compilation nature of Julia, `BetaML` is reasonably fast for datasets that fits in memory.
+
+Aside the algorithms themselves, `BetaML` provides many "utility" functions. Because algorithms are all self-contained in the library itself (you are invited to explore their source code by typing `@edit functionOfInterest(par1,par2)`), the utility functions have APIs that are coordinated with the algorithms, facilitating the "preparation" of the data for the analysis, the evaluation of the models or the implementation of several models in chains (pipelines).
+While `BetaML` doesn't provide itself tools for hyper-parameters optimisation or complex pipeline building tools, most models have an interface for the [`MLJ`](https://github.com/alan-turing-institute/MLJ.jl) framework that allows it.
+
+Aside Julia, BetaML can be accessed in R or Python using respectively [JuliaCall](https://github.com/Non-Contradiction/JuliaCall) and [PyJulia](https://github.com/JuliaPy/pyjulia). See [here](https://syl1.gitbook.io/julia-language-a-concise-tutorial/language-core/interfacing-julia-with-other-languages#use-julia-in-python) for a tutorial or the examples for some actual use of the _Beta Machine Learning Toolkit_ in R/Python.
 
 ## Installation
 
