+++ docs/Project.toml
@@ -1,21 +1,22 @@
 [deps]
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
+PDMats = "90014a1f-27ba-587c-ab20-58faa44d9150"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-PDMats = "90014a1f-27ba-587c-ab20-58faa44d9150"
-Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 
 [compat]
-julia = "1.4"
+Distributions = "0.23"
+Documenter = "0.25"
+PDMats = "0.9"
 ProgressMeter = "1.3"
 Reexport = "0.2"
 Zygote = "0.4"
-Documenter = "0.25"
-PDMats = "0.9"
-Distributions = "0.23"
+julia = "1.4"
