### 2021-03-11 655461e
Renamed (again) /examples -> /docs/src/tutorials, started random ri-thinking
1. Examples has been moved to /docs/src/tutorials. Related links are still all broken and the actual "tutorial" need to be rewritten.
There are also a few isues remaining in building the documentation

2. The constants FIXEDSEED and FIXEDRNG has been added to Utils, and the package StableRNGs has been added as dependency.
The new constants however has not yet be used yet, nor any algorithm has been yet modified
