+++ docs/src/tutorials/Getting started/betaml_tutorial_getting_started.md
@@ -1,0 +1,23 @@
+```@meta
+EditURL = "<unknown>/docs/src/tutorials/Getting started/betaml_tutorial_getting_started.jl"
+```
+
+## This is markdown title
+This is also markdown
+
+```@example betaml_tutorial_getting_started
+# This is a normal comment
+```
+
+```@example betaml_tutorial_getting_started
+a = 1
+b = a + 1
+println(b)
+```
+
+[View this file on Github](<unknown>/docs/src/tutorials/Getting started/betaml_tutorial_getting_started.jl).
+
+---
+
+*This page was generated using [Literate.jl](https://github.com/fredrikekre/Literate.jl).*
+
