+++ Project.toml
@@ -14,6 +14,7 @@ PDMats = "90014a1f-27ba-587c-ab20-58faa44d9150"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
+StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
@@ -28,3 +29,4 @@ ProgressMeter = "1.3"
 Reexport = "0.2, 1.0"
 Zygote = "0.4, 0.5, 0.6"
 julia = "1.3.1"
+StableRNGs = "0.1, 1"
