### 2021-03-11 75ea785
Added parameter rng to all stochastic algorithms used  on tests
Use it with:
- `myAlgorithm(;rng=FIXEDRNG)`             # always produce the same sequence of results
on each run of the script ("pulling" from the same rng object on different calls)
- `myAlgorithm(;rng=StableRNG(FIXEDSEED))` # always produce the same result
(new rng object on each call)

Tests use the first version
