+++ src/Perceptron.jl
@@ -62,9 +62,10 @@ The perceptron is a _linear_ classifier. Multiclass is supported using a one-vs-
 * `T`:           Maximum number of iterations across the whole set (if the set
                  is not fully classified earlier) [def: 1000]
 * `nMsg`:        Maximum number of messages to show if all iterations are done [def: `0`]
-* `shuffle`:    Whether to randomly shuffle the data at each iteration [def: `false`]
+* `shuffle`:     Whether to randomly shuffle the data at each iteration [def: `false`]
 * `forceOrigin`: Whether to force `θ₀` to remain zero [def: `false`]
 * `returnMeanHyperplane`: Whether to return the average hyperplane coefficients instead of the final ones  [def: `false`]
+* `rng`:         Random Number Generator (@see Utils.FIXEDSEED) [deafult: `Random.GLOBAL_RNG`]
 
 # Return a named tuple with:
 * `θ`:          The weights of the classifier
@@ -81,7 +82,7 @@ julia> model = perceptron([1.1 2.1; 5.3 4.2; 1.8 1.7], [-1,1,-1])
 julia> ŷ     = predict([2.1 3.1; 7.3 5.2], model.θ, model.θ₀, model.classes)
 ```
 """
-function perceptron(x::AbstractMatrix, y::AbstractVector; θ=zeros(size(x,2)),θ₀=0.0, T=1000, nMsgs=0, shuffle=false, forceOrigin=false, returnMeanHyperplane=false)
+function perceptron(x::AbstractMatrix, y::AbstractVector; θ=zeros(size(x,2)),θ₀=0.0, T=1000, nMsgs=0, shuffle=false, forceOrigin=false, returnMeanHyperplane=false, rng = Random.GLOBAL_RNG)
     yclasses = unique(y)
     nCl      = length(yclasses)
     if nCl == 2
@@ -93,7 +94,7 @@ function perceptron(x::AbstractMatrix, y::AbstractVector; θ=zeros(size(x,2)),θ
     end
     for (i,c) in enumerate(yclasses)
         ybin = ((y .== c) .*2 .-1)  # conversion to -1/+1
-        outBinary = perceptronBinary(x, ybin; θ=θ,θ₀=θ₀, T=T, nMsgs=nMsgs, shuffle=shuffle, forceOrigin=forceOrigin)
+        outBinary = perceptronBinary(x, ybin; θ=θ,θ₀=θ₀, T=T, nMsgs=nMsgs, shuffle=shuffle, forceOrigin=forceOrigin, rng=rng)
         if returnMeanHyperplane
             outθ[i]  = outBinary.avgθ
             outθ₀[i] = outBinary.avgθ₀
@@ -124,6 +125,7 @@ Train the binary classifier "perceptron" algorithm based on x and y (labels)
 * `nMsg`:        Maximum number of messages to show if all iterations are done
 * `shuffle`:     Whether to randomly shuffle the data at each iteration [def: `false`]
 * `forceOrigin`: Whether to force `θ₀` to remain zero [def: `false`]
+* `rng`:         Random Number Generator (@see Utils.FIXEDSEED) [deafult: `Random.GLOBAL_RNG`]
 
 # Return a named tuple with:
 * `θ`:          The final weights of the classifier
@@ -143,7 +145,7 @@ Train the binary classifier "perceptron" algorithm based on x and y (labels)
 julia> model = perceptronBinary([1.1 2.1; 5.3 4.2; 1.8 1.7], [-1,1,-1])
 ```
 """
[…]
