### 2021-03-12 e73fac8
Added `rng` parameter info  in the various docstrings
