### 2021-03-12 7724c0a
Reimplemented GMM as supervised model
