+++ src/Clustering_MLJ.jl
@@ -60,7 +60,7 @@ GMM(;
     rng           = Random.GLOBAL_RNG,
 ) = GMM(K,p₀,mixtures, tol, minVariance, minCovariance,initStrategy,rng)
 
-mutable struct MissingImputator{TM <: AbstractMixture} <: MMI.Static
+mutable struct MissingImputator{TM <: AbstractMixture} <: MMI.Unsupervised
     K::Int64
     p₀::Union{Nothing,AbstractArray{Float64,1}}
     mixtures::AbstractArray{TM,1}
@@ -100,6 +100,15 @@ function MMI.fit(m::GMM, verbosity, X, y)
     # X is nothing, y is the data: https://alan-turing-institute.github.io/MLJ.jl/dev/adding_models_for_general_use/#Models-that-learn-a-probability-distribution-1
     y          = MMI.matrix(y) # convert table to matrix
     res        = gmm(y,m.K,p₀=m.p₀,mixtures=m.mixtures, minVariance=m.minVariance, minCovariance=m.minCovariance,initStrategy=m.initStrategy,verbosity=NONE)
+    fitResults = (res.pₙₖ,res.pₖ,res.mixtures)
+    cache      = nothing
+    report     = (res.ϵ,res.lL,res.BIC,res.AIC)
+    return (fitResults, cache, report)
+end
+
+function MMI.fit(m::MissingImputator, verbosity, X)
+    x          = MMI.matrix(X) # convert table to matrix
+    res        = gmm(x,m.K,p₀=m.p₀,mixtures=m.mixtures, minVariance=m.minVariance, minCovariance=m.minCovariance,initStrategy=m.initStrategy,verbosity=NONE)
     fitResults = (res.pₙₖ,res.pₖ,res.mixtures)
     cache      = nothing
     report     = (res.ϵ,res.lL,res.BIC,res.AIC)
@@ -148,6 +157,17 @@ function MMI.predict(m::GMM, fitResults, X)
     return predictions
 end
 
+""" predict(m::MissingImputator, fitResults, X) - Given a trained imputator model fill the missing data of some new observations"""
+function MMI.predict(m::MissingImputator, fitResults, X)
+    x               = MMI.matrix(X) # convert table to matrix
+    (N,D)           = size(x)
+    (pₙₖ,pₖ,mixtures) = fitResults
+    nCl             = length(pₖ)
+    # Fill the missing data of this "new X" using the mixtures computed in the fit stage
+    xout         = predictMissing(x,nCl,p₀=pₖ,mixtures=mixtures,tol=m.tol,verbosity=NONE,minVariance=m.minVariance,minCovariance=m.minCovariance,initStrategy="given",maxIter=1)
+    return MMI.table(xout.X̂)
+end
+
 """ transform(m::GMM, fitResults, X) - Given a trained clustering model and some observations, predict the class of the observation"""
 function MMI.transform(m::GMM, fitResults, X)
     return MMI.predict(m::GMM, fitResults, X)
@@ -188,7 +208,7 @@ MMI.metadata_model(GMM,
 )
 
 MMI.metadata_model(MissingImputator,
-    input_scitype    = MMI.Table(MMI.Continuous,MMI.Missing),
+    input_scitype    = MMI.Table(Union{MMI.Continuous,MMI.Missing}),
     output_scitype   = MMI.Table(MMI.Continuous),     # for an unsupervised, what output?
     supports_weights = false,                         # does the model support sample weights?
     descr            = "Impute missing values using an Expectation-Maximisation clustering algorithm, from the Beta Machine Learning Toolkit (BetaML).",
