### 2021-03-15 4ada16c
Solved problem with random number generator and multithread (buildForest)
New Utils.generateParallelRngs() function that generate new rangs, one per thread.
It is used in buildForest
However it doesn't work with StableRNGs
Hence the FIXEDRNG constant has been reverted to MersenneTwister(FIXEDSEED)

All tests now use rng=copy(TESTRNG) where TESTRNG is just an alias of FIXEDRNG. I am two minds if keeping the tests using StableRNG
and making an exception for Random Forests or use directly MersenneTwister for everything.
In case it will be enought to change TESTRNG
