+++ docs/src/tutorials/A regression task: sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.jl
@@ -1,8 +1,175 @@
 
-# ## This is markdown title
-# This is also markdown
-## This is a normal comment
+# # A regression task: the prediction of  bike  sharing demand
+# The task is to estimate the influence of several variables (like the weather, the season, the day of the week..) on the demand of shared bicycles, so that the authority in charge of the service can organise the service in the best way.
 #
-a = 1
-b = a + 1
-println(b)
+# Data origin:
+# - original full dataset (by hour, not used here): https://archive.ics.uci.edu/ml/datasets/Bike+Sharing+Dataset
+# - simplified dataset (by day, with some simple scaling): https://www.hds.utc.fr/~tdenoeux/dokuwiki/en/aec
+# - description: https://www.hds.utc.fr/~tdenoeux/dokuwiki/_media/en/exam_2019_ace_.pdf
+# - data: https://www.hds.utc.fr/~tdenoeux/dokuwiki/_media/en/bike_sharing_day.csv.zip
+
+# Note that even if we are estimating a time serie, we are not using here a recurrent neural network as we assume the temporal dependence to be negligible (i.e. $Y_t = f(X_t)$ alone).
+
+## Library and data loading
+
+using LinearAlgebra, Random, DataFrames, CSV, StatsPlots, StableRNGs, BetaML.Perceptron, BetaML.Nn, BetaML.Trees
+import Distributions: Uniform
+
+using BetaML  # only needed to get the relative path in importing data
+baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","A regression task: sharing bike demand prediction")
+
+# Data loading and cleaning..
+data = CSV.File(joinpath(baseDir,"data","bike_sharing_day.csv"),delim=',') |> DataFrame
+describe(data)
+
+The variable we want to learn to predict is `cnt`, the total demand of bikes for a given day. Even if it is indeed an integer, we treat it as a continuous variale.
+
+
+
+## Data preparation
+The first step is to prepare the data for the analysis. This indeed depends already on the model we want to employ, as some models "accept" everything as input, no matter if the data is numerical or categorical, if it has missing values or not... Other models are instead more exigents, and require more or less work.
+Here we start with a Decision Tree and a Random Forest models that belong to the first category, those of the models for which any kind of input is fine.
+Hence, the only things we have to do is to select the variables in input and those representing our output:
+
+x    = convert(Matrix,hcat(data[:,[:instant,:season,:yr,:mnth,:holiday,:weekday,:workingday,:weathersit,:temp,:atemp,:hum,:windspeed]]))
+y    = data[:,16]
+
+We can now split the dataset between the data we will use for training the algorithm (xtrain/ytrain), those for selecting the hyperparameters (xval/yval) and finally those for testing the quality of the algoritm with the optimal hyperparameters (xtest/ytest). For doing that we use the function `Utils.partition` specifying the share we want to use for these shares. As our data represent indeed a time serie, we want our model to be able to predict _future_ demand of bike sharing from _past_, observed rented bikes, so we do not shuffle the datasets as it would be by default.
+
[…]
