### 2021-03-16 7b75b45
Some Project.toml updates and documentation
Revoved dependency from StableRNGs, set version to 0.5
