+++ docs/src/tutorials/A regression task: sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.jl
@@ -10,10 +10,11 @@
 
 # Note that even if we are estimating a time serie, we are not using here a recurrent neural network as we assume the temporal dependence to be negligible (i.e. $Y_t = f(X_t)$ alone).
 
-## Library and data loading
+# ## Library and data loading
 
 using LinearAlgebra, Random, DataFrames, CSV, StatsPlots, StableRNGs, BetaML.Perceptron, BetaML.Nn, BetaML.Trees
 import Distributions: Uniform
+using DecisionTree # for comparision
 
 using BetaML  # only needed to get the relative path in importing data
 baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","A regression task: sharing bike demand prediction")
@@ -22,60 +23,84 @@ baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","A regr
 data = CSV.File(joinpath(baseDir,"data","bike_sharing_day.csv"),delim=',') |> DataFrame
 describe(data)
 
-The variable we want to learn to predict is `cnt`, the total demand of bikes for a given day. Even if it is indeed an integer, we treat it as a continuous variale.
+# The variable we want to learn to predict is `cnt`, the total demand of bikes for a given day. Even if it is indeed an integer, we treat it as a continuous variable, so each single prediction will be a scalar $$Y \in R$$.
 
 
-
-## Data preparation
-The first step is to prepare the data for the analysis. This indeed depends already on the model we want to employ, as some models "accept" everything as input, no matter if the data is numerical or categorical, if it has missing values or not... Other models are instead more exigents, and require more or less work.
-Here we start with a Decision Tree and a Random Forest models that belong to the first category, those of the models for which any kind of input is fine.
-Hence, the only things we have to do is to select the variables in input and those representing our output:
+# ## Data preparation
+# The first step is to prepare the data for the analysis. This indeed depends already on the model we want to employ, as some models "accept" everything as input, no matter if the data is numerical or categorical, if it has missing values or not... Other models are instead more exigents, and require more or less work.
+# Here we use Decision Tree and Random Forest models that belong to the first category, those of the models for which any kind of input is fine.
+# Hence, the only things we have to do is to select the variables in input and those representing our output:
 
 x    = convert(Matrix,hcat(data[:,[:instant,:season,:yr,:mnth,:holiday,:weekday,:workingday,:weathersit,:temp,:atemp,:hum,:windspeed]]))
 y    = data[:,16]
 
-We can now split the dataset between the data we will use for training the algorithm (xtrain/ytrain), those for selecting the hyperparameters (xval/yval) and finally those for testing the quality of the algoritm with the optimal hyperparameters (xtest/ytest). For doing that we use the function `Utils.partition` specifying the share we want to use for these shares. As our data represent indeed a time serie, we want our model to be able to predict _future_ demand of bike sharing from _past_, observed rented bikes, so we do not shuffle the datasets as it would be by default.
[…]
