+++ docs/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+DecisionTree = "7806a523-6efd-50cb-b5f6-3fa6f1930dbb"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
@@ -19,4 +20,4 @@ PDMats = "0.9"
 ProgressMeter = "1.3"
 Reexport = "0.2"
 Zygote = "0.4"
-julia = "1.4"
+julia = "1.5"
