### 2021-03-17 fd70f66
Restructured the loadiang of the module so that "using BataML" is enought
It is no longer necessary to load BetaML.Nn, BetaML.Trees, etcc.

- new ForceImport module
- removed StableRNGs (problems with multithreads)
- new file "Api.jl" with the names used in multiple (sub)modules

Todo: update the documentation to just use "using BetaML"
