+++ test/Utils_tests.jl
@@ -1,7 +1,7 @@
 using Test, Statistics, CategoricalArrays, Random
 #using StableRNGs
 #rng = StableRNG(123)
-using BetaML.Utils
+using BetaML
 
 TESTRNG = FIXEDRNG # This could change...
 
@@ -289,7 +289,7 @@ m2 = convert(Array{Float64,2},[41:50 51:60])
 m3 = makeMatrix(collect(61:70))
 m4 = collect(71:80)
 parts = [0.33,0.27,0.4]
-out = Utils.partition([m1,m2,m3,m4],parts,shuffle=true,rng=copy(TESTRNG))
+out = partition([m1,m2,m3,m4],parts,shuffle=true,rng=copy(TESTRNG))
 @test size(out,1) == 4 && size(out[1][3]) == (4,3)
 x = [1:10 11:20]
 y = collect(31:40)
