+++ test/Nn_tests.jl
@@ -3,7 +3,7 @@ using DelimitedFiles, LinearAlgebra, MLDatasets
 
 #using StableRNGs
 #rng = StableRNG(123)
-using BetaML.Nn
+using BetaML
 
 TESTRNG = FIXEDRNG # This could change...
 
@@ -40,7 +40,7 @@ mynn = buildNetwork([l1,l2],squaredCost,name="Simple Multinomial logistic regres
 o1 = forward(l1,x)
 #@code_warntype forward(l1,x)
 o2 = forward(l2,o1)
-orig = Nn.predict(mynn,x')[1,:]
+orig = predict(mynn,x')[1,:]
 #@code_warntype Nn.predict(mynn,x')[1,:]
 ϵ = squaredCost(o2,y)
 #@code_warntype  squaredCost(o2,y)
@@ -66,7 +66,7 @@ l1dw = getGradient(l1,x,dϵ_do1)
 #@code_warntype
 dw = getGradient(mynn,x,y)
 #@code_warntype getGradient(mynn,x,y)
-y_deltax1 = Nn.predict(mynn,[x[1]+0.001 x[2]])[1,:]
+y_deltax1 = predict(mynn,[x[1]+0.001 x[2]])[1,:]
 #@code_warntype Nn.predict(mynn,[x[1]+0.001 x[2]])[1,:]
 lossDeltax1 = loss(mynn,[x[1]+0.001 x[2]],y')
 #@code_warntype
@@ -98,7 +98,7 @@ for i in 1:10000
     setParams!(mynn,w)
 end
 lossFinal = loss(mynn,x',y')
-@test Nn.predict(mynn,x')[1,1]>0.96
+@test predict(mynn,x')[1,1]>0.96
 setParams!(mynn,origW)
 train!(mynn,x',y',epochs=10000,batchSize=1,sequential=true,verbosity=NONE,optAlg=SGD(η=t->η,λ=1),rng=copy(TESTRNG))
 #@code_warntype train!(mynn,x',y',epochs=10000,batchSize=1,sequential=true,verbosity=NONE,optAlg=SGD(η=t->η,λ=1))
@@ -126,7 +126,7 @@ train!(mynn,xtrain,ytrain,batchSize=1,sequential=true,epochs=100,verbosity=NONE,
 avgLoss = loss(mynn,xtest,ytest)
 @test  avgLoss ≈ 1.599729991966362
 expectedOutput = [0.7360644412052633, 0.7360644412052633, 0.7360644412052633, 2.47093434438514]
-predicted = dropdims(Nn.predict(mynn,xtest),dims=2)
+predicted = dropdims(predict(mynn,xtest),dims=2)
 @test any(isapprox(expectedOutput,predicted))
 
 # With the ADAM optimizer...
@@ -138,7 +138,7 @@ train!(mynn,xtrain,ytrain,batchSize=1,sequential=true,epochs=100,verbosity=NONE,
 avgLoss = loss(mynn,xtest,ytest)
 @test  avgLoss ≈ 0.9497779759064725
 expectedOutput = [1.7020525792404175, -0.1074729043392682, 1.4998367847079956, 3.3985794704732717]
-predicted = dropdims(Nn.predict(mynn,xtest),dims=2)
+predicted = dropdims(predict(mynn,xtest),dims=2)
 @test any(isapprox(expectedOutput,predicted))
 
 # ==================================
@@ -160,8 +160,8 @@ train!(mynn,xtrain,ytrain,epochs=1000,sequential=true,batchSize=1,verbosity=NONE
 avgLoss = loss(mynn,xtest,ytest)
[…]
