+++ docs/src/tutorials/A regression task: sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.jl
@@ -12,9 +12,9 @@
 
 # ## Library and data loading
 
-using LinearAlgebra, Random, DataFrames, CSV, StatsPlots, StableRNGs, BetaML.Perceptron, BetaML.Nn, BetaML.Trees
+using LinearAlgebra, Random, DataFrames, CSV, StatsPlots,  BetaML
 import Distributions: Uniform
-using DecisionTree # for comparision
+import DecisionTree # for comparision
 
 using BetaML  # only needed to get the relative path in importing data
 baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","A regression task: sharing bike demand prediction")
@@ -43,63 +43,71 @@ y    = data[:,16]
 # BetaML doesn't have a diedicated function for hyperparameters optimisation, but it is easy to write some custom julia code, at least for a simple grid-based "search". Indeed the reason that a dedicated function exists in other Machine Learning libraries is that loops in other languages are slow, but this is not a problem in Julia, so you can retain the flexibility to write the kind of hyperparameter tuning that best fits your needs.
 # Below if an example of a possible such function. Note there are more "elegant" ways to code it, but this one does the job. We will see the various functions inside `tuneHyperParameters()` in a moment. For now let's going just to observe that `tuneHyperParameters` just loops over all the possible hyperparameter and select the one where the error between xval and yval is minimised.
 
-function tuneHyperParameters(model,xtrain,ytrain,xval,yval;maxDepthsRange=1:15,maxFeaturesRange=5:12,nTreesRange=20:20,repetitions=5,rng=Random.GLOBAL_RNG)
+function tuneHyperParameters(model,xtrain,ytrain,xval,yval;maxDepthsRange=15:15,maxFeaturesRange=size(xtrain,2):size(xtrain,2),nTreesRange=20:20,βRange=0:0,minRecordsRange=2:2,repetitions=5,rng=Random.GLOBAL_RNG)
     # We start with an infinititly high error
     bestMre         = +Inf
     bestMaxDepth    = 1
     bestMaxFeatures = 1
+    bestMinRecords  = 2
     bestNTrees      = 1
+    bestβ           = 0
     # We loop over all possible hyperparameter combinations...
     for maxDepth in maxDepthsRange
         for maxFeatures in maxFeaturesRange
-            for nTrees in nTreesRange
-               totAttemptError = 0.0
-               # We run several repetitions with the same hyperparameter combination to account for stochasticity...
-               for r in 1:repetitions
-                  if model == "DecisionTree"
-                     # here we train the Decition Tree model
-                     myTrainedModel = buildTree(xtrain,ytrain, maxDepth=maxDepth,maxFeatures=maxFeatures,rng=rng)
-                  else
-                     # Here we train the Random Forest model
-                     myTrainedModel = buildForest(xtrain,ytrain,nTrees,maxDepth=maxDepth,maxFeatures=maxFeatures, rng=rng)
-                  end
-                  # Here we make prediciton with this trained model and we compute its error
-                  ŷval   = Trees.predict(myTrainedModel, xval)
-                  mreVal = meanRelError(ŷval,yval)
-                  totAttemptError += mreVal
-               end
-               avgAttemptedDepthError = totAttemptError / repetitions
[…]
