+++ docs/src/tutorials/A regression task: sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.md
@@ -1,23 +1,0 @@
-```@meta
-EditURL = "<unknown>/src/tutorials/A regression task: sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.jl"
-```
-
-## This is markdown title
-This is also markdown
-
-```@example betaml_tutorial_regression_sharingBikes
-# This is a normal comment
-```
-
-```@example betaml_tutorial_regression_sharingBikes
-a = 1
-b = a + 1
-println(b)
-```
-
-[View this file on Github](<unknown>/src/tutorials/A regression task: sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.jl).
-
----
-
-*This page was generated using [Literate.jl](https://github.com/fredrikekre/Literate.jl).*
-
