### 2021-03-19 39a2ca0
Working on regression tutorial, added factors to `integerEncoder` and allowed non-numerical as input of `oneHotEncoder`
