+++ src/Utils.jl
@@ -94,6 +94,7 @@ Use it with `rngs = generateParallelRngs(rng,Threads.nthreads())`
 
 """
 function generateParallelRngs(rng::AbstractRNG, n::Integer)
+    # see also here for a workaround that would allow StableRNGs back: https://github.com/JuliaRandom/StableRNGs.jl/issues/8#issuecomment-801892721
     step = rand(rng,big(10)^20:big(10)^40) # making the step random too !
     rngs = Vector{Union{MersenneTwister,Random._GLOBAL_RNG}}(undef, n) # Vector{typeof(rng)}(undef, n)
     rngs[1] = copy(rng)
@@ -109,7 +110,7 @@ end
 import Base.reshape
 """ reshape(myNumber, dims..) - Reshape a number as a n dimensional Array """
 reshape(x::T, dims...) where {T <: Number} =   (x = [x]; reshape(x,dims) )
-makeColVector(x::T) where {T <: Number} =  [x]
+makeColVector(x::T) where {T} =  [x]
 makeColVector(x::T) where {T <: AbstractArray} =  reshape(x,length(x))
 makeRowVector(x::T) where {T <: Number} = return [x]'
 makeRowVector(x::T) where {T <: AbstractArray} =  reshape(x,1,length(x))
@@ -148,36 +149,92 @@ function getPermutations(v::AbstractArray{T,1};keepStructure=false) where {T}
 end
 
 
-function oneHotEncoderRow(y,d;count = false)
-    y = makeColVector(y)
+""" singleUnique(x) Return the unique values of x whether x is an array of arrays, an array or a scalar"""
+function singleUnique(x::Union{T,AbstractArray{T}}) where {T <: Union{Any,AbstractArray{T2}} where T2 <: Any }
+    if typeof(x) <: AbstractArray{T2} where {T2 <: AbstractArray}
+        return unique(vcat(unique.(x)...))
+    elseif typeof(x) <: AbstractArray{T2} where {T2}
+        return unique(x)
+    else
+        return [x]
+    end
+end
+
+function oneHotEncoderRow(x::Union{AbstractArray{T},T}; d=maximum(x), factors=1:d,count = false) where {T <: Integer}
+    x = makeColVector(x)
     out = zeros(Int64,d)
-    for j in y
+    for j in x
         out[j] = count ? out[j] + 1 : 1
     end
     return out
 end
-"""
-    oneHotEncoder(y,d;count)
 
-Encode arrays (or arrays of arrays) of integer data as 0/1 matrices
+function oneHotEncoderRow(x::Union{AbstractArray{T},T};factors=singleUnique(x),d=length(factors),count = false) where {T}
+    x = makeColVector(x)
+    return oneHotEncoderRow(integerEncoder(x;factors=factors),d=length(factors);count=count)
+end
+
+
+
+
+"""
+    oneHotEncoder(x;d,factors,count)
+
+Encode arrays (or arrays of arrays) of categorical data as matrices of one column per factor.
+
+The case of arrays of arrays is for when at each record you have more than one categorical output. You can then decide to encode just the presence of the factors or their counting
 
 # Parameters:
-- `y`: The data to convert (integer, array or array of arrays of integers)
-- `d`: The number of dimensions in the output matrik. [def: `maximum(maximum.(Y))`]
-- `count`: Wether to count multiple instances on the same dimension/record or indicate just presence. [def: `false`]
+- `x`: The data to convert (array or array of arrays)
+- `d`: The number of dimensions in the output matrix [def: `maximum(x)` for integers and `length(factors)` otherwise]
+- `factors`: The factors from which to encode [def: `1:d` for integer x or `unique(x)` otherwise]
[…]
