### 2021-03-19 fe19dc5
Changed default initStrategy for KMeans to random, MLJ Interface clusters improvements
- the grid initStrategy should be refactored to avoid starting centers at the extremes
- in `MMI.predict(m::Union{KMeans,KMedoids},...)` the output categorical array contain the inforlation of all the factors, not just those returned
- updated metadata_model of KMeans and KMedoids
