### 2021-03-19 cf529ba
Wirking on cluster initialisations, expecially for MLJ
- MLJ:KMeans and MLJ:KMedoids default to "shuffle" initStrategy
- improved "grid" initStrategy, now all centers have the same initial "space"
- added the `rng=m.rng` parameter to MLJ cluster training functions
