### 2021-03-19 863113c
Refactored generateParallelRngs and went back to StableRNGs for testing
- generateParallelRngs is now independent on the specific RNG used, as it uses just seed!() that is implemented in almost all RNGs.
- test now switched back to `StableRND(123)`
- still, some problems with the MLJ interface on Julia 1.5 !
- removed module Future
