### 2021-03-22 29994d4
Readded StableRNG, doc (tutorial) and MLJ GMM Interface
- readdd StableRNG so that the toolkit provide a guaranteed fixed RNG. FIXEDRNG redefault to StableRNG(123)
- working on multithread the hyperparameter tuning function in the doc
- changed MLJ interface for GMM, from "fit return the mixtures, predict use them to predict post prob from them and new or old feature" to "fit provide the prob, predict
just wrap them"
