+++ docs/src/tutorials/A regression task: sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.jl
@@ -12,10 +12,12 @@
 
 # ## Library and data loading
 
-using LinearAlgebra, Random, Statistics, DataFrames, CSV, Plots, BetaML
+using LinearAlgebra, Random, Statistics, DataFrames, CSV, Plots, BetaML,StableRNGs
 import Distributions: Uniform
 import DecisionTree # for comparision
 using Test     #src
+
+FIXEDRNG =  StableRNG(123)
 
 baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","A regression task: sharing bike demand prediction")
 
@@ -44,7 +46,7 @@ y    = data[:,16]
 # We can now "tune" our model so-called hyperparameters, i.e. choose the best exogenous parameters of our algorithm, where "best" refer to some minimisation of a "loss" function between the true and the predicted value, where a dedicated "validation" subset of the input is used (xval and yval) in our case.
 # BetaML doesn't have a diedicated function for hyperparameters optimisation, but it is easy to write some custom julia code, at least for a simple grid-based "search". Indeed the reason that a dedicated function exists in other Machine Learning libraries is that loops in other languages are slow, but this is not a problem in Julia, so we can retain the flexibility to write the kind of hyperparameter tuning that best fits our needs.
 # Below if an example of a possible such function. Note there are more "elegant" ways to code it, but this one does the job. We will see the various functions inside `tuneHyperParameters()` in a moment. For now let's going just to observe that `tuneHyperParameters` just loops over all the possible hyperparameter and select the one where the error between xval and yval is minimised. For the meaning of the various hyperparameter, consult the documentation of the `buildTree` and `buildForest` functions.
-function tuneHyperParameters(model,xtrain,ytrain,xval,yval;maxDepthsRange=15:15,maxFeaturesRange=size(xtrain,2):size(xtrain,2),nTreesRange=20:20,βRange=0:0,minRecordsRange=2:2,repetitions=5,rng=Random.GLOBAL_RNG)
+function tuneHyperParameters(model,xtrain,ytrain,xval,yval;maxDepthRange=15:15,maxFeaturesRange=size(xtrain,2):size(xtrain,2),nTreesRange=20:20,βRange=0:0,minRecordsRange=2:2,repetitions=5,rng=Random.GLOBAL_RNG)
     ## We start with an infinititly high error
     bestMre         = +Inf
     bestMaxDepth    = 1
@@ -52,51 +54,90 @@ function tuneHyperParameters(model,xtrain,ytrain,xval,yval;maxDepthsRange=15:15,
     bestMinRecords  = 2
     bestNTrees      = 1
     bestβ           = 0
+    compLock  = ReentrantLock()
+
     ## We loop over all possible hyperparameter combinations...
-    for maxDepth in maxDepthsRange
-        for maxFeatures in maxFeaturesRange
-            for minRecords in minRecordsRange
-                for nTrees in nTreesRange
-                   for β in βRange
-                       totAttemptError = 0.0
-                       ## We run several repetitions with the same hyperparameter combination to account for stochasticity...
-                       for r in 1:repetitions
-                          if model == "DecisionTree"
-                             ## Here we train the Decition Tree model
-                             myTrainedModel = buildTree(xtrain,ytrain, maxDepth=maxDepth,maxFeatures=maxFeatures,minRecords=minRecords,rng=rng)
-                          else
-                             ## Here we train the Random Forest model
[…]
