+++ test/Clustering_tests.jl
@@ -1,15 +1,15 @@
 using Test
 #using DelimitedFiles, LinearAlgebra
 
-using StableRNGs
+#using StableRNGs
 #rng = StableRNG(123)
 
 import MLJBase
 const Mlj = MLJBase
 using BetaML
 
-#TESTRNG = FIXEDRNG # This could change...
-TESTRNG = StableRNG(123)
+TESTRNG = FIXEDRNG # This could change...
+#TESTRNG = StableRNG(123)
 
 println("*** Testing Clustering...")
 
@@ -115,8 +115,8 @@ acc = BetaML.accuracy(Mlj.levelcode.(yhat),Mlj.levelcode.(y),ignoreLabels=true)
 model                       =  GMM(rng=copy(TESTRNG))
 modelMachine                =  Mlj.machine(model, nothing, X) # DimensionMismatch
 (fitResults, cache, report) =  Mlj.fit(model, 0, nothing, X)
-yhat_prob                   =  Mlj.transform(model,fitResults,X)
-yhat_prob                   =  Mlj.predict(model, fitResults, X)
+yhat_prob                   =  Mlj.transform(model,fitResults,nothing) # Mlj.transform(model,fitResults,X)
+yhat_prob                   =  Mlj.predict(model, fitResults, nothing)  # Mlj.transform(model,fitResults,X)
 @test length(yhat_prob)     == size(Mlj.matrix(X),1)
 
 X = [1 10.5;1.5 missing; 1.8 8; 1.7 15; 3.2 40; missing missing; 3.3 38; missing -2.3; 5.2 -2.4]
