### 2021-03-22 df133d8
Completed the code (but not the comments) of the regression tutorial
