+++ docs/src/tutorials/A regression task: sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.jl
@@ -12,12 +12,10 @@
 
 # ## Library and data loading
 
-using LinearAlgebra, Random, Statistics, DataFrames, CSV, Plots, BetaML,StableRNGs
+using LinearAlgebra, Random, Statistics, DataFrames, CSV, Plots, BetaML
 import Distributions: Uniform
 import DecisionTree # for comparision
 using Test     #src
-
-FIXEDRNG =  StableRNG(123)
 
 baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","A regression task: sharing bike demand prediction")
 
@@ -44,8 +42,9 @@ y    = data[:,16]
 (ntrain, nval, ntest) = size.([ytrain,yval,ytest],1)
 
 # We can now "tune" our model so-called hyperparameters, i.e. choose the best exogenous parameters of our algorithm, where "best" refer to some minimisation of a "loss" function between the true and the predicted value, where a dedicated "validation" subset of the input is used (xval and yval) in our case.
-# BetaML doesn't have a diedicated function for hyperparameters optimisation, but it is easy to write some custom julia code, at least for a simple grid-based "search". Indeed the reason that a dedicated function exists in other Machine Learning libraries is that loops in other languages are slow, but this is not a problem in Julia, so we can retain the flexibility to write the kind of hyperparameter tuning that best fits our needs.
+# BetaML doesn't have a dedicated function for hyperparameters optimisation, but it is easy to write some custom julia code, at least for a simple grid-based "search". Indeed the reason that a dedicated function exists in other Machine Learning libraries is that loops in other languages are slow, but this is not a problem in julia, so we can retain the flexibility to write the kind of hyperparameter tuning that best fits our needs.
 # Below if an example of a possible such function. Note there are more "elegant" ways to code it, but this one does the job. We will see the various functions inside `tuneHyperParameters()` in a moment. For now let's going just to observe that `tuneHyperParameters` just loops over all the possible hyperparameter and select the one where the error between xval and yval is minimised. For the meaning of the various hyperparameter, consult the documentation of the `buildTree` and `buildForest` functions.
+# The function uses multiple threads, so it calls `generateParallelRngs()` (in the `BetaML.Utils`) to generate thread-safe random number generators and locks the comparision step.
 function tuneHyperParameters(model,xtrain,ytrain,xval,yval;maxDepthRange=15:15,maxFeaturesRange=size(xtrain,2):size(xtrain,2),nTreesRange=20:20,βRange=0:0,minRecordsRange=2:2,repetitions=5,rng=Random.GLOBAL_RNG)
     ## We start with an infinititly high error
     bestMre         = +Inf
@@ -54,12 +53,12 @@ function tuneHyperParameters(model,xtrain,ytrain,xval,yval;maxDepthRange=15:15,m
     bestMinRecords  = 2
     bestNTrees      = 1
     bestβ           = 0
-    compLock  = ReentrantLock()
+    compLock        = ReentrantLock()
 
-    ## We loop over all possible hyperparameter combinations...
+    ## Generate one random number generator per thread
     rngs = generateParallelRngs(rng,Threads.nthreads())
 
-    #for (maxDepth,maxFeatures,minRecords,nTrees,β) in Iterators.product(maxDepthRange,maxFeaturesRange,minRecordsRange,nTreesRange,βRange)
+    ## We loop over all possible hyperparameter combinations...
     Threads.@threads for ij in CartesianIndices((length(maxDepthRange),length(maxFeaturesRange),length(minRecordsRange),length(nTreesRange),length(βRange)))
[…]
