### 2021-03-22 1e92cd4
DOC: github workflow set to 4 threads, resetted manifest/simplified Project.toml for the doc foder, renamed tutorual folders
