+++ docs/make.jl
@@ -6,8 +6,8 @@ push!(LOAD_PATH,"../src/")
 const _TUTORIAL_DIR = joinpath(@__DIR__, "src", "tutorials")
 const _TUTORIAL_SUBDIR = [
     "Getting started",
-    "A regression task: sharing bike demand prediction",
-    "A classification task when labels are known: determining the plant species giving floreal measures",
+    "A regression task - sharing bike demand prediction",
+    "A classification task when labels are known - determining the plant species giving floreal measures",
 ]
 
 function link_example(content)
