+++ docs/src/tutorials/A regression task - sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.md
@@ -1,0 +1,515 @@
+```@meta
+EditURL = "<unknown>/docs/src/tutorials/ARegressionTask_sharingBikeDemandPrediction/betaml_tutorial_regression_sharingBikes.jl"
+```
+
+# A regression task: the prediction of  bike  sharing demand
+The task is to estimate the influence of several variables (like the weather, the season, the day of the week..) on the demand of shared bicycles, so that the authority in charge of the service can organise the service in the best way.
+
+Data origin:
+- original full dataset (by hour, not used here): [https://archive.ics.uci.edu/ml/datasets/Bike+Sharing+Dataset](https://archive.ics.uci.edu/ml/datasets/Bike+Sharing+Dataset)
+- simplified dataset (by day, with some simple scaling): [https://www.hds.utc.fr/~tdenoeux/dokuwiki/en/aec]( https://www.hds.utc.fr/~tdenoeux/dokuwiki/en/aec)
+- description: [https://www.hds.utc.fr/~tdenoeux/dokuwiki/_media/en/exam_2019_ace_.pdf](https://www.hds.utc.fr/~tdenoeux/dokuwiki/_media/en/exam_2019_ace_.pdf)
+- data: [https://www.hds.utc.fr/~tdenoeux/dokuwiki/_media/en/bike_sharing_day.csv.zip](https://www.hds.utc.fr/~tdenoeux/dokuwiki/_media/en/bike_sharing_day.csv.zip)
+
+Note that even if we are estimating a time serie, we are not using here a recurrent neural network as we assume the temporal dependence to be negligible (i.e. $Y_t = f(X_t)$ alone).
+
+## Library and data loading
+
+```@example betaml_tutorial_regression_sharingBikes
+using LinearAlgebra, Random, Statistics, DataFrames, CSV, Plots, BetaML
+import Distributions: Uniform
+import DecisionTree ## For comparisions
+
+baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","ARegressionTask_sharingBikeDemandPrediction")
+```
+
+Data loading
+
+```@example betaml_tutorial_regression_sharingBikes
+data = CSV.File(joinpath(baseDir,"data","bike_sharing_day.csv"),delim=',') |> DataFrame
+describe(data)
+```
+
+The variable we want to learn to predict is `cnt`, the total demand of bikes for a given day. Even if it is indeed an integer, we treat it as a continuous variable, so each single prediction will be a scalar $Y \in \mathbb{R}$.
+
+## Decision Trees and Random Forests
+
+We start our regression task with Decision Trees
+
+### Data preparation
+The first step is to prepare the data for the analysis. This indeed depends already on the model we want to employ, as some models "accept" everything as input, no matter if the data is numerical or categorical, if it has missing values or not... Other models are instead more exigents, and require more or less work.
+Here we use Decision Tree and Random Forest models that belong to the first category, those of the models for which any kind of input is fine.
+Hence, the only things we have to do is to select the variables in input and those representing our output:
+
+```@example betaml_tutorial_regression_sharingBikes
+x    = convert(Matrix,hcat(data[:,[:instant,:season,:yr,:mnth,:holiday,:weekday,:workingday,:weathersit,:temp,:atemp,:hum,:windspeed]]))
+y    = data[:,16]
+```
+
[…]
