+++ docs/src/tutorials/A classification task when labels are known: determining the plant species giving floreal measures/tutorial_classification_iris.md
@@ -1,23 +1,0 @@
-```@meta
-EditURL = "<unknown>/src/tutorials/A classification task when labels are known: determining the plant species giving floreal measures/tutorial_classification_iris.jl"
-```
-
-## This is markdown title
-This is also markdown
-
-```@example tutorial_classification_iris
-# This is a normal comment
-```
-
-```@example tutorial_classification_iris
-a = 1
-b = a + 1
-println(b)
-```
-
-[View this file on Github](<unknown>/src/tutorials/A classification task when labels are known: determining the plant species giving floreal measures/tutorial_classification_iris.jl).
-
----
-
-*This page was generated using [Literate.jl](https://github.com/fredrikekre/Literate.jl).*
-
