### 2021-03-22 f32bddc
gitub workflow for documentation: corrected bug in setting the option -t
