### 2021-03-25 70f7d69
Added par reSeed to generatePArallelRngs and added non-AD example to the regression tutorial
