### 2021-03-25 5ea564f
Improved initMixtures! for the "grid" option and removed mutation of GMM-based MLJ models in fit
