+++ src/Clustering_MLJ.jl
@@ -3,7 +3,7 @@
 import MLJModelInterface       # It seems that having done this in the top module is not enought
 const MMI = MLJModelInterface  # We need to repoeat it here
 
-export KMeans, KMedoids, GMM, MissingImputator
+export KMeans, KMedoids, GMMFitter, GMMClusterer, MissingImputator
 
 # ------------------------------------------------------------------------------
 # Model Structure declarations..
@@ -39,7 +39,7 @@ KMeans(;
   ) = KMedoids(K,dist,initStrategy,Z₀,rng)
 
 # function gmm(X,K;p₀=nothing,mixtures=[DiagonalGaussian() for i in 1:K],tol=10^(-6),verbosity=STD,minVariance=0.05,minCovariance=0.0,initStrategy="grid")
-mutable struct GMM{TM <: AbstractMixture} <: MMI.Probabilistic
+mutable struct GMMFitter{TM <: AbstractMixture} <: MMI.Probabilistic
   K::Int64
   p₀::Union{Nothing,AbstractArray{Float64,1}}
   mixtures::AbstractArray{TM,1}
@@ -49,7 +49,7 @@ mutable struct GMM{TM <: AbstractMixture} <: MMI.Probabilistic
   initStrategy::String
   rng::AbstractRNG
 end
-GMM(;
+GMMFitter(;
     K             = 3,
     p₀            = nothing,
     mixtures      = [DiagonalGaussian() for i in 1:K],
@@ -58,7 +58,28 @@ GMM(;
     minCovariance = 0.0,
     initStrategy  = "kmeans",
     rng           = Random.GLOBAL_RNG,
-) = GMM(K,p₀,mixtures, tol, minVariance, minCovariance,initStrategy,rng)
+) = GMMFitter(K,p₀,mixtures, tol, minVariance, minCovariance,initStrategy,rng)
+
+mutable struct GMMClusterer{TM <: AbstractMixture} <: MMI.Unsupervised
+  K::Int64
+  p₀::Union{Nothing,AbstractArray{Float64,1}}
+  mixtures::AbstractArray{TM,1}
+  tol::Float64
+  minVariance::Float64
+  minCovariance::Float64
+  initStrategy::String
+  rng::AbstractRNG
+end
+GMMClusterer(;
+    K             = 3,
+    p₀            = nothing,
+    mixtures      = [DiagonalGaussian() for i in 1:K],
+    tol           = 10^(-6),
+    minVariance   = 0.05,
+    minCovariance = 0.0,
+    initStrategy  = "kmeans",
+    rng           = Random.GLOBAL_RNG,
+) = GMMClusterer(K,p₀,mixtures, tol, minVariance, minCovariance,initStrategy,rng)
 
 mutable struct MissingImputator{TM <: AbstractMixture} <: MMI.Unsupervised
     K::Int64
@@ -98,7 +119,7 @@ end
 MMI.fitted_params(model::Union{KMeans,KMedoids}, fitresult) = (centers=fitesult[2], cluster_labels=CategoricalArrays.categorical(fitresults[1]))
 
 
-function MMI.fit(m::GMM, verbosity, X, y)
+function MMI.fit(m::GMMFitter, verbosity, X, y)
     # X is nothing, y is the data: https://alan-turing-institute.github.io/MLJ.jl/dev/adding_models_for_general_use/#Models-that-learn-a-probability-distribution-1
     y          = MMI.matrix(y) # convert table to matrix
[…]
