### 2021-03-26 648de38
Changed mixture options in GMMClusterer MLJ interface
