### 2021-03-26 6ccd5ab
Added RDatasets do docs project and almost ended regression tutorial
