+++ docs/src/tutorials/A regression task - sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.md
@@ -16,21 +16,24 @@ Note that even if we are estimating a time serie, we are not using here a recurr
 ## Library and data loading
 
 ```@example betaml_tutorial_regression_sharingBikes
-using LinearAlgebra, Random, Statistics, DataFrames, StableRNGs, CSV, Plots, BetaML
+using LinearAlgebra, Random, Statistics, DataFrames, CSV, Plots, Pipe, BetaML, BenchmarkTools
 import Distributions: Uniform
-import DecisionTree ## For comparisions
-
-baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","A regression task - sharing bike demand prediction")
+import DecisionTree, Flux ## For comparisions
 ```
 
 Data loading
 
 ```@example betaml_tutorial_regression_sharingBikes
+baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","A regression task - sharing bike demand prediction")
 data = CSV.File(joinpath(baseDir,"data","bike_sharing_day.csv"),delim=',') |> DataFrame
 describe(data)
 ```
 
 The variable we want to learn to predict is `cnt`, the total demand of bikes for a given day. Even if it is indeed an integer, we treat it as a continuous variable, so each single prediction will be a scalar $Y \in \mathbb{R}$.
+
+```@example betaml_tutorial_regression_sharingBikes
+plot(data.cnt, title="Daily bike sharing rents (2Y)", label=nothing)
+```
 
 ## Decision Trees and Random Forests
 
@@ -38,7 +41,7 @@ We start our regression task with Decision Trees
 
 ### Data preparation
 The first step is to prepare the data for the analysis. This indeed depends already on the model we want to employ, as some models "accept" everything as input, no matter if the data is numerical or categorical, if it has missing values or not... Other models are instead more exigents, and require more or less work.
-Here we use Decision Tree and Random Forest models that belong to the first category, those of the models for which any kind of input is fine.
+Here we start using  Decision Tree and Random Forest models that belong to the first category, those of the models for which any kind of input is fine.
 Hence, the only things we have to do is to select the variables in input and those representing our output:
 
 ```@example betaml_tutorial_regression_sharingBikes
@@ -46,7 +49,7 @@ x    = convert(Matrix,hcat(data[:,[:instant,:season,:yr,:mnth,:holiday,:weekday,
 y    = data[:,16]
 ```
 
-We can now split the dataset between the data we will use for training the algorithm (xtrain/ytrain), those for selecting the hyperparameters (xval/yval) and finally those for testing the quality of the algoritm with the optimal hyperparameters (xtest/ytest). For doing that we use the function `Utils.partition` specifying the share we want to use for these three different jobs. As our data represent indeed a time serie, we want our model to be able to predict _future_ demand of bike sharing from _past_, observed rented bikes, so we do not shuffle the datasets as it would be the default.
+We can now split the dataset between the data we will use for training the algorithm (xtrain/ytrain), those for selecting the hyperparameters (xval/yval) and finally those for testing the quality of the algoritm with the optimal hyperparameters (xtest/ytest). For doing that we use the `partition` function specifying the share we want to use for these three different jobs. As our data represents indeed a time serie, we want our model to be able to predict _future_ demand of bike sharing from _past_, observed rented bikes, so we do not shuffle the datasets as it would be the default.
 
 ```@example betaml_tutorial_regression_sharingBikes
[…]

[…]
