### 2021-03-29 8ed7183
Documentation and reverted min version to Julia 1.3.1
