+++ Project.toml
@@ -30,5 +30,5 @@ PDMats = "0.9, 0.10, 0.11"
 ProgressMeter = "1.3"
 Reexport = "0.2, 1.0"
 Zygote = "0.4, 0.5, 0.6"
-julia = "1.5.0"
+julia = "1.3.1"
 StableRNGs ="1.0"
