+++ docs/src/tutorials/A classification task when labels are known - determining the plant species giving floreal measures/betaml_tutorial_classification_iris.jl
@@ -20,8 +20,6 @@ iris = dataset("datasets", "iris")
 describe(y)
 
 
-
-
 # ## Decision Trees and Random Forests
 x = convert(Matrix,iris[:,1:4])
 y = convert(Array{String,1},iris[:,5])
