### 2021-03-29 ad4badf
Doc: Updated flux example to latest API; optimised some NN code
