### 2021-03-29 2e11e93
CI doc: removed the working classigfication to focus only on regression doc
