+++ dev/tutorials/A classification task when labels are known - determining the plant species giving floreal measures/tutorial_classification_iris.jl
@@ -1,8 +1,0 @@
-
-# ## This is markdown title
-# This is also markdown
-## This is a normal comment
-#
-a = 1
-b = a + 1
-println(b)
