### 2021-03-29 cf34197
DOC: Improved visualisation of the regression example
