+++ docs/src/tutorials/A regression task - sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.jl
@@ -36,7 +36,7 @@ plot(data.cnt, title="Daily bike sharing rents (2Y)", label=nothing)
 # Here we start using  Decision Tree and Random Forest models that belong to the first group, so the only things we have to do is to select the variables in input (the "feature matrix", we wil lindicate it with "X") and those representing our output (the values we want to learn to predict, we call them "y"):
 
 x    = convert(Matrix,hcat(data[:,[:instant,:season,:yr,:mnth,:holiday,:weekday,:workingday,:weathersit,:temp,:atemp,:hum,:windspeed]]))
-y    = data[:,16]
+y    = data[:,16];
 
 # We can now split the dataset between the data we will use for training the algorithm (`xtrain`/`ytrain`), those for selecting the hyperparameters (`xval`/`yval`) and finally those for testing the quality of the algoritm with the optimal hyperparameters (`xtest`/`ytest`). We use the `partition` function specifying the share we want to use for these three different subsets, here 75%, 12.5% and 12.5 respectively. As our data represents indeed a time serie, we want our model to be able to predict _future_ demand of bike sharing from _past_, observed rented bikes, so we do not shuffle the datasets as it would be the default.
 
@@ -126,9 +126,9 @@ myTree = buildTree(xtrain,ytrain, maxDepth=bestMaxDepth, maxFeatures=bestMaxFeat
 (ŷtrain,ŷval,ŷtest) = predict.([myTree], [xtrain,xval,xtest])
 
 # Note that the above code uses the "dot syntax" to "broadcast" `predict()` over an array of label matrices. It is exactly equivalent to:
-ŷtrain = predict(myTree, xtrain)
-ŷval   = predict(myTree, xval)
-ŷtest  = predict(myTree, xtest)
+ŷtrain = predict(myTree, xtrain);
+ŷval   = predict(myTree, xval);
+ŷtest  = predict(myTree, xtest);
 
 # We now compute the relative mean error for the training, the validation and the test set. The `meanRelError` is a very flexible error function. Without additional parameter, it computes, as the name says, the _mean relative error_, also known as the "mean absolute percentage error" (MAPE)](https://en.wikipedia.org/wiki/Mean_absolute_percentage_error)") between an estimated and a true vector.
 # However it can also compute the _relative mean error_ (as we do here), or use a p-norm higher than 1.
@@ -136,9 +136,9 @@ ŷtest  = predict(myTree, xtest)
 # In this exercise we use the later, as our data has clearly some outlier days with very small rents, and we care more of avoiding our customers finding empty bike racks than having unrented bikes on the rack. Targeting a low mean average error would push all our predicitons down to try accomodate the low-level predicitons (to avoid a large relative error), and that's not what we want.
 
 # For example let'c consider the following example:
-y     = [30,28,27,3,32,38]
-ŷpref = [32,30,28,10,31,40]
-ŷbad  = [29,25,24,5,28,35]
+y     = [30,28,27,3,32,38];
+ŷpref = [32,30,28,10,31,40;]
+ŷbad  = [29,25,24,5,28,35];
 
 # Here ŷpref is an ipotetical output of a model that minimise the relative mean error, while ŷbad minimise the mean realative error
[…]
