### 2021-03-29 1f37828
Typo in doc tutorial regression
