### 2021-03-30 3dcb709
Trying to optimise NN getGradient
