+++ docs/src/tutorials/A regression task - sharing bike demand prediction/betaml_tutorial_regression_sharingBikes.md
@@ -47,7 +47,8 @@ Here we start using  Decision Tree and Random Forest models that belong to the f
 
 ```@example betaml_tutorial_regression_sharingBikes
 x    = convert(Matrix,hcat(data[:,[:instant,:season,:yr,:mnth,:holiday,:weekday,:workingday,:weathersit,:temp,:atemp,:hum,:windspeed]]))
-y    = data[:,16]
+y    = data[:,16];
+nothing #hide
 ```
 
 We can now split the dataset between the data we will use for training the algorithm (`xtrain`/`ytrain`), those for selecting the hyperparameters (`xval`/`yval`) and finally those for testing the quality of the algoritm with the optimal hyperparameters (`xtest`/`ytest`). We use the `partition` function specifying the share we want to use for these three different subsets, here 75%, 12.5% and 12.5 respectively. As our data represents indeed a time serie, we want our model to be able to predict _future_ demand of bike sharing from _past_, observed rented bikes, so we do not shuffle the datasets as it would be the default.
@@ -124,7 +125,7 @@ function tuneHyperParameters(model,xtrain,ytrain,xval,yval;maxDepthRange=15:15,m
 end
 ```
 
-We can now run the hyperparameter optimisation function with some "reasonable" ranges. To obtain repetable results we call `tuneHyperParameters` with `rng=copy(FIXEDRNG)`, where `FIXEDRNG` is a fixed-seeded random number generator guaranteed to maintain the same stream of random numbers even between different julia versions. That's also what we use for our unit tests.
+We can now run the hyperparameter optimisation function with some "reasonable" ranges. To obtain replicable results we call `tuneHyperParameters` with `rng=copy(FIXEDRNG)`, where `FIXEDRNG` is a fixed-seeded random number generator guaranteed to maintain the same stream of random numbers even between different julia versions. That's also what we use for our unit tests.
 
 ```@example betaml_tutorial_regression_sharingBikes
 (bestRme,bestMaxDepth,bestMaxFeatures,bestMinRecords) = tuneHyperParameters("DecisionTree",xtrain,ytrain,xval,yval,
@@ -156,9 +157,10 @@ The above `buildTree`function produces a DecisionTree object that can be used to
 Note that the above code uses the "dot syntax" to "broadcast" `predict()` over an array of label matrices. It is exactly equivalent to:
 
 ```@example betaml_tutorial_regression_sharingBikes
-ŷtrain = predict(myTree, xtrain)
-ŷval   = predict(myTree, xval)
-ŷtest  = predict(myTree, xtest)
+ŷtrain = predict(myTree, xtrain);
+ŷval   = predict(myTree, xval);
+ŷtest  = predict(myTree, xtest);
+nothing #hide
 ```
 
 We now compute the relative mean error for the training, the validation and the test set. The `meanRelError` is a very flexible error function. Without additional parameter, it computes, as the name says, the _mean relative error_, also known as the "mean absolute percentage error" (MAPE)](https://en.wikipedia.org/wiki/Mean_absolute_percentage_error)") between an estimated and a true vector.
@@ -169,9 +171,10 @@ In this exercise we use the later, as our data has clearly some outlier days wit
 For example let'c consider the following example:
 
[…]
