### 2021-03-30 b7fdb7b
Removed multi-threading to NN (lots of @simd and @inbounds instead) and completed regression tutorial
