### 2021-03-31 84a572b
Implemented the possibility that the output of VectorFunctionLayer has different size as the input
This allows the layer to be used to "pool" neurons from the previous layer, with the newly implemented
`pool1d` function
