+++ test/Nn_tests.jl
@@ -1,12 +1,12 @@
 using Test
-using DelimitedFiles, LinearAlgebra, MLDatasets
+using DelimitedFiles, LinearAlgebra #, MLDatasets
 
 using StableRNGs
 #rng = StableRNG(123)
 using BetaML
 
-#TESTRNG = FIXEDRNG # This could change...
-TESTRNG = StableRNG(123)
+TESTRNG = FIXEDRNG # This could change...
+#TESTRNG = StableRNG(123)
 
 println("*** Testing Neural Network...")
 
@@ -36,22 +36,29 @@ println("Testing basic NN behaviour...")
 x = [0.1,1]
 y = [1,0]
 l1   = DenseNoBiasLayer(2,2,w=[2 1;1 1],f=identity,rng=copy(TESTRNG))
-l2   = VectorFunctionLayer(2,2,f=softmax)
+l2   = VectorFunctionLayer(2,f=softmax)
 mynn = buildNetwork([l1,l2],squaredCost,name="Simple Multinomial logistic regression")
 o1 = forward(l1,x)
+@test o1 == [1.2,1.1]
 #@code_warntype forward(l1,x)
 o2 = forward(l2,o1)
+@test o2 ≈ [0.5249791874789399, 0.47502081252106] ≈ softmax([1.2,1.1])
 orig = predict(mynn,x')[1,:]
+@test orig == o2
 #@code_warntype Nn.predict(mynn,x')[1,:]
 ϵ = squaredCost(o2,y)
 #@code_warntype  squaredCost(o2,y)
 lossOrig = loss(mynn,x',y')
+@test ϵ == lossOrig
 #@code_warntype  loss(mynn,x',y')
 dϵ_do2 = dSquaredCost(o2,y)
+@test dϵ_do2 == [-0.4750208125210601,0.47502081252106]
 #@code_warntype dSquaredCost(o2,y)
 dϵ_do1 = backward(l2,o1,dϵ_do2)
+@test dϵ_do1 ≈ [-0.23691761847142412, 0.23691761847142412]
 #@code_warntype backward(l2,o1,dϵ_do2)
 dϵ_dX = backward(l1,x,dϵ_do1)
+@test dϵ_dX ≈ [-0.23691761847142412, 0.0]
 #@code_warntype backward(l1,x,dϵ_do1)
 l1w = getParams(l1)
 #@code_warntype
@@ -59,11 +66,13 @@ l2w = getParams(l2)
 #@code_warntype
 w = getParams(mynn)
 #@code_warntype getParams(mynn)
-#typeof(w)
+#typeof(w)2-element Vector{Float64}:
 origW = deepcopy(w)
 l2dw = getGradient(l2,o1,dϵ_do2)
+@test length(l2dw.data) == 0
 #@code_warntype getGradient(l2,o1,dϵ_do2)
 l1dw = getGradient(l1,x,dϵ_do1)
+@test l1dw.data[1] ≈ [-0.023691761847142414 -0.23691761847142412; 0.023691761847142414 0.23691761847142412]
 #@code_warntype
 dw = getGradient(mynn,x,y)
 #@code_warntype getGradient(mynn,x,y)
@@ -207,7 +216,7 @@ xtest = scale(xtest)
 
 l1   = DenseLayer(4,10, w=ones(10,4), wb=zeros(10),f=celu,rng=copy(TESTRNG))
 l2   = DenseLayer(10,3, w=ones(3,10), wb=zeros(3),rng=copy(TESTRNG))
-l3   = VectorFunctionLayer(3,3,f=softmax)
+l3   = VectorFunctionLayer(3,f=softmax)
[…]
