+++ dev/tutorials/A classification task when labels are known - determining the plant species giving floreal measures/betaml_tutorial_classification_iris.jl
@@ -133,3 +133,101 @@ testAccuracy   = accuracy(ŷtest,ytest)   # 0.956
 # ## Perceptron
 
 # ## Neural networks
+
+
+using HTTP, CSV, DataFrames, Random
+model = RDatasets.datasets("datasets")
+data = dataset("datasets", "mtcars")
+describe(data)
+
+# -------------------------------------------------------------
+baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","cars")
+data    = CSV.File(joinpath(baseDir,"data","auto-mpg.data-original_edited.csv"),delim=' ',missingstring="NA", ignorerepeated=true, header=false) |> DataFrame
+
+import Pipe: @pipe
+
+https://www.rpubs.com/dksmith01/cars
+https://archive.ics.uci.edu/ml/datasets/auto+mpg
+
+urlDataOriginal = "https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data-original"
+urlData         = "https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data"
+
+
+data = @pipe HTTP.get(urlDataOriginal).body |>
+          replace!(_, UInt8('\t') => UInt8(' ')) |>
+          CSV.File(_, delim=' ', missingstring="NA", ignorerepeated=true, header=false) |>
+          DataFrame
+data = @pipe HTTP.get(urlData).body |>
+        replace!(_, UInt8('\t') => UInt8(' ')) |>
+        CSV.File(_, delim=' ', missingstring="?", ignorerepeated=true, header=false) |>
+        DataFrame
+
+
+data[shuffle(copy(FIXEDRNG),axes(data, 1)), :]
+
+#urlData = "https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data-original"
+#data    = CSV.File(replace(IOBuffer(String(HTTP.get(urlData).body)),'\t' => ' '), delim=' ',missingstring="NA", ignorerepeated=true, header=false) |> DataFrame
+describe(data)
+
+x = Matrix{Union{Missing,Float64}}(data[:,1:7])
+x2 = predictMissing(x,3,mixtures=[FullGaussian() for i in 1:3],minVariance=0.002,rng=copy(FIXEDRNG)).X̂
+
+y = Vector{Int64}(data[:,8])
+
+#((xtrain,xval,xtest),(ytrain,yval,ytest)) = partition([x,y],[0.75,0.125,1-0.75-0.125],rng=copy(FIXEDRNG))
+((xtrain,xtest),(ytrain,ytest)) = partition([x2,y],[0.8,1-0.8],rng=copy(FIXEDRNG))
+myForest       = buildForest(xtrain,ytrain,5, minRecords=2,  rng=FIXEDRNG,forceClassification=true);
+ŷtrain         = predict(myForest, xtrain,rng=FIXEDRNG)
+#ŷval           = predict(myForest, xval,rng=FIXEDRNG)
+ŷtest          = predict(myForest, xtest,rng=FIXEDRNG)
+trainAccuracy  = accuracy(parse.(Int64,mode(ŷtrain)),ytrain)
+#valAccuracy    = accuracy(parse.(Int64,mode(ŷval)),yval)
+testAccuracy   = accuracy(parse.(Int64,mode(ŷtest)),ytest)
+
+
+#-
+
+y_oh            = oneHotEncoder(y) # Convert to One-hot representation (e.g. 2 => [0 1 0], 3 => [0 0 1])
[…]
