### 2021-04-01 5348ded
VectorFunctionLayer: moved computation of output size at construction time
