### 2021-04-02 9c89764
DOC: classification tutorial done
