### 2021-04-02 026fe70
Docs: completed classificaton tutorial
