+++ docs/src/tutorials/Regression - bike sharing/betaml_tutorial_regression_sharingBikes.jl
@@ -10,7 +10,7 @@
 # Note that even if we are estimating a time serie, we are not using here a recurrent neural network as we assume the temporal dependence to be negligible (i.e. $Y_t = f(X_t)$ alone).
 
 # ## Library and data loading
-#=
+
 # We first load all the packages we are going to use
 using  LinearAlgebra, Random, Statistics, DataFrames, CSV, Plots, Pipe, BenchmarkTools, BetaML
 import Distributions: Uniform
@@ -19,7 +19,7 @@ using  Test     #src
 
 
 # Here we load the data from a csv provided by the BataML package
-baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","A regression task - sharing bike demand prediction")
+baseDir = joinpath(dirname(pathof(BetaML)),"..","docs","src","tutorials","Regression - bike sharing")
 data    = CSV.File(joinpath(baseDir,"data","bike_sharing_day.csv"),delim=',') |> DataFrame
 describe(data)
 
@@ -201,7 +201,7 @@ oobError, trueTestMeanRelativeError  = myForest.oobError,meanRelError(ŷtest,yt
 #+
 (ŷtrain,ŷval,ŷtest)         = predict.([myForest], [xtrain,xval,xtest])
 (rmeTrain, rmeVal, rmeTest) = meanRelError.([ŷtrain,ŷval,ŷtest],[ytrain,yval,ytest],normRec=false)
-
+#src 0.06512088063750156, 0.09389025098838212, 0.22229683782214169
 @test rmeTest <= 0.23 #src
 
 # In this case we found an error very similar to the one employing a single decision tree. Let's print the observed data vs the estimated one using the random forest and then along the temporal axis:
@@ -256,6 +256,7 @@ model = DecisionTree.build_forest(ytrain, convert(Matrix,xtrain),
 # DecisionTrees.jl makes a good job in optimising the Random Forest algorithm, as it is over 3 times faster that BetaML.
 
 (rmeTrain, rmeVal, rmeTest) = meanRelError.([ŷtrain,ŷval,ŷtest],[ytrain,yval,ytest],normRec=false)
+#src 0.031235291992120055,0.17271800981298888,0.3141747075399198
 # However the error on the test set remains relativly high. The very low error level on the training set is a sign that it overspecialised on the training set, and we should have better ran a dedicated hyperparameter optimisation for the model.
 
 @test rmeTest <= 0.36 #src
@@ -414,6 +415,8 @@ ŷtest  = @pipe predict(mynn,xtestScaled)  |> scale(_ ,yScaleFactors,rev=true);
 
 #-
 (mreTrain, mreVal, mreTest) = meanRelError.([ŷtrain,ŷval,ŷtest],[ytrain,yval,ytest],normRec=false)
+#src 0.08839103036501561, 0.1482673840724729, 0.17613463255050987
+
 # The error is much lower. Let's plot our predictions:
 @test mreTest < 0.18 #src
 
@@ -472,6 +475,7 @@ ŷtestf  = @pipe Flux_nn(xtestScaled')'  |> scale(_,yScaleFactors,rev=true);
 
 # ..and we compute the mean relative errors..
 (mreTrain, mreVal, mreTest) = meanRelError.([ŷtrainf,ŷvalf,ŷtestf],[ytrain,yval,ytest],normRec=false)
+#src 0.09808947560569008, 0.13646196170811473,0.16181699665523128
 # .. finding an error not significantly different than the one obtained from BetaML.Nn.
 
 #-
@@ -503,4 +507,19 @@ plot(data[stc:endc,:dteday],[data[stc:endc,:cnt] ŷvalfullf[stc:endc] ŷtestfu
 
[…]
