+++ docs/make.jl
@@ -6,9 +6,9 @@ push!(LOAD_PATH,"../src/")
 const _TUTORIAL_DIR = joinpath(@__DIR__, "src", "tutorials")
 const _TUTORIAL_SUBDIR = [
     "Getting started",
-    "A regression task - sharing bike demand prediction",
-    "A classification task when labels are known - determining the country of origin of cars given the cars characteristics",
-    "A clusterisation task  - determining the plant species giving floreal measures"
+    "Regression - bike sharing",
+    "Classification - cars",
+    "Clusterisation - Iris"
 ]
 
 function link_example(content)
