+++ dev/tutorials/Classification - cars/betaml_tutorial_classification_cars.jl
@@ -1,0 +1,193 @@
+# # [A classification task when labels are known - determining the country of origin of cars given the cars characteristics](@id classification_tutorial)
+
+# In this exercise we have some car technical characteristics (mpg, horsepower,weight, model year...) and the country of origin and we want to create a model such that the country of origin can be accurately predicted given the technical characteristics.
+
+#
+# Data origin:
+# - dataset description: [https://archive.ics.uci.edu/ml/datasets/auto+mpg](https://archive.ics.uci.edu/ml/datasets/auto+mpg)
+#src Also useful: https://www.rpubs.com/dksmith01/cars
+# - data source we use here: [https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data](https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data)
+
+# field description
+
+# 1. mpg:           continuous
+# 2. cylinders:     multi-valued discrete
+# 3. displacement:  continuous
+# 4. horsepower:    continuous
+# 5. weight:        continuous
+# 6. acceleration:  continuous
+# 7. model year:    multi-valued discrete
+# 8. origin:        multi-valued discrete
+# 9. car name:      string (unique for each instance) - not used here
+
+
+
+# ## Library and data loading
+
+# We load a buch of packages that we'll use during this tutorial..
+using Random, HTTP, CSV, DataFrames, BenchmarkTools, BetaML
+import DecisionTree, Flux
+import Pipe: @pipe
+using  Test     #src
+
+# To load the data from the internet our workflow is
+# (1) Retrieve the data -> (2) Clean it -> (3) Load it -> (4) Output it as a DataFrame
+
+# For step 1 we use HTTP.get(), for step (2) we use `replace!`, for steps (3) and (4) we uses the CSV package, and we use the "pip" `|>` operator to chain these operations:
+
+urlDataOriginal = "https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data-original"
+data = @pipe HTTP.get(urlDataOriginal).body                                                |>
+             replace!(_, UInt8('\t') => UInt8(' '))                                        |>
+             CSV.File(_, delim=' ', missingstring="NA", ignorerepeated=true, header=false) |>
+             DataFrame;
+
+# This results in a table where the rows are the observations (the various cars) and the column the fields. All BetaML models expect this layout.
+# As the dataset is ordered, we randomly shuffle the data. Note that we pass to shuffle `copy(FIXEDRNG)` as the random nuber generator in order to obtain reproducible output.
+data[shuffle(copy(FIXEDRNG),axes(data, 1)), :]
+describe(data)
+
+# Columns 1 to 7 contain  characteristics of the car, while column 8 encodes the country or origin ("1" -> US, "2" -> EU, "3" -> Japan) that we want to be able to predict.
+# Columns 9 contains the car name, but we are not going to use this information in this tutorial.
+# Note also that some fields have missing data.
+# Our first step is hence to divide the dataset in features (the x) and the labels (the y) we want to predict. The `x` is then a Julia standard `Matrix` of 406 rows by 7 columns and the `y` is a vector of the 406 observations:
+x     = Matrix{Union{Missing,Float64}}(data[:,1:7]);
+y     = Vector{Int64}(data[:,8]);
+
[…]
