### 2021-04-05 6593cc9
Doc: reviewed class tutorial, doc compilation without code evaluation. Utils: started confusion matrix
Documentation:
- partial review of the classification tutorial
- implemented a way to build documlentation without having to execute/test the example code

Utils:
- Started the work on the confusion matrix to superseed accuracy
