### 2021-04-05 58696d0
Added mode methods and added randomness if multiple same makimum values, added ConfusionMatrix
TODO: A few things to add to ConfusionMatrix (incl nice printing), and appy it to the tutorial
