+++ src/Utils.jl
@@ -33,7 +33,7 @@ export Verbosity, NONE, LOW, STD, HIGH, FULL,
        bic, aic,
        autoJacobian,
        squaredCost, dSquaredCost, crossEntropy, dCrossEntropy, classCounts, meanDicts, mode, gini, entropy, variance,
-       error, accuracy, meanRelError,
+       error, accuracy, meanRelError, ConfusionMatrix,
        l1_distance,l2_distance, l2²_distance, cosine_distance, lse, sterling,
        #normalFixedSd, logNormalFixedSd,
        radialKernel, polynomialKernel
@@ -65,8 +65,6 @@ Use it with:
 """
 const FIXEDRNG  = StableRNG(FIXEDSEED) #StableRNG(FIXEDSEED) Random.default_rng() #MersenneTwister(FIXEDSEED)
 #const FIXEDRNG  = MersenneTwister(FIXEDSEED) #StableRNG(FIXEDSEED) Random.default_rng()
-
-
 
 
 macro codeLocation()
@@ -772,12 +770,37 @@ struct ConfusionMatrix{T}
     labels::Vector{T}             # Array of categorical labels
     actualCount::Vector{Int64}    # Array of counts per lebel in the actual data
     predictedCount::Vector{Int64} # Array of counts per label in the predicted data
-    scores::Array{Int64,2}        # Matrix predicted (rows) vs actual (columns)
+    scores::Array{Int64,2}        # Matrix actual (rows) vs predicted (columns)
     normalisedScores::Array{Float64,2} # Normalised scores
+    tp::Vector{Int64}                  # True positive (by class)
+    tn::Vector{Int64}                  # True negative (by class)
+    fp::Vector{Int64}                  # False positive (by class)
+    fn::Vector{Int64}                  # False negative (by class)
+    accuracy::Float64                  # Overall accuracy rate
+    error::Float64                     # Overall error rate
 end
+# https://towardsdatascience.com/confusion-matrix-for-your-multi-class-machine-learning-model-ff9aa3bf7826
+# https://en.wikipedia.org/wiki/Confusion_matrix
 
-
-
+function ConfusionMatrix(ŷ,y::AbstractArray{T};rng=Random.GLOBAL_RNG,labels=unique(y)) where {T}
+    nCl = length(labels)
+    ŷ = typeof(ŷ) <: AbstractVector{T} ? ŷ : mode(ŷ,rng=rng) # get the mode if needed
+    N = length(y)
+    length(ŷ) == N || @error "ŷ and y must have the same length in ConfusionMatrix"
+    actualCount = [get(classCounts(y),i,0) for i in labels]
+    predictedCount = [get(classCounts(ŷ),i,0) for i in labels]
+    scores = zeros(Int64,(nCl,nCl))
+    normalisedScores = zeros(Float64,(nCl,nCl))
+    [scores[findfirst(x -> x == y[i],labels),findfirst(x -> x == ŷ[i],labels)] += 1 for i in 1:N]
+    [normalisedScores[r,:] = scores[r,:] ./ actualCount[r] for r in 1:nCl]
+    tp = [scores[i,i] for i in 1:nCl]
+    tn = [sum(scores[r,c] for r in 1:nCl, c in 1:nCl if r != i && c != i)  for i in 1:nCl]
+    fp = [sum(scores[r,c] for r in 1:nCl, c in 1:nCl if r != i && c == i)  for i in 1:nCl]
+    fn = [sum(scores[r,c] for r in 1:nCl, c in 1:nCl if r == i && c != i)  for i in 1:nCl]
+    accuracy = sum(tp)/N
+    error    = 1-accuracy
[…]
