+++ dev/tutorials/Classification - cars/betaml_tutorial_classification_cars.jl
@@ -1,24 +1,25 @@
 # # [A classification task when labels are known - determining the country of origin of cars given the cars characteristics](@id classification_tutorial)
 
-# In this exercise we have some car technical characteristics (mpg, horsepower,weight, model year...) and the country of origin and we want to create a model such that the country of origin can be accurately predicted given the technical characteristics.
+# In this exercise we have some car technical characteristics (mpg, horsepower,weight, model year...) and the country of origin and we would like to create a model such that the country of origin can be accurately predicted given the technical characteristics.
+# As the information to predict is a multi-class one, this is a _[classification]_(https://en.wikipedia.org/wiki/Statistical_classification) task.
 
 #
 # Data origin:
 # - dataset description: [https://archive.ics.uci.edu/ml/datasets/auto+mpg](https://archive.ics.uci.edu/ml/datasets/auto+mpg)
 #src Also useful: https://www.rpubs.com/dksmith01/cars
-# - data source we use here: [https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data](https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data)
+# - data source we use here: [https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data](https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data-original)
 
-# field description
+# Field description:
 
-# 1. mpg:           continuous
-# 2. cylinders:     multi-valued discrete
-# 3. displacement:  continuous
-# 4. horsepower:    continuous
-# 5. weight:        continuous
-# 6. acceleration:  continuous
-# 7. model year:    multi-valued discrete
-# 8. origin:        multi-valued discrete
-# 9. car name:      string (unique for each instance) - not used here
+# 1. mpg:           _continuous_
+# 2. cylinders:     _multi-valued discrete_
+# 3. displacement:  _continuous_
+# 4. horsepower:    _continuous_
+# 5. weight:        _continuous_
+# 6. acceleration:  _continuous_
+# 7. model year:    _multi-valued discrete_
+# 8. origin:        _multi-valued discrete_
+# 9. car name:      _string (unique for each instance)_ - not used here
 
 
 
@@ -31,9 +32,9 @@ import Pipe: @pipe
 using  Test     #src
 
 # To load the data from the internet our workflow is
-# (1) Retrieve the data -> (2) Clean it -> (3) Load it -> (4) Output it as a DataFrame
+# (1) Retrieve the data --> (2) Clean it --> (3) Load it --> (4) Output it as a DataFrame.
 
-# For step 1 we use HTTP.get(), for step (2) we use `replace!`, for steps (3) and (4) we uses the CSV package, and we use the "pip" `|>` operator to chain these operations:
+# For step (1) we use `HTTP.get()``, for step (2) we use `replace!`, for steps (3) and (4) we uses the `CSV` package, and we use the "pip" `|>` operator to chain these operations:
 
 urlDataOriginal = "https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data-original"
 data = @pipe HTTP.get(urlDataOriginal).body                                                |>
@@ -42,31 +43,33 @@ data = @pipe HTTP.get(urlDataOriginal).body                                     
              DataFrame;
 
 # This results in a table where the rows are the observations (the various cars) and the column the fields. All BetaML models expect this layout.
-# As the dataset is ordered, we randomly shuffle the data. Note that we pass to shuffle `copy(FIXEDRNG)` as the random nuber generator in order to obtain reproducible output.
[…]
