### 2021-04-06 e420dbb
Completed ConfusionMatrix with constructor and `print`/`println` definitions
