+++ src/Utils.jl
@@ -19,7 +19,7 @@ Provide shared utility functions for various machine learning algorithms. You do
 """
 module Utils
 
-using LinearAlgebra, Random, Statistics, Combinatorics, Zygote, CategoricalArrays, StableRNGs
+using LinearAlgebra, Printf, Random, Statistics, Combinatorics, Zygote, CategoricalArrays, StableRNGs
 
 using ForceImport
 @force using ..Api
@@ -764,45 +764,164 @@ error(ŷ::Array{T,2},y::Array{Int64,1};tol=1) where {T <: Number} = 1 - accurac
 """ error(ŷ,y) - Categorical error with with probabilistic predictions of a dataset given in terms of a dictionary of probabilities (Dict{T,Float64} vs T). """
 error(ŷ::Array{Dict{T,Float64},1},y::Array{T,1};tol=1) where {T} = 1 - accuracy(ŷ,y;tol=tol)
 
+"""
+    ConfusionMatrix
 
 
+Scores and measures resulting from a comparation between true and predicted categorical variables
+
+Use the function `ConfusionMatrix(ŷ,y;classes,labels,rng)` to build it and `print(cm::ConfusionMatrix;what)` to visualise it, or use the individual parts of interest, e.g. `display(cm.scores)`.
+
+# Fields:
+- `labels`: Array of categorical labels
+- `accuracy`: Overall accuracy rate
+- `misclassification`: Overall misclassification rate
+- `actualCount`: Array of counts per lebel in the actual data
+- `predictedCount`: Array of counts per label in the predicted data
+- `scores`: Matrix actual (rows) vs predicted (columns)
+- `normalisedScores`: Normalised scores
+- `tp`: True positive (by class)
+- `tn`: True negative (by class)
+- `fp`: False positive (by class), aka "type I error" or "false allarm"
+- `fn`: False negative (by class), aka "type II error" or "miss"
+- `precision`: True class i over predicted class i (by class)
+- `recall`: Predicted class i over true class i (by class), aka "True Positive Rate (TPR)", "Sensitivity" or "Probability of detection"
+- `specificity`: Predicted not class i over true not class i (by class), aka "True Negative Rate (TNR)"
+- `f1Score`: Harmonic mean of precision and recall
+- `meanPrecision`: Mean by class, respectively unweighted and weighted by actualCount
+- `meanRecall`: Mean by class, respectively unweighted and weighted by actualCount
+- `meanSpecificity`: Mean by class, respectively unweighted and weighted by actualCount
+- `meanF1Score`: Mean by class, respectively unweighted and weighted by actualCount
+
+
+"""
 struct ConfusionMatrix{T}
-    labels::Vector{T}             # Array of categorical labels
-    actualCount::Vector{Int64}    # Array of counts per lebel in the actual data
-    predictedCount::Vector{Int64} # Array of counts per label in the predicted data
-    scores::Array{Int64,2}        # Matrix actual (rows) vs predicted (columns)
-    normalisedScores::Array{Float64,2} # Normalised scores
-    tp::Vector{Int64}                  # True positive (by class)
-    tn::Vector{Int64}                  # True negative (by class)
-    fp::Vector{Int64}                  # False positive (by class)
-    fn::Vector{Int64}                  # False negative (by class)
-    accuracy::Float64                  # Overall accuracy rate
-    error::Float64                     # Overall error rate
+    classes::Vector{T}                      # Array of categorical labels
+    labels::Vector{String}                  # String representation of the categories
[…]
