+++ Project.toml
@@ -12,6 +12,7 @@ ForceImport = "9dda63f9-cce7-5873-89fa-eccbb2fffcde"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MLJModelInterface = "e80e1ace-859a-464e-9ed9-23947d8ae3ea"
 PDMats = "90014a1f-27ba-587c-ab20-58faa44d9150"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
@@ -29,6 +30,6 @@ MLJModelInterface = "0.3.7, 0.4.0"
 PDMats = "0.9, 0.10, 0.11"
 ProgressMeter = "1.3"
 Reexport = "0.2, 1.0"
+StableRNGs = "1.0"
 Zygote = "0.4, 0.5, 0.6"
 julia = "1.4"
-StableRNGs ="1.0"
