### 2021-04-08 f2bc902
Added new pkg LoopVectorization, various utility function tweeks/bugfix, bugfix in MLJ
- Added pkg LoopVectorization and use of its @avx macro in NN internals
- Solved a bug in MLJ interface `fit(m::GMMClusterer,...)`
- Solved bug in accuracy when PMF was multimodal (including all values the same). This affected `mode`
- Renamed `classcounts` to `classCountsWithLabels` and added `classCounts` for when we don't care to the labels. Rewrite of `gini` and `entropy`
in accordance of it
- Solved a problem with `print(cm::ConfusionMatrix)` and `println`
- Doc tutorial classification: solve a bug in N
[…]
