+++ docs/src/tutorials/Classification - cars/betaml_tutorial_classification_cars.jl
@@ -2,6 +2,7 @@
 
 # In this exercise we have some car technical characteristics (mpg, horsepower,weight, model year...) and the country of origin and we would like to create a model such that the country of origin can be accurately predicted given the technical characteristics.
 # As the information to predict is a multi-class one, this is a _[classification]_(https://en.wikipedia.org/wiki/Statistical_classification) task.
+# It is a challenging exercise due to the simultaneous presence of three factors: (1) presence of missing data; (2) unbalanced data - 254 out of 406 cars are US made; (3) small dataset.
 
 #
 # Data origin:
@@ -76,13 +77,24 @@ myForest       = buildForest(xtrain,ytrain,30, rng=copy(FIXEDRNG),forceClassific
 #src [`predict`](@ref BetaML.Trees.predict)  [`predict`](@ref forest_prediction)
 # with our `myForest` model and either the training or testing data.
 ŷtrain,ŷtest   = predict.(Ref(myForest), [xtrain,xtest],rng=copy(FIXEDRNG));
-# Finally we can measure the _accuracy_ of our predictions with the [`accuracy`](@ref) function:
-trainAccuracy,testAccuracy  = accuracy.([parse.(Int64,mode(ŷtrain)),parse.(Int64,mode(ŷtest))],[ytrain,ytest])
+# Finally we can measure the _accuracy_ of our predictions with the [`accuracy`](@ref) function, with the sidenote that we need first to "parse" the ŷs as forcing the classification job transformed automatically them in strings from the original integers:
+trainAccuracy,testAccuracy  = accuracy.([parse.(Int64,mode(ŷtrain,rng=copy(FIXEDRNG))),parse.(Int64,mode(ŷtest,rng=copy(FIXEDRNG)))],[ytrain,ytest])
 #src (0.9969230769230769,0.8024691358024691)
 
 @test testAccuracy > 0.8 #src
 
+
 # The predictions are quite good, for the training set the algoritm predicted almost all cars' origins correctly, while for the testing set (i.e. those records that has **not** been used to train the algorithm), the correct prediction level is still quite high, at 80%
+
+# While accuracy can sometimes suffice, we may often want to better understand which categories our model has trouble to predict correctly.
+# We can investigate the output of a multi-class classifier more in-deep with a [`ConfusionMatrix`](@ref) where the true values (`y`) are given in rows and the predicted ones (`ŷ`) in columns, together to some per-class metrics like the _precision_ (true class _i_ over predicted in class _i_), the _recall_ (predicted class _i_ over the true class _i_) and others.
+# We fist build the [`ConfusionMatrix`](@ref BetaML.Utils.ConfusionMatrix) object between `ŷ` and `y` and then we print it (we do it here for the test subset):
+
+cm = ConfusionMatrix(parse.(Int64,mode(ŷtest,rng=copy(FIXEDRNG))),ytest,classes=[1,2,3],labels=["US","EU","Japan"])
+BetaML.Utils.report(cm;what="all")
+
+# From the report we can see that are the Japanese cars that have more trouble in being correctly classified, and in particular many Japanease cars are classified as US ones.
+
 # When we benchmark the resourse used (time and memory) we find that Random Forests remain pretty fast, expecially when we compare them with neural networks (see later)
 @btime buildForest(xtrain,ytrain,30, rng=copy(FIXEDRNG),forceClassification=true);
[…]

[…]
