### 2021-04-08 deb16a4
Restricted test of pool1D function in NN to Julia >= 1.6
See https://github.com/FluxML/Zygote.jl/issues/943
