### 2021-04-08 d00327c
MLJ Interface: Removed unused GMMFitter instances and correcte scitype of GMMClusterer
