+++ src/Clustering_MLJ.jl
@@ -3,7 +3,7 @@
 import MLJModelInterface       # It seems that having done this in the top module is not enought
 const MMI = MLJModelInterface  # We need to repoeat it here
 
-export KMeans, KMedoids, GMMFitter, GMMClusterer, MissingImputator
+export KMeans, KMedoids, GMMClusterer, MissingImputator
 
 # ------------------------------------------------------------------------------
 # Model Structure declarations..
@@ -37,28 +37,6 @@ KMeans(;
     Z₀           = nothing,
     rng          = Random.GLOBAL_RNG,
   ) = KMedoids(K,dist,initStrategy,Z₀,rng)
-
-# function gmm(X,K;p₀=nothing,mixtures=[DiagonalGaussian() for i in 1:K],tol=10^(-6),verbosity=STD,minVariance=0.05,minCovariance=0.0,initStrategy="grid")
-mutable struct GMMFitter <: MMI.Probabilistic
-  K::Int64
-  p₀::Union{Nothing,AbstractArray{Float64,1}}
-  mixtures::Symbol
-  tol::Float64
-  minVariance::Float64
-  minCovariance::Float64
-  initStrategy::String
-  rng::AbstractRNG
-end
-GMMFitter(;
-    K             = 3,
-    p₀            = nothing,
-    mixtures      = :diag_gaussian,
-    tol           = 10^(-6),
-    minVariance   = 0.05,
-    minCovariance = 0.0,
-    initStrategy  = "kmeans",
-    rng           = Random.GLOBAL_RNG,
-) = GMMFitter(K,p₀,mixtures, tol, minVariance, minCovariance,initStrategy,rng)
 
 mutable struct GMMClusterer <: MMI.Unsupervised
   K::Int64
@@ -119,26 +97,6 @@ end
 MMI.fitted_params(model::Union{KMeans,KMedoids}, fitresult) = (centers=fitesult[2], cluster_labels=CategoricalArrays.categorical(fitresults[1]))
 
 
-function MMI.fit(m::GMMFitter, verbosity, X, y)
-    # X is nothing, y is the data: https://alan-turing-institute.github.io/MLJ.jl/dev/adding_models_for_general_use/#Models-that-learn-a-probability-distribution-1
-    y          = MMI.matrix(y) # convert table to matrix
-
-    if m.mixtures == :diag_gaussian
-        mixtures = [DiagonalGaussian() for i in 1:m.K]
-    elseif m.mixtures == :full_gaussian
-        mixtures = [FullGaussian() for i in 1:m.K]
-    elseif m.mixtures == :spherical_gaussian
-        mixtures = [SphericalGaussian() for i in 1:m.K]
-    else
-        error("Usupported mixture. Supported mixtures are either `:diag_gaussian`, `:full_gaussian` or `:spherical_gaussian`.")
-    end
-    res        = gmm(y,m.K,p₀=deepcopy(m.p₀),mixtures=mixtures, minVariance=m.minVariance, minCovariance=m.minCovariance,initStrategy=m.initStrategy,verbosity=NONE,rng=m.rng)
-    fitResults = (pₙₖ=res.pₙₖ,pₖ=res.pₖ,mixtures=res.mixtures) # res.pₙₖ
-    cache      = nothing
-    report     = (res.ϵ,res.lL,res.BIC,res.AIC)
-    return (fitResults, cache, report)
-end
-
 function MMI.fit(m::GMMClusterer, verbosity, X)
     # X is nothing, y is the data: https://alan-turing-institute.github.io/MLJ.jl/dev/adding_models_for_general_use/#Models-that-learn-a-probability-distribution-1
     x          = MMI.matrix(X) # convert table to matrix
[…]
