### 2021-04-09 1550ac6
Removed transform(model::GMMClusterer)
