### 2021-04-09 d298e7b
Adapted tests
- removed `transform(GMMClusterer,...)`  test
- as now mode is stochastic, added the rng to all the accuracy functions in random trees
