+++ test/Clustering_tests.jl
@@ -102,6 +102,7 @@ out3 = predictMissing(X,3,mixtures=[FullGaussian() for i in 1:3],verbosity=NONE,
 # NEW TEST
 println("Testing MLJ interface for Clustering models....")
 X, y                           = Mlj.@load_iris
+
 model                          = KMeans(rng=copy(TESTRNG))
 modelMachine                   = Mlj.machine(model, X)
 (fitResults, cache, report)    = Mlj.fit(model, 0, X)
@@ -121,11 +122,9 @@ acc = BetaML.accuracy(Mlj.levelcode.(yhat),Mlj.levelcode.(y),ignoreLabels=true)
 model                       =  GMMClusterer(mixtures=:diag_gaussian,rng=copy(TESTRNG))
 modelMachine                =  Mlj.machine(model, X) # DimensionMismatch
 (fitResults, cache, report) =  Mlj.fit(model, 0, X)
-yhat_prob                   =  Mlj.transform(model,fitResults,nothing) # Mlj.transform(model,fitResults,X)
-yhat_prob2                  =  Mlj.predict(model, fitResults, X)  # Mlj.transform(model,fitResults,X)
+yhat_prob                   =  Mlj.predict(model, fitResults, X)  # Mlj.transform(model,fitResults,X)
 # how to get this ??? Mlj.predict_mode(yhat_prob)
-@test Distributions.pdf(yhat_prob[end],2) ≈ 0.5937332769039277
-@test Distributions.pdf(yhat_prob2[end],2) ≈ 0.5937443601647852
+@test Distributions.pdf(yhat_prob[end],2) ≈ 0.5937443601647852
 
 X = [1 10.5;1.5 missing; 1.8 8; 1.7 15; 3.2 40; missing missing; 3.3 38; missing -2.3; 5.2 -2.4]
 X = Mlj.table(X)
