+++ Project.toml
@@ -34,4 +34,4 @@ ProgressMeter = "1.3"
 Reexport = "0.2, 1.0"
 StableRNGs = "1.0"
 Zygote = "0.4, 0.5, 0.6"
-julia = "1.4"
+julia = "1.3"
