### 2021-04-09 284af4d
Removed LoopVectorization as just its "using" create problems in Julia 1.5
For some reasons only "using LoopVectorization" creates problem in unrelated function as
MLJ::predict(GMMClustering).
As the advantage was minimal and for now we want to keep compatibility with Julia1.5, it
has been removed.
