+++ Project.toml
@@ -10,7 +10,6 @@ DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ForceImport = "9dda63f9-cce7-5873-89fa-eccbb2fffcde"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-LoopVectorization = "bdcacae8-1622-11e9-2a5c-532679323890"
 MLJModelInterface = "e80e1ace-859a-464e-9ed9-23947d8ae3ea"
 PDMats = "90014a1f-27ba-587c-ab20-58faa44d9150"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
@@ -27,7 +26,6 @@ CategoricalArrays = "0.8, 0.9"
 Combinatorics = "1"
 Distributions = "0.23, 0.24"
 ForceImport = "0.0.3"
-LoopVectorization = "0.8, 0.9, 0.10, 0.11, 0.12"
 MLJModelInterface = "0.3.7, 0.4.0"
 PDMats = "0.9, 0.10, 0.11"
 ProgressMeter = "1.3"
