+++ docs/src/tutorials/Classification - cars/betaml_tutorial_classification_cars.jl
@@ -91,15 +91,13 @@ trainAccuracy,testAccuracy  = accuracy.([parse.(Int64,mode(ŷtrain,rng=copy(FIX
 # We fist build the [`ConfusionMatrix`](@ref BetaML.Utils.ConfusionMatrix) object between `ŷ` and `y` and then we print it (we do it here for the test subset):
 
 cm = ConfusionMatrix(parse.(Int64,mode(ŷtest,rng=copy(FIXEDRNG))),ytest,classes=[1,2,3],labels=["US","EU","Japan"])
-BetaML.Utils.report(cm;what="all")
+print(cm;what="all")
 
-# From the report we can see that are the Japanese cars that have more trouble in being correctly classified, and in particular many Japanease cars are classified as US ones.
+# From the report we can see that Japanese cars have more trouble in being correctly classified, and in particular many Japanease cars are classified as US ones. This is likely a result of the class imbalance of the data set, and could be solved by balancing the dataset with various sampling tecniques before training the model.
 
 # When we benchmark the resourse used (time and memory) we find that Random Forests remain pretty fast, expecially when we compare them with neural networks (see later)
 @btime buildForest(xtrain,ytrain,30, rng=copy(FIXEDRNG),forceClassification=true);
-#src   128.335 ms (781027 allocations: 196.30 MiB)
-
-
+#src   134.096 ms (781027 allocations: 196.30 MiB)
 
 # ### Comparision with DecisionTree.jl
 
@@ -119,9 +117,7 @@ model = DecisionTree.build_forest(ytrain, xtrainFull,-1,30,rng=123)
 # While the accuracy on the training set is exactly the same as for `BetaML` random forets, `DecisionTree.jl` random forests are slighly less accurate in the testing sample.
 # Where however `DecisionTrees.jl` excell is in the efficiency: they are extremelly fast and memory parse, even if here to this benchmark we should add the resources need to impute the missing values. Also, one of the reasons DecisionTrees are such efficient is that internally they sort the data to avoid repeated comparision, but in this way they work only with features that are sortable, while BetaML random forests accept virtually any kind of input without the need of adapt it.
 @btime  DecisionTree.build_forest(ytrain, xtrainFull,-1,30,rng=123);
-#src 1.451 ms (10875 allocations: 1.52 MiB)
-
-
+#src 1.431 ms (10875 allocations: 1.52 MiB)
 
 # ### Neural network
 
@@ -132,8 +128,8 @@ D               = size(xtrainFull,2)
 classes         = unique(y)
 nCl             = length(classes)
 
-# The second "inconvenient" of NN i that, while not requiring feature engineering, they stil lneed a bit of practice on the way to build the network. It's not as simple as `train(model,x,y)`. We need here to specify how we want our layers, _chain_ the layers together and then decide a _loss_ overall function. Only when we done these steps, we have the model ready for training.
-# Here we define 3 [`DenseLayer`](@ref) zwhere, for each of them, we specify the number of neurons in input (the first layer being equal to the dimensions of the data), the output layer (for a classification task, the last layer output size beying equal to the number of classes) and an _activation function for each layer (default the `identity` function).
[…]
