### 2021-04-13 #2849 @sylvaticus
Yes, as you can see [in the commit log](https://github.com/sylvaticus/BetaML.jl/commits/master), I am actually still implementing the modifications required by the reviewers.. I created an interface to my models for one of the toolbox cited (these interfface has been already being pushed but still needs to be included in a release of BetaML) and I am implementing a more detailed set of tutorials.
I would still need a 1-2 weeks to complete it and update the JOSS paper.
