### 2021-04-13 92191fc
Added utils function `shuffle` of an array of n-dimensional arrays on a given dimension
