+++ docs/src/tutorials/Clusterisation - Iris/betaml_tutorial_cluster_iris.jl
@@ -1,38 +1,64 @@
 # # A classification task: the prediction of  plant species from floreal measures (the iris tdataset)
-# The task is to estimate the species of a plant given some floreal measurements. It is a very
+# The task is to estimate the species of a plant given some floreal measurements. It use the classical "Iris" dataset.
+# Note that in this example we are using clustering approaches, so we try to understand the "structure" of our data, without relying to actually knowing the true labels ("classes" or "factors"). However we have chosen a dataset for which the true labels are actually known, so to compare the accuracy of the algorithms we use, but these labels will not be used during the algorithms training.
+
 #
 # Data origin:
 # - dataset description: [https://en.wikipedia.org/wiki/Iris_flower_data_set](https://en.wikipedia.org/wiki/Iris_flower_data_set)
 # - data source we use here: [https://github.com/JuliaStats/RDatasets.jl](https://github.com/JuliaStats/RDatasets.jl)
 
-# Note that even if we are estimating a time serie, we are not using here a recurrent neural network as we assume the temporal dependence to be negligible (i.e. $Y_t = f(X_t)$ alone).
 
 # ## Library and data loading
 
-using LinearAlgebra, Random, Statistics, DataFrames, CSV, Plots, Pipe, BetaML, BenchmarkTools, RDatasets
+using  LinearAlgebra, Random, Statistics, DataFrames, CSV, Plots, Pipe, BetaML, BenchmarkTools, RDatasets
 import Distributions: Uniform
-import DecisionTree, Flux ## For comparisions
+
 using  Test     #src
 
 
-# Differently from the [regression tutorial](../A%20regression%20task%20-%20sharing%20bike%20demand%20prediction/betaml_tutorial_regression_sharingBikes.html), we load the data here from `RDatasets`, a package providing standard datasets.
+
+
+
+
+# Differently from the [regression tutorial](@ref regression_tutorial), we load the data here from `RDatasets`, a package providing standard datasets.
 iris = dataset("datasets", "iris")
+
+
 describe(iris)
 
-#=
-# ## Decision Trees and Random Forests
 
 # ### Data preparation
-# The first step is to prepare the data for the analysis. This indeed depends already on the model we want to employ, as some models "accept" everything as input, no matter if the data is numerical or categorical, if it has missing values or not... while other models are instead much more exigents, and require more work to "clean up" our dataset.
-# Here we start using  Decision Tree and Random Forest models that belong to the first group, so the only things we have to do is to select the variables in input (the "feature matrix", we wil lindicate it with "X") and those representing our output (the values we want to learn to predict, we call them "y"):
+# The first step is to prepare the data for the analysis.
 
-x = Matrix{Float64}(iris[:,1:4])
-y = Vector{String}(iris[:,5])
 
[…]
