### 2021-04-14 fef2e06
CompatHelper: bump compat for "MLJModelInterface" to "1.0"
