+++ Project.toml
@@ -26,7 +26,7 @@ CategoricalArrays = "0.8, 0.9"
 Combinatorics = "1"
 Distributions = "0.23, 0.24"
 ForceImport = "0.0.3"
-MLJModelInterface = "0.3.7, 0.4.0"
+MLJModelInterface = "0.3.7, 0.4.0, 1.0"
 PDMats = "0.9, 0.10, 0.11"
 ProgressMeter = "1.3"
 Reexport = "0.2, 1.0"
