### 2021-04-14 9b10e19
Added `dims` to partition
Now we can partition over to an arbitrary dimension
