+++ src/Utils.jl
@@ -33,7 +33,7 @@ export Verbosity, NONE, LOW, STD, HIGH, FULL,
        bic, aic,
        autoJacobian,
        squaredCost, dSquaredCost, crossEntropy, dCrossEntropy, classCounts, classCountsWithLabels, meanDicts, mode, gini, entropy, variance,
-       error, accuracy, meanRelError, ConfusionMatrix,
+       error, accuracy, meanRelError, ConfusionMatrix, crossValidate,
        l1_distance,l2_distance, l2²_distance, cosine_distance, lse, sterling,
        #normalFixedSd, logNormalFixedSd,
        radialKernel, polynomialKernel
@@ -313,8 +313,9 @@ function batch(n::Integer,bSize::Integer;sequential=false,rng = Random.GLOBAL_RN
     return batches
 end
 
+# TODO: add dims to partition (see shuffle)
 """
-    partition(data,parts;shuffle=true)
+    partition(data,parts;shuffle,dims,rng)
 
 Partition (by rows) one or more matrices according to the shares in `parts`.
 
@@ -322,7 +323,12 @@ Partition (by rows) one or more matrices according to the shares in `parts`.
 * `data`: A matrix/vector or a vector of matrices/vectors
 * `parts`: A vector of the required shares (must sum to 1)
 * `shufle`: Whether to randomly shuffle the matrices (preserving the relative order between matrices)
+* `dims`: The dimension for which to partition [def: `1`]
 * `rng`: Random Number Generator (see [`FIXEDSEED`](@ref)) [deafult: `Random.GLOBAL_RNG`]
+
+# Notes:
+* The sum of parts must be equal to 1
+* The number of elements in the specified dimension must be the same for all the arrays in `data`
 
 # Example:
 ```julia
@@ -331,40 +337,35 @@ julia> y = collect(31:40)
 julia> ((xtrain,xtest),(ytrain,ytest)) = partition([x,y],[0.7,0.3])
  ```
  """
-function partition(data::AbstractArray{T,1},parts::AbstractArray{Float64,1};shuffle=true,rng = Random.GLOBAL_RNG) where T <: AbstractArray
+function partition(data::AbstractArray{T,1},parts::AbstractArray{Float64,1};shuffle=true,dims=1,rng = Random.GLOBAL_RNG) where T <: AbstractArray
         # the sets of vector/matrices
-        n = size(data[1],1)
-        if !all(size.(data,1) .== n)
-            @error "All matrices passed to `partition` must have the same number of rows"
-        end
-        ridx = shuffle ? Random.shuffle(rng,1:n) : collect(1:n)
-        return partition.(data,Ref(parts);shuffle=shuffle, fixedRIdx = ridx,rng=rng)
+        N = size(data[1],dims)
+        all(size.(data,dims) .== N) || @error "All matrices passed to `partition` must have the same number of elements for the required dimension"
+        ridx = shuffle ? Random.shuffle(rng,1:N) : collect(1:N)
+        return partition.(data,Ref(parts);shuffle=shuffle,dims=dims,fixedRIdx = ridx,rng=rng)
 end
 
-function partition(data::AbstractArray{T,N}, parts::AbstractArray{Float64,1};shuffle=true,fixedRIdx=Int64[],rng = Random.GLOBAL_RNG) where {T,N}
-    # the individual vecto/matrix
-    data = makeMatrix(data)
-    n = size(data,1)
-    nParts = size(parts)
-    toReturn = []
[…]
