### 2021-04-14 65b2b38
Added a `copy` parameter to `partition` (def to true)
