+++ src/Utils.jl
@@ -33,7 +33,7 @@ export Verbosity, NONE, LOW, STD, HIGH, FULL,
        bic, aic,
        autoJacobian,
        squaredCost, dSquaredCost, crossEntropy, dCrossEntropy, classCounts, classCountsWithLabels, meanDicts, mode, gini, entropy, variance,
-       error, accuracy, meanRelError, ConfusionMatrix, crossValidate,
+       error, accuracy, meanRelError, ConfusionMatrix, crossValidation,
        l1_distance,l2_distance, l2²_distance, cosine_distance, lse, sterling,
        #normalFixedSd, logNormalFixedSd,
        radialKernel, polynomialKernel
@@ -175,9 +175,6 @@ function oneHotEncoderRow(x::Union{AbstractArray{T},T};factors=singleUnique(x),d
     return oneHotEncoderRow(integerEncoder(x;factors=factors),d=length(factors);count=count)
 end
 
-
-
-
 """
     oneHotEncoder(x;d,factors,count)
 
@@ -313,7 +310,7 @@ function batch(n::Integer,bSize::Integer;sequential=false,rng = Random.GLOBAL_RN
     return batches
 end
 
-# TODO: add dims to partition (see shuffle)
+
 """
     partition(data,parts;shuffle,dims,rng)
 
@@ -324,6 +321,7 @@ Partition (by rows) one or more matrices according to the shares in `parts`.
 * `parts`: A vector of the required shares (must sum to 1)
 * `shufle`: Whether to randomly shuffle the matrices (preserving the relative order between matrices)
 * `dims`: The dimension for which to partition [def: `1`]
+* `copy`: Wheter to _copy_ the actual data or only create a reference [def: `true`]
 * `rng`: Random Number Generator (see [`FIXEDSEED`](@ref)) [deafult: `Random.GLOBAL_RNG`]
 
 # Notes:
@@ -337,19 +335,19 @@ julia> y = collect(31:40)
 julia> ((xtrain,xtest),(ytrain,ytest)) = partition([x,y],[0.7,0.3])
  ```
  """
-function partition(data::AbstractArray{T,1},parts::AbstractArray{Float64,1};shuffle=true,dims=1,rng = Random.GLOBAL_RNG) where T <: AbstractArray
+function partition(data::AbstractArray{T,1},parts::AbstractArray{Float64,1};shuffle=true,dims=1,copy=true,rng = Random.GLOBAL_RNG) where T <: AbstractArray
         # the sets of vector/matrices
         N = size(data[1],dims)
         all(size.(data,dims) .== N) || @error "All matrices passed to `partition` must have the same number of elements for the required dimension"
         ridx = shuffle ? Random.shuffle(rng,1:N) : collect(1:N)
-        return partition.(data,Ref(parts);shuffle=shuffle,dims=dims,fixedRIdx = ridx,rng=rng)
+        return partition.(data,Ref(parts);shuffle=shuffle,dims=dims,fixedRIdx = ridx,copy=copy,rng=rng)
 end
 
-function partition(data::AbstractArray{T,Ndims}, parts::AbstractArray{Float64,1};shuffle=true,dims=1,fixedRIdx=Int64[],rng = Random.GLOBAL_RNG) where {T <: Number,Ndims}
+function partition(data::AbstractArray{T,Ndims}, parts::AbstractArray{Float64,1};shuffle=true,dims=1,fixedRIdx=Int64[],copy=true,rng = Random.GLOBAL_RNG) where {T <: Number,Ndims}
     # the individual vector/matrix
     N        = size(data,dims)
     nParts   = size(parts)
-    toReturn = Array{Array{T,Ndims},1}(undef,nParts)
[…]
