### 2021-04-15 fb13775
Added crossValidation (to clean up)
- partition now accepts non-numerics
- created AbstractDataSampler, SamplerWithData, and KFold types
- Implemented the iteration API for KFold (`iterate(iter::SamplerWithData{KFold})` and `iterate(iter::SamplerWithData{KFold},state)` )
- Implemented crossValidation
- Moved samplers to their own Samplers.jl file

TODO: doc crossValidation, reorganisation of the code in different folders and (for utils) files
