### 2021-04-15 a11b06f
Completed doc of crossValidation and cleaned up a bit
