### 2021-04-15 b426414
Reorganised src files in folders (Utils, Nn, ...) and removed Clustering.py and the empty Svn.jl
