### 2021-04-16 932030e
Completed clustering tutorial and set default kmedoids initialisation to grid
