+++ src/Utils/Utils.jl
@@ -1321,50 +1321,4 @@ lse(x) = maximum(x)+log(sum(exp.(x .- maximum(x))))
 sterling(n::BigInt,k::BigInt) = (1/factorial(k)) * sum((-1)^i * binomial(k,i)* (k-i)^n for i in 0:k)
 sterling(n::Int64,k::Int64)   = sterling(BigInt(n),BigInt(k))
 
-#=
-# https://github.com/simonster/Reexport.jl/blob/master/src/Reexport.jl
-macro reexport(ex)
-    isa(ex, Expr) && (ex.head == :module ||
-                      ex.head == :using ||
-                      (ex.head == :toplevel &&
-                       all(e->isa(e, Expr) && e.head == :using, ex.args))) ||
-        error("@reexport: syntax error")
-
-    if ex.head == :module
-        modules = Any[ex.args[2]]
-        ex = Expr(:toplevel, ex, :(using .$(ex.args[2])))
-    elseif ex.head == :using && all(e->isa(e, Symbol), ex.args)
-        modules = Any[ex.args[end]]
-    elseif ex.head == :using && ex.args[1].head == :(:)
-        symbols = [e.args[end] for e in ex.args[1].args[2:end]]
-        return esc(Expr(:toplevel, ex, :(eval(Expr(:export, $symbols...)))))
-    else
-        modules = Any[e.args[end] for e in ex.args]
-    end
-
-    esc(Expr(:toplevel, ex,
-             [:(eval(Expr(:export, names($mod)...))) for mod in modules]...))
-end
-=#
-
-#=
-
-"""
-    gradientDescentSingleUpdate(θ,▽,η)
-
-Perform a single update of parameter θ using the gradient descent method with gradient ▽ and learning rate η.
-
-# Notes:
-- The parameter and the gradient can be either numbers, arrays or tuple of arrays.
-"""
-#- For Arrays and tuple of floats it is available also as inplace modification as
-#  `gradientDescentSingleUpdate!(θ,▽,η)`
-gradientDescentSingleUpdate(θ::Number,▽::Number,η) = θ .- (η .* ▽)
-gradientDescentSingleUpdate(θ::AbstractArray,▽::AbstractArray,η) = gradientDescentSingleUpdate.(θ,▽,Ref(η))
-gradientDescentSingleUpdate(θ::Tuple,▽::Tuple,η) = gradientDescentSingleUpdate.(θ,▽,Ref(η))
-#gradientDescentSingleUpdate!(θ::AbstractArray{AbstractFloat},▽::AbstractArray{AbstractFloat},η) = (θ .= θ .- (η .* ▽))
-#gradientDescentSingleUpdate!(θ::AbstractArray{AbstractFloat},▽::AbstractArray{Number},η) = (θ .= θ .- (η .* ▽))
-#gradientDescentSingleUpdate!(θ::Tuple,▽::Tuple,η) = gradientDescentSingleUpdate!.(θ,▽,Ref(η))
-=#
-
 end # end module
