### 2021-04-18 d4576cb
Solved the problem of documentation breacking docmunetation build
Taht was due to @btime of GaussianMixtures.jl that was crashing
