+++ docs/src/tutorials/Clusterisation - Iris/betaml_tutorial_cluster_iris.jl
@@ -21,7 +21,7 @@ using  Test     #src
 Random.seed!(123)
 #logger  = Logging.SimpleLogger(stdout, Logging.Error); global_logger(logger); ## For suppressing GaussianMixtures output
 
-# Differently from the [regression tutorial](@ref regression_tutorial), we load the data here from `RDatasets`, a package providing standard datasets.
+# Differently from the [regression tutorial](@ref regression_tutorial), we load the data here from [`RDatasets`](https://github.com/JuliaStats/RDatasets.jl](https://github.com/JuliaStats/RDatasets.jl), a package providing standard datasets.
 iris = dataset("datasets", "iris")
 describe(iris)
 
@@ -127,7 +127,7 @@ report = DataFrame(mName = modelLabels, avgAccuracy = dropdims(round.(μs',digit
 
 # Up to now we used the real labels to compare the model accuracies. But in real clustering examples we don't have the true classes, or we wouln't need to do clustering in the first instance, so we don't know the number of classes to use.
 # There are several methods to judge clusters algorithms goodness, perhaps the simplest one, at least for the expectation-maximisation algorithm employed in `gmm` to fit the data to the unknown mixture, is to use a information criteria that trade the goodness of the lickelyhood with the parameters used to do the fit.
-# BetaML provide by default in the gmm clustering outputs both the _Bayesian information criterion_  ([`BIC`](@ref)) and the _Akaike information criterion_  ([`AIC`](@ref)), where for both a lower value is better.
+# BetaML provide by default in the gmm clustering outputs both the _Bayesian information criterion_  ([`BIC`](@ref bic)) and the _Akaike information criterion_  ([`AIC`](@ref aic)), where for both a lower value is better.
 
 # We can then run the model with different number of classes and see which one leads to the lower BIC or AIC.
 # We run hence `crossValidation` again with the `FullGaussian` gmm model
@@ -171,34 +171,29 @@ plot(1:K,[μsBICS' μsAICS'], labels=["BIC" "AIC"], title="Information criteria 
 
 # We now benchmark the time and memory required by the various models by using the `@btime` macro of the `BenchmarkTools` package:
 
-@btime kmeans($xs,3);
-#-
-@btime kmedoids($xs,3);
-#-
-@btime gmm($xs,3,mixtures=[SphericalGaussian() for i in 1:3], verbosity=NONE);
-#-
-@btime gmm($xs,3,mixtures=[DiagonalGaussian() for i in 1:3], verbosity=NONE);
-#-
-@btime gmm($xs,3,mixtures=[FullGaussian() for i in 1:3], verbosity=NONE);
-#-
-@btime Clustering.kmeans($xs', 3);
-#-
-@btime begin dGMM = GaussianMixtures.GMM(3, $xs; method=:kmeans, kind=:diag); GaussianMixtures.em!(dGMM, $xs) end;
-#-
-@btime begin fGMM = GaussianMixtures.GMM(3, $xs; method=:kmeans, kind=:full); GaussianMixtures.em!(fGMM, $xs) end;
+# ```
+# @btime kmeans($xs,3);
+# # 261.540 μs (3777 allocations: 442.53 KiB)
+# @btime kmedoids($xs,3);
+# 4.576 ms (97356 allocations: 10.42 MiB)
+# @btime gmm($xs,3,mixtures=[SphericalGaussian() for i in 1:3], verbosity=NONE);
+# # 5.498 ms (133365 allocations: 8.42 MiB)
+# @btime gmm($xs,3,mixtures=[DiagonalGaussian() for i in 1:3], verbosity=NONE);
[…]
