+++ src/Utils/Processing.jl
@@ -1,0 +1,598 @@
+# Part of submodule Utils of BetaML - The Beta Machine Learning Toolkit
+# Functions typically used for processing (manipulating) data, typically preprocessing data before running a ML model
+
+
+# ------------------------------------------------------------------------------
+# Various reshaping functions
+import Base.reshape
+""" reshape(myNumber, dims..) - Reshape a number as a n dimensional Array """
+reshape(x::T, dims...) where {T <: Number} =   (x = [x]; reshape(x,dims) )
+makeColVector(x::T) where {T} =  [x]
+makeColVector(x::T) where {T <: AbstractArray} =  reshape(x,length(x))
+makeRowVector(x::T) where {T <: Number} = return [x]'
+makeRowVector(x::T) where {T <: AbstractArray} =  reshape(x,1,length(x))
+"""Transform an Array{T,1} in an Array{T,2} and leave unchanged Array{T,2}."""
+makeMatrix(x::AbstractArray) = ndims(x) == 1 ? reshape(x, (size(x)...,1)) : x
+
+
+"""Return wheather an array is sortable, i.e. has methos issort defined"""
+issortable(::AbstractArray{T,N})  where {T,N} = hasmethod(isless, Tuple{nonmissingtype(T),nonmissingtype(T)})
+
+
+"""
+    getPermutations(v::AbstractArray{T,1};keepStructure=false)
+
+Return a vector of either (a) all possible permutations (uncollected) or (b) just those based on the unique values of the vector
+
+Useful to measure accuracy where you don't care about the actual name of the labels, like in unsupervised classifications (e.g. clustering)
+
+"""
+function getPermutations(v::AbstractArray{T,1};keepStructure=false) where {T}
+    if !keepStructure
+        return Combinatorics.permutations(v)
+    else
+        classes       = unique(v)
+        nCl           = length(classes)
+        N             = size(v,1)
+        pSet          = Combinatorics.permutations(1:nCl)
+        nP            = length(pSet)
+        vPermutations = fill(similar(v),nP)
+        vOrigIdx      = [findfirst(x -> x == v[i] , classes) for i in 1:N]
+        for (pIdx,perm) in enumerate(pSet)
+            vPermutations[pIdx] = classes[perm[vOrigIdx]] # permuted specific version
+        end
+        return vPermutations
+    end
+end
+
+
+""" singleUnique(x) Return the unique values of x whether x is an array of arrays, an array or a scalar"""
+function singleUnique(x::Union{T,AbstractArray{T}}) where {T <: Union{Any,AbstractArray{T2}} where T2 <: Any }
+    if typeof(x) <: AbstractArray{T2} where {T2 <: AbstractArray}
+        return unique(vcat(unique.(x)...))
+    elseif typeof(x) <: AbstractArray{T2} where {T2}
+        return unique(x)
+    else
+        return [x]
+    end
+end
+
+function oneHotEncoderRow(x::Union{AbstractArray{T},T}; d=maximum(x), factors=1:d,count = false) where {T <: Integer}
+    x = makeColVector(x)
+    out = zeros(Int64,d)
+    for j in x
+        out[j] = count ? out[j] + 1 : 1
+    end
+    return out
+end
+
+function oneHotEncoderRow(x::Union{AbstractArray{T},T};factors=singleUnique(x),d=length(factors),count = false) where {T}
+    x = makeColVector(x)
+    return oneHotEncoderRow(integerEncoder(x;factors=factors),d=length(factors);count=count)
+end
+
+"""
+    oneHotEncoder(x;d,factors,count)
+
+Encode arrays (or arrays of arrays) of categorical data as matrices of one column per factor.
+
[…]
