+++ src/Utils/Miscellaneous.jl
@@ -1,0 +1,2 @@
+# Part of submodule Utils of BetaML _ the Beta Machine Learning Toolkit
+# Miscelaneaous funcitons / types

[…]
