+++ dev/tutorials/Clusterisation - Iris/betaml_tutorial_cluster_iris.jl
@@ -1,49 +1,199 @@
-# # A classification task: the prediction of  plant species from floreal measures (the iris tdataset)
-# The task is to estimate the species of a plant given some floreal measurements. It is a very
+# # [A classification task: the prediction of  plant species from floreal measures (the iris dataset)](@id classification_clustering)
+# The task is to estimate the species of a plant given some floreal measurements. It use the classical "Iris" dataset.
+# Note that in this example we are using clustering approaches, so we try to understand the "structure" of our data, without relying to actually knowing the true labels ("classes" or "factors"). However we have chosen a dataset for which the true labels are actually known, so to compare the accuracy of the algorithms we use, but these labels will not be used during the algorithms training.
+
 #
 # Data origin:
 # - dataset description: [https://en.wikipedia.org/wiki/Iris_flower_data_set](https://en.wikipedia.org/wiki/Iris_flower_data_set)
 # - data source we use here: [https://github.com/JuliaStats/RDatasets.jl](https://github.com/JuliaStats/RDatasets.jl)
 
-# Note that even if we are estimating a time serie, we are not using here a recurrent neural network as we assume the temporal dependence to be negligible (i.e. $Y_t = f(X_t)$ alone).
 
 # ## Library and data loading
+# We load the Beta Machine Learning Toolkit as well as some other packages that we use in this tutorial
+using BetaML
+using Random, Statistics, Logging, BenchmarkTools, RDatasets, Plots, DataFrames
 
-using LinearAlgebra, Random, Statistics, DataFrames, CSV, Plots, Pipe, BetaML, BenchmarkTools, RDatasets
-import Distributions: Uniform
-import DecisionTree, Flux ## For comparisions
+# We are also going to compare our results with two other leading packages in Julia for clustering analysis, [`Clustering.jl`](https://github.com/JuliaStats/Clustering.jl) that provides (inter alia) kmeans and kmedoids algorithms and [`GaussianMixtures.jl`](https://github.com/davidavdav/GaussianMixtures.jl) that provides, as the name says, Gaussian Mixture Models. So we import them (we "import" them, rather than "use", not to bound their full names into namespace as some would collide with BetaML).
+import Clustering, GaussianMixtures
 using  Test     #src
 
+# We do a few tweeks for the Clustering and GaussianMixtures packages. Note that in BetaML we can also control both the random seed and the verbosity in the algorithm call, not only globally
+Random.seed!(123)
+#logger  = Logging.SimpleLogger(stdout, Logging.Error); global_logger(logger); ## For suppressing GaussianMixtures output
 
-# Differently from the [regression tutorial](../A%20regression%20task%20-%20sharing%20bike%20demand%20prediction/betaml_tutorial_regression_sharingBikes.html), we load the data here from `RDatasets`, a package providing standard datasets.
+# Differently from the [regression tutorial](@ref regression_tutorial), we load the data here from [`RDatasets`](https://github.com/JuliaStats/RDatasets.jl](https://github.com/JuliaStats/RDatasets.jl), a package providing standard datasets.
 iris = dataset("datasets", "iris")
 describe(iris)
 
-#=
-# ## Decision Trees and Random Forests
+# The iris dataset  provides floreal measures in columns 1 to 4 and the assigned species name in column 5. There are no missing values
 
-# ### Data preparation
[…]
