### 2021-04-18 130616f
Reorganised documentation. Extended doc of Utilities module
