+++ src/Utils/Utils.jl
@@ -1,3 +1,5 @@
+
+
 """
   Utils.jl File
 
@@ -5,7 +7,7 @@ Machine Learning shared utility functions (Module BetaML.Utils)
 
 `?BetaML.Utils` for documentation
 
-- [Importable source code (most up-to-date version)](https://github.com/sylvaticus/BetaML.jl/blob/master/src/Utils.jl) - [Julia Package](https://github.com/sylvaticus/Utils.jl)
+- Part of [BetaML](https://github.com/sylvaticus/BetaML.jl)
 - New to Julia? [A concise Julia tutorial](https://github.com/sylvaticus/juliatutorial) - [Julia Quick Syntax Reference book](https://julia-book.com)
 
 """
@@ -14,7 +16,34 @@ Machine Learning shared utility functions (Module BetaML.Utils)
 """
     Utils module
 
-Provide shared utility functions for various machine learning algorithms. You don't usually need to import from this module, as each other module (Nn, Perceptron, Clusters,...) reexport it.
+Provide shared utility functions for various machine learning algorithms.
+
+For the complete list of functions provided see below. The main ones are:
+
+## Helper functions for logging
+- Most BetAML functions accept a parameter `verbosity` that expect one of the element in the `Verbosity` enoum (`NONE`, `LOW`, `STD`, `HIGH` or `FULL`)
+- Writing complex code and need to find where something is executed ? Use the macro [`@codeLocation`](@ref)
+
+## Stochasticity management
+- Utils provide [`FIXEDSEED`], [`FIXEDRNG`] and [`generateParallelRngs`](@ref). All stochastic functions accept a `rng` paraemter. See the "Getting started" section in the tutorial for details.
+
+## Data processing
+- Various small and large utilities for helping processing the data, expecially before running a ML algorithm
+- Includes [`getPermutaitons`](@ref), [`oneHotEncoder`](@ref), [`integerEncoder`](@ref) (and [`integerDecoder)`](@ref)), [`partition`](@ref), [`scale`](@ref) (and [`getScaleFactors`](@ref)), [`pca`](@ref), [`crossValidation`](@ref)
+
+## Samplers
+- Utilities to sample from data (e.g. for neural network training or for cross-validation)
+- Include the "generic" type [`SamplerWithData`](@ref), together with the sampler implementation [`KFold`](@ref) and the function [`batch`](@ref)
+
+## Transformers
+- Funtions that "transform" a single input (that can be also a vector or a matrix)
+- Includes varios NN "activation" functions ([`relu`](@ref), [`celu`](@ref), [`sigmoid`](@ref), [`softmax`](@ref), [`pool1d`](@ref)) and their derivatives (`d[FunctionName]`), but also [`gini`](@ref), [`entropy`](@ref), [`variance`](@ref), [`BIC`](@ref bic), [`AIC`](@ref aic)
+
+## Measures
+- Several functions of a pair of parameters (often `y` and `ŷ`) to measure the goodness of `ŷ`, the distance between the two elements of the pair, ...
+- Includes "classical" distance functions ([`l1_distance`](@ref), [`l2_distance`](@ref), [`l2²_distance`](@ref) [`cosine_distance`](@ref)), "cost" functions for continuous variables ([`squaredCost`](@ref), [`meanRelError`](@ref)) and comparision functions for multui-class variables ([`crossEntropy`](@ref), [`accuracy`](@ref), [`ConfusionMatrix`](@ref)).
+
+
 
 """
 module Utils
