### 2021-04-19 faf2dd5
Documentation checked. This is the first attempt for release v0.5.
