+++ docs/src/tutorials/Classification - cars/betaml_tutorial_classification_cars.jl
@@ -35,7 +35,7 @@ using  Test     #src
 # To load the data from the internet our workflow is
 # (1) Retrieve the data --> (2) Clean it --> (3) Load it --> (4) Output it as a DataFrame.
 
-# For step (1) we use `HTTP.get()``, for step (2) we use `replace!`, for steps (3) and (4) we uses the `CSV` package, and we use the "pip" `|>` operator to chain these operations:
+# For step (1) we use `HTTP.get()`, for step (2) we use `replace!`, for steps (3) and (4) we uses the `CSV` package, and we use the "pip" `|>` operator to chain these operations:
 
 urlDataOriginal = "https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data-original"
 data = @pipe HTTP.get(urlDataOriginal).body                                                |>
@@ -44,40 +44,49 @@ data = @pipe HTTP.get(urlDataOriginal).body                                     
              DataFrame;
 
 # This results in a table where the rows are the observations (the various cars) and the column the fields. All BetaML models expect this layout.
-# As the dataset is ordered, we randomly shuffle the data. Note that we pass to shuffle `copy(FIXEDRNG)` as the random nuber generator in order to obtain reproducible output ( [`FIXEDRNG`](@ref BetaML.Utils.FIXEDRNG) is nothing else than an istance of `StableRNG(123)` defined in the [`BetaML.Utils`](@ref) sub-module, but you can choose of course your own "fixed" RNG). See the [Dealing with stochasticity](@ref dealing_with_stochasticity) section in the [Getting started](@ref) tutorial for details.
+
+# As the dataset is ordered, we randomly shuffle the data. Note that we pass to shuffle `copy(FIXEDRNG)` as the random nuber generator in order to obtain reproducible output ( [`FIXEDRNG`](@ref BetaML.Utils.FIXEDRNG) is nothing else than an istance of `StableRNG(123)` defined in the [`BetaML.Utils`](@ref utils_module) sub-module, but you can choose of course your own "fixed" RNG). See the [Dealing with stochasticity](@ref dealing_with_stochasticity) section in the [Getting started](@ref getting_started) tutorial for details.
 data[shuffle(copy(FIXEDRNG),axes(data, 1)), :]
 describe(data)
 
-# Columns 1 to 7 contain  characteristics of the car, while column 8 encodes the country or origin ("1" -> US, "2" -> EU, "3" -> Japan). That's what we want to be able to predict.
+# Columns 1 to 7 contain  characteristics of the car, while column 8 encodes the country or origin ("1" -> US, "2" -> EU, "3" -> Japan). That's the variable we want to be able to predict.
+
 # Columns 9 contains the car name, but we are not going to use this information in this tutorial.
 # Note also that some fields have missing data.
+
 # Our first step is hence to divide the dataset in features (the x) and the labels (the y) we want to predict. The `x` is then a Julia standard `Matrix` of 406 rows by 7 columns and the `y` is a vector of the 406 observations:
 x     = Matrix{Union{Missing,Float64}}(data[:,1:7]);
 y     = Vector{Int64}(data[:,8]);
 
-# Some algorithms that we will use today don't like missing data, so we need to _impute_ them. Foir this we are using the [`predictMissing`](@ref) function provided by the [`BetaML.Clustering`](@ref) sub-module. Internally the function uses a Gaussian Mixture Model to assign to the missing walue of a given record an average of the values of the non-missing records weighted for how close they are to our specific record.
[…]
