### 2021-04-19 26e3a0b
Fixed documentation minor points, updated JOSS paper
