+++ docs/paper/paper.bib
@@ -139,3 +139,51 @@
   year = {2020},
   url = {https://github.com/IBM/AutoMLPipeline.jl},
 }
+
+% JuliaCall package
+@Article{Li:2019,
+    author = {Changcheng Li},
+    title = {{JuliaCall}: an {R} package for seamless integration between {R} and {Julia}},
+    journal = {The Journal of Open Source Software},
+    publisher = {The Open Journal},
+    year = {2019},
+    volume = {4},
+    number = {35},
+    pages = {1284},
+    doi = {10.21105/joss.01284},
+  }
+
+% PyJulia
+@software{Arakaki:2020,
+  author       = {Takafumi Arakaki and
+                  Jake Bolewski and
+                  Robin Deits and
+                  Keno Fischer and
+                  Steven G. Johnson and
+                  Matthias Bussonnier and
+                  Isaiah Norton and
+                  Páll Haraldsson and
+                  Matthew Rocklin and
+                  Tsur and
+                  Viral B. Shah and
+                  Daniel Soto and
+                  eslgastal and
+                  Elias Kuthe and
+                  jakirkham and
+                  Marius Millea and
+                  grahamgill and
+                  fnmdx111 and
+                  Alex Arslan and
+                  Darren Christopher Lukas and
+                  David Nadlinger and
+                  Lilian Besson and
+                  Sheehan Olver and
+                  Taine Zhao and
+                  scls19fr},
+  title        = {JuliaPy/pyjulia: PyJulia},
+  month        = nov,
+  year         = 2020,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.4294939},
+
+}
