+++ docs/src/tutorials/Betaml_tutorial_getting_started.md
@@ -2,30 +2,33 @@
 
 ## Introduction
 
-This page presents some general information concerning BetaML. For detailed information on the algorithms provided by the Toolkit refer to the individual modules API or to the tutorial sections below:
+This "tutorial" part of the documentation presents a step-by-step guide to the main algorithms and utility functions provided by BetaML and comparisons with the leading packages in each field.
+Aside this page, the tutorial is divided in the following sections:
 -  [Regression tutorial](@ref regression_tutorial) - Arguments: _Decision trees, Random forests, neural networks, hyper-parameter tuning, continuous error measures_
 -  [Classification tutorial](@ref classification_tutorial) - Arguments: _Decision trees and random forests, neural networks (softmax), pre-processing workflow, confusion matrix_
 -  [Clustering tutorial](@ref clustering_tutorial) - Arguments: _k-means, kmedoids, generative gaussain models, cross-validation_
 
-The following modules are currently implemented: [Perceptron](https://sylvaticus.github.io/BetaML.jl/dev/Perceptron.html) (linear and kernel-based classifiers), [Trees](https://sylvaticus.github.io/BetaML.jl/dev/Trees.html) (Decision Trees and Random Forests), [Nn](https://sylvaticus.github.io/BetaML.jl/dev/Nn.html) (Neural Networks), [Clustering](https://sylvaticus.github.io/BetaML.jl/dev/Clustering.html) (Kmean, Kmenoids, Expectation-Maximisation, Missing value imputation, ...) and [Utils](https://sylvaticus.github.io/BetaML.jl/dev/Utils.html).
+Detailed information on the algorithms can be instead found in the API (Reference manual) of the individual modules. The following modules are currently implemented: [Perceptron](@ref perceptron_module) (linear and kernel-based classifiers), [Trees](@ref trees_module) (Decision Trees and Random Forests), [Nn](@ref nn_module) (Neural Networks), [Clustering](@ref clustering_module) (Kmean, Kmenoids, Expectation-Maximisation, Missing value imputation, ...) and [Utils](@ref utils_module).
 
-The overall "philosophy" of BetaML is to be able to make simple things easy and complex things possible. On the most basic level, most algorithms have default parameters suitable for a basic analysis. A great level of flexibility can be already achieved by just employing the full set of model parameters, for example changing the distance function in `kmedoids` to `l1_distance` (aka "Manhattan distance").
-Finally, the greatest flexibility can be obtained by customising BetaML and writing for example your own neural network layer type (by subclassing `AbstractLayer`), your own sampler (by subclassing `AbstractDataSampler`) or your own mixture component (by subclassing `AbstractMixture`),
+Finally, theoretical notes describing most of these algorithms can be found at the companion repository [https://github.com/sylvaticus/MITx_6.86x](https://github.com/sylvaticus/MITx_6.86x).
+
+The overall "philosophy" of BetaML is to support simple machine learning tasks easily and make complex tasks possible. An the most basic level, the majority of  algorithms have default parameters suitable for a basic analysis. A great level of flexibility can be already achieved by just employing the full set of model parameters, for example changing the distance function in `kmedoids` to `l1_distance` (aka "Manhattan distance").
+Finally, the greatest flexibility can be obtained by customising BetaML and writing, for example, its own neural network layer type (by subclassing `AbstractLayer`), its own sampler (by subclassing `AbstractDataSampler`) or its own mixture component (by subclassing `AbstractMixture`),
 In such a case, while not required by any means, please consider to give it back to the community and open a pull request to integrate your types in BetaML.
 
-If you are looking for an introductory book on Julia, have a look on "[Julia Quick Syntax Reference](https://www.julia-book.com/)" (Apress,2019).
[…]

[…]
