### 2021-04-19 #2849 @sylvaticus
Dear editor and reviewers, I have updated the library and the paper to account for the reviewers' comments:

- a detailed [step-by-step tutorial](https://sylvaticus.github.io/BetaML.jl/dev/tutorials/Betaml_tutorial_getting_started.html) has been added to show the usage of the library (and, more in general, of ML techniques) and compare it with existing libraries. As I did expect, BetaML is surely less computationally performant, but it is not less accurate nor (with the notable exception of neural networks) less flexible;
- I added several interfaces to BetaML models to be used within the MLJ framework;
- In the meantime I updated/added several utility functions to the algorithms. For example, the set of `scale`, `pca`, `oneHotEncoder`,  `crossValidation` (with a configurable user-provided function / "do block") and other functions allows to easily make a "workflow" from the data to the ML models, even if not exactly a "pipeline".

I am confident that the modifications introduced will help the users of the library and I thanks the reviewers for the time they spent in suggesting the improvements to the library and their guidance in implementing them.
