+++ v0.5.0/tutorials/Clustering - Iris/betaml_tutorial_cluster_iris.jl
@@ -1,0 +1,207 @@
+# # [A classification task: the prediction of  plant species from floreal measures (the iris dataset)](@id clustering_tutorial)
+# The task is to estimate the species of a plant given some floreal measurements. It use the classical "Iris" dataset.
+# Note that in this example we are using clustering approaches, so we try to understand the "structure" of our data, without relying to actually knowing the true labels ("classes" or "factors"). However we have chosen a dataset for which the true labels are actually known, so to compare the accuracy of the algorithms we use, but these labels will not be used during the algorithms training.
+
+#
+# Data origin:
+# - dataset description: [https://en.wikipedia.org/wiki/Iris_flower_data_set](https://en.wikipedia.org/wiki/Iris_flower_data_set)
+# - data source we use here: [https://github.com/JuliaStats/RDatasets.jl](https://github.com/JuliaStats/RDatasets.jl)
+
+
+# ## Library and data loading
+# We load the Beta Machine Learning Toolkit as well as some other packages that we use in this tutorial
+using BetaML
+using Random, Statistics, Logging, BenchmarkTools, RDatasets, Plots, DataFrames
+
+# We are also going to compare our results with two other leading packages in Julia for clustering analysis, [`Clustering.jl`](https://github.com/JuliaStats/Clustering.jl) that provides (inter alia) kmeans and kmedoids algorithms and [`GaussianMixtures.jl`](https://github.com/davidavdav/GaussianMixtures.jl) that provides, as the name says, Gaussian Mixture Models. So we import them (we "import" them, rather than "use", not to bound their full names into namespace as some would collide with BetaML).
+import Clustering, GaussianMixtures
+using  Test     #src
+
+# We do a few tweeks for the Clustering and GaussianMixtures packages. Note that in BetaML we can also control both the random seed and the verbosity in the algorithm call, not only globally
+Random.seed!(123)
+#logger  = Logging.SimpleLogger(stdout, Logging.Error); global_logger(logger); ## For suppressing GaussianMixtures output
+
+# Differently from the [regression tutorial](@ref regression_tutorial), we load the data here from [`RDatasets`](https://github.com/JuliaStats/RDatasets.jl](https://github.com/JuliaStats/RDatasets.jl), a package providing standard datasets.
+iris = dataset("datasets", "iris")
+describe(iris)
+
+# The iris dataset  provides floreal measures in columns 1 to 4 and the assigned species name in column 5. There are no missing values
+
+# ## Data preparation
+# The first step is to prepare the data for the analysis. We collect the first 4 columns as our _feature_ `x` matrix and the last one as our `y` label vector.
+# As we are using clustering algorithms, we are not actually using the labels to train the algorithms, we'll behave like we do not know them, we'll just let the algorithm "learn" fro mthe structure of the data itself. We'll however use it to judge the accuracy that they did reach.
+x       = Matrix{Float64}(iris[:,1:4]);
+yLabels = unique(iris[:,5]);
+# As the labels are expressed as strings, the first thing we do is encode them as integers for our analysis using the function [`integerEncoder`](@ref).
+y       = integerEncoder(iris[:,5],factors=yLabels);
+
+# The dataset from RDatasets is ordered by species, so we need to shuffle it to avoid biases.
+# Shuffling happens by default in crossValidation, but we are keeping here a copy of the shuffled version for later.
[…]
