+++ paper.md
@@ -39,7 +39,7 @@ bibliography: docs/paper/paper.bib
 
 # Summary
 
-A serie of _machine learning_ algorithms has been implemented and bundled in a single package for the Julia programming language.
+A serie of _machine learning_ algorithms has been implemented and bundled together with several "utility" functions in a single package for the Julia programming language.
 Currently, algorithms are available in the areas of classification (perceptron, kernel perceptron, pegasos), neural networks (feed-forward), clustering (kmeans, kmenoids, gmm, missing values imputation) and decision trees/random forests. Development of these algorithms started following the theoretical notes of the MOOC class "Machine Learning with Python: from Linear Models to Deep Learning" from MITx/edX.
 
 This paper presents the motivations and the general approach of the package and gives an overview of its organisation. We refer the reader to the [package documentation](https://sylvaticus.github.io/BetaML.jl/stable) for instructions on how to use the various algorithms provided or to the MOOC notes available on GitHub [@Lobianco:2020] for their mathematical backgrounds.
@@ -107,16 +107,16 @@ Knet: @Yuret:2016
 
 The BetaML toolkit is currently composed of 5 modules: `Utils` provides common data-science utility functions to be used in the other modules, `Perceptron` supplies linear and non-linear classifiers based on the classical Perceptron algorithm, `Nn` allows implementing and training artificial neural networks, `Clustering` includes several clustering algorithms and missing value attribution / collaborative filtering algorithms based on clustering and finally `Trees` implements decision trees classifiers/regressors together with their most common ensemble method, random forests.
 
-All sub-module functionalities are re-exported at the root level, so the user normally doesn't need to deal with the sub_modules, but just load the main `BetaML` module.
+All sub-module functionalities are re-exported at the root level, so the user doesn't need to deal with the sub-modules, but just load the main `BetaML` module.
 
 
 ## The `Utils` module
 
 The `Utils` module is intended to provide functionalities that are either: (a) used in other modules but are not strictly part of that specific module's logic (for example activation functions would be most likely used in neural networks, but could be of more general usage); (b) general methods that are used alongside the ML algorithms, e.g. to improve their predictions capabilities; or (c ) general methods to assess the goodness of fits of ML algorithms.
 
-Concerning the fist category `Utils` provides "classical" activation functions (and their respective derivatives) like `relu`, `sigmoid`, `softmax`, but also more recent implementations like elu [@Clevert:2015], celu [@Barron:2017], plu [@Nicolae:2018], softplus [@Glorot:2011] and mish [@Misra:2019].  Kernel functions (`radialKernel` - aka "KBF", `polynomialKernel`), distance metrics (`l1_distance` - aka "Manhattan", `l2_distance`, `l2²_distance`, `cosine_distance`), and functions typically used to improve numerical stability (`lse`) are also provided with the intention to be available in the different ML algorithms.
+Concerning the fist category `Utils` provides "classical" activation functions (and their respective derivatives) like `relu`, `sigmoid`, `softmax`, but also more recent implementations like `elu` [@Clevert:2015], `celu` [@Barron:2017], `plu` [@Nicolae:2018], `softplus` [@Glorot:2011] and `mish` [@Misra:2019].  Kernel functions (`radialKernel` - aka "KBF", `polynomialKernel`), distance metrics (`l1_distance` - aka "Manhattan", `l2_distance`, `l2²_distance`, `cosine_distance`), and functions typically used to improve numerical stability (`lse`) are also provided with the intention to be available in the different ML algorithms.
 
[…]
