### 2021-04-20 6a02de0
Minor docs updates/corrections
