### 2021-04-23 #2849 @sylvaticus
@arfon what are the next steps now ? Should I create a software deposit on zenodo ?
