### 2021-04-23 #2849 @arfon
@sylvaticus - could you please clean up all of the comments in your paper? I was trying to give the `paper.md` a read, and realized that lots of the content that I was struggling with was actually commented out.

Also, please add more information to your affiliations - I'm not sure what many of them are.
