### 2021-04-23 52f4e5d
JOSS paper: removed comments and put full affiliation names
