+++ paper.md
@@ -19,13 +19,13 @@ affiliations:
    index: 1
  - name: Université de Strasbourg
    index: 2
- - name: AgroParisTech
+ - name: Institut des sciences et industries du vivant et de l'environnement (AgroParisTech)
    index: 3
- - name: CNRS
+ - name: Centre national de la recherche scientifique (CNRS)
    index: 4
- - name: INRAE
+ - name: Institut national de recherche pour l’agriculture, l’alimentation et l’environnement (INRAE)
    index: 5
- - name: BETA
+ - name: Bureau d'économie théorique et appliquée (BETA)
    index: 6
 date: 19 April 2021
 bibliography: docs/paper/paper.bib
@@ -64,43 +64,15 @@ While much better results can be obtained (in general) by scaling the variables 
 
 Still BetaML offers a fair level of flexibility. As we didn't aim for heavy optimisation, we were able to keep the API (Application Programming Interface) both beginner-friendly and flexible.
 
-<!--The overall "philosophy" of BetaML is to be able to make simple things easy and complex things possible. -->
-
 If a great level of flexibility can already be achieved by just employing the full set of model parameters, the greatest flexibility is obtained by customising BetaML and writing, for example, its own neural network layer type (by subclassing `AbstractLayer`), its own sampler (by subclassing `AbstractDataSampler`) or its own mixture component (by subclassing `AbstractMixture`).
 While the library is designed for Julia users, the documentation provides examples for using the package from R or Python (thanks to JuliaCall [@Li:2019] and PyJulia [@Arakaki:2020] respectively).
 
-<!--To help beginners, many parameters and functions have pretty longer but more explicit names than usual. For example the Dense layer is a `DenseLayer`, the RBF kernel is `radialKernel`, etc.-->
-
 A few packages try to provide a common Julia framework of the various ML algorithms available in Julia, like  ScikitLearn.jl [@St-Jean:2020], AutoMLPipeline.jl [@Paulito:2020] or MLJ.jl [@Blaom:2019]. They build up on existing Julia (and/or Python) ML specialised packages. While avoiding the problem of "reinventing the wheel", the wrapping level may unintentionally introduces complications for the end-user, like the need to load the models and learn framework-specific concepts as _model_ or _machine_ in MLJ or `@pipeline` and `fit_transform!` in AutoMLPipeline.
-
-<!--Also it make difficult for the user to trace back the code performing the computations and modify for his own needs.-->
 
 We chose instead to bundle the main ML algorithms directly within the package. This offers a complementary approach that we feel it is more beginner-friendly.
 
-<!--
-Parameters Start gradually
-
-Simplicity not like Flux that rather than passing the neural network object to the train function in order to flexibility and optimise pass ....
-
-
-
-Of course you can get much better results (in general) by scaling the variables, adding further layer(s) and/or tuning their activation functions or the optimisation algorithm (have a look at the notebooks 1 or at the documentation for that), but the idea is that while we can offer a fair level of flexibility (you can choose or define your own activation function, easy define your own layers, choose weight initialisation, choose or implement the optimisation algorithm and its parameters, choose the training parameters - epochs, batchsize,…-, the callback function to get informations during (long) training,…), still we try to keep it one step at the time. So for most stuff we provide default parameters that can be overridden when needed rather than pretend that the user already know and provide all the needed parameters.
--->
-
 We believe that the BetaML flexibility and simplicity, together with the efficiency and usability of a Just in Time compiled language like Julia and the convenience to have several ML algorithms and data-science utilities all in the same package,
[…]
