### 2021-04-23 #2849 @sylvaticus
Done it.  I have removed the comments and specified the full affiliation names.

I am very sorry for the 6 affiliations (it's crazy, I know..) but that's the way we have been asked to sign our papers :-/ :

> 1) Tout ce qui est publié dans le BETA par les chercheurs INRA et AgroParisTech et tout ce qui traite de la forêt et du bois dans le BETA doit être signé:
>
> “Université de Lorraine, Université de Strasbourg, AgroParisTech, CNRS,  INRA, BETA, 54000, Nancy, France” pour les lorrains
>
> Et
>
> “Université de Strasbourg, Université de Lorraine, AgroParisTech, CNRS,  INRA, BETA, 67000, Strasbourg, France” pour les strasbourgeois
