### 2021-04-30 #2849 @arfon
@sylvaticus - I made a few minor changes to your paper here: https://github.com/sylvaticus/BetaML.jl/pull/23 . Once you have merged this PR, could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
