### 2021-04-30 e7a6113
JOSS  paper changes (for acceptance), bump version to v0.5.1
