+++ v0.5.1/tutorials/Classification - cars/betaml_tutorial_classification_cars.jl
@@ -1,0 +1,271 @@
+# # [A classification task when labels are known - determining the country of origin of cars given the cars characteristics](@id classification_tutorial)
+
+# In this exercise we have some car technical characteristics (mpg, horsepower,weight, model year...) and the country of origin and we would like to create a model such that the country of origin can be accurately predicted given the technical characteristics.
+# As the information to predict is a multi-class one, this is a _[classification]_(https://en.wikipedia.org/wiki/Statistical_classification) task.
+# It is a challenging exercise due to the simultaneous presence of three factors: (1) presence of missing data; (2) unbalanced data - 254 out of 406 cars are US made; (3) small dataset.
+
+#
+# Data origin:
+# - dataset description: [https://archive.ics.uci.edu/ml/datasets/auto+mpg](https://archive.ics.uci.edu/ml/datasets/auto+mpg)
+#src Also useful: https://www.rpubs.com/dksmith01/cars
+# - data source we use here: [https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data](https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data-original)
+
+# Field description:
+
+# 1. mpg:           _continuous_
+# 2. cylinders:     _multi-valued discrete_
+# 3. displacement:  _continuous_
+# 4. horsepower:    _continuous_
+# 5. weight:        _continuous_
+# 6. acceleration:  _continuous_
+# 7. model year:    _multi-valued discrete_
+# 8. origin:        _multi-valued discrete_
+# 9. car name:      _string (unique for each instance)_ - not used here
+
+
+
+# ## Library and data loading
+
+# We load a buch of packages that we'll use during this tutorial..
+using Random, HTTP, CSV, DataFrames, BenchmarkTools, BetaML
+import DecisionTree, Flux
+import Pipe: @pipe
+using  Test     #src
+
+# To load the data from the internet our workflow is
+# (1) Retrieve the data --> (2) Clean it --> (3) Load it --> (4) Output it as a DataFrame.
+
+# For step (1) we use `HTTP.get()`, for step (2) we use `replace!`, for steps (3) and (4) we uses the `CSV` package, and we use the "pip" `|>` operator to chain these operations:
+
+urlDataOriginal = "https://archive.ics.uci.edu/ml/machine-learning-databases/auto-mpg/auto-mpg.data-original"
+data = @pipe HTTP.get(urlDataOriginal).body                                                |>
+             replace!(_, UInt8('\t') => UInt8(' '))                                        |>
+             CSV.File(_, delim=' ', missingstring="NA", ignorerepeated=true, header=false) |>
+             DataFrame;
+
+# This results in a table where the rows are the observations (the various cars) and the column the fields. All BetaML models expect this layout.
+
+# As the dataset is ordered, we randomly shuffle the data. Note that we pass to shuffle `copy(FIXEDRNG)` as the random nuber generator in order to obtain reproducible output ( [`FIXEDRNG`](@ref BetaML.Utils.FIXEDRNG) is nothing else than an istance of `StableRNG(123)` defined in the [`BetaML.Utils`](@ref utils_module) sub-module, but you can choose of course your own "fixed" RNG). See the [Dealing with stochasticity](@ref dealing_with_stochasticity) section in the [Getting started](@ref getting_started) tutorial for details.
+data[shuffle(copy(FIXEDRNG),axes(data, 1)), :]
+describe(data)
+
+# Columns 1 to 7 contain  characteristics of the car, while column 8 encodes the country or origin ("1" -> US, "2" -> EU, "3" -> Japan). That's the variable we want to be able to predict.
+
[…]
